### 2021-12-03 #3976 @whedon
[PRE REVIEW]: DataAssimilationBenchmarks.jl: a data assimilation research framework
**Submitting author:** @cgrudz (<a href="http://orcid.org/0000-0002-3084-3178">Colin Grudzien</a>)
**Repository:** <a href="https://github.com/cgrudz/DataAssimilationBenchmarks.jl" target ="_blank">https://github.com/cgrudz/DataAssimilationBenchmarks.jl</a>
**Version:** v0.2.0
