### 2021-12-03 #3976 @kthyng
@cgrudz in the meantime, could you state here how the GMD paper is different from the JOSS submission? The manuscript for GMD itself certainly doesn't look like an in-depth review of the software like in JOSS, but can you speak to what the reviewers of the paper will be doing as part of their review? From my experience with GMD, there are different "types" of papers which have different levels of rules regarding the software.
