### 2021-12-03 #3976 @cgrudz
@kthyng thanks for handling this so quickly, @taless474, thanks for editing!

> @cgrudz in the meantime, could you state here how the GMD paper is different from the JOSS submission? The manuscript for GMD itself certainly doesn't look like an in-depth review of the software like in JOSS, but can you speak to what the reviewers of the paper will be doing as part of their review? From my experience with GMD, there are different "types" of papers which have different levels of rules regarding the software.

Regarding the scope of the different manuscripts, the GMD work is more mathematical where we develop the theory for a novel iterative estimator, the SIEnKS.  Particularly, this includes a review of several similar mathematical methods all within a Bayesian estimation framework.  We also present a variety of numerical simulations to demonstrate the performance of the scheme versus other similar methods.  Because this involved a significant amount of development to rigorously derive the numerics, the DataAssimilationBenchmarks.jl software package grew out of this mathematical study concurrently.  We wish submit our manuscript to JOSS to better describe the software implementation aspects which are not discussed in GMD, together giving a comprehensive reference on the study.  Writing the JOSS manuscript particularly involved the work of my undergraduate research assistant who was the primary supporting developer, building test cases and developing the overall package structure.

Please let me know if you have any further questions.

Cheers,
Colin
