### 2021-12-13 #3976 @cgrudz
@taless474, I would like to suggest the following:

John Maclean, Lecturer in Data Science and Statistics, School of Mathematical Sciences, University of Adelaide, john.maclean@adelaide.edu.au

Patrick N. Raanes, Researcher, NORCE Norwegian Research Institute, para@norceresearch.no, @patricknraanes

Artem Moiseev, Researcher, Nansen Environmental and Remote Sensing Center, Artem.Moiseev@nersc.no, @korvinos

Yumeng Chen, Data Assimilation Scientific Programmer (PDRA), National Centre for Earth Observation, University of Reading, yumeng.chen@reading.ac.uk, @yumengch

Please let me know if you would like further suggestions.
