### 2021-12-15 #3976 @cgrudz
Ah, sorry @taless474, I missed the list at the top, certainly:

Henning Bonart, henning.bonart@tu-berlin.de, bonh, does research in data assimilation and has Julia listed as a secondary language.

Lindsey Heagy,  lheagy@eoas.ubc.ca, lheagy, does research in geophysical inverse problems, which is a closely related area to the data assimilation methods studied in this package, and has Julia listed as a secondary language.

Felix Cremer, felix.cremer@dlr.de, felixcremer, does research in remote sensing and earth observation time series analysis, which is a related topic to data assimilation, and has Julia as a preferred language.

Lukas Riedel, lukas.riedel@iwr.uni-heidelberg.de, peanutfun, does research in data assimilation, and although their listed languages does not include Julia, I think they would be familiar with the mathematical methods that underpin the numerics.
 Hopefully the software will be easy to understand with this background, as this would represent a target audience.

Please let me know if you need any further information or suggestions.

Cheers,
Colin
