### 2022-01-12 ead3d7b
Updating citation for version number
Revised citation after release.
