+++ CITATION.bib
@@ -2,7 +2,7 @@
 	author  = {Colin Grudzien <cgrudz@mailbox.org> and contributors},
 	title   = {DataAssimilationBenchmarks.jl},
 	url     = {https://github.com/Colin Grudzien/DataAssimilationBenchmarks.jl},
-	version = {v0.1.0},
+	version = {v0.2.0},
 	year    = {2021},
-	month   = {6}
+	month   = {12}
 }
