### 2022-01-13 #3976 @peanutfun
Yes, I would like to review! I feel confident in reviewing the data assimilation methods and the repository overall, but I haven't worked with Julia yet. I will agree to review if another reviewer joins who is "fluent" in Julia.
