### 2022-02-03 #4129 @whedon
[REVIEW]: DataAssimilationBenchmarks.jl: a data assimilation research framework
**Submitting author:** @cgrudz (<a href="http://orcid.org/0000-0002-3084-3178">Colin Grudzien</a>)
**Repository:** https://github.com/cgrudz/DataAssimilationBenchmarks.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.4
