### 2022-02-04 #4129 @peanutfun
Hi @cgrudz 👋 I'll review your software in the next couple of days, and open up issues in the repository as I go along. Some of them might only be suggestions and not crucial to the review. Once I'm finished, I'll report back here and summarize the issues with respect to the review criteria and the checklist above.

Happy coding! 🚀
