### 2022-02-11 cgrudz/DataAssimilationBenchmarks.jl#7 @peanutfun
Add usage examples and workflow to README.md
The `README.md` documents the API of the module. As the argument lists are quite extensive, I think it would be helpful to add "typical" usage examples for some of the methods and solvers. In particular, this would help to demonstrate some kind of workflow. Starting out with generating a time series, an example could show how to run different data assimilation methods included in this package, and then how to evaluate the output and compare the methods against each other.

Currently, one has to look up a typical argument list in the tests, which is cumbersome.

openjournals/joss-reviews#4129
