### 2022-02-11 cgrudz/DataAssimilationBenchmarks.jl#8 @peanutfun
Consider adding docstrings to exported functions
The public API of the `DataAssimilationBenchmarks.jl` module is currently documented in the `README.md`. This clutters the document and does not adhere to a standardized documentation formatting.

Julia enables object documentation via [docstrings](https://docs.julialang.org/en/v1/manual/documentation/), and advertises a standard to write such documentations. This has the advantage that the documentation is part of the actual source code, and that a formatted documentation can always be generated from the source code with [`Documenter.jl`](https://juliadocs.github.io/Documenter.jl/stable/).

Please consider removing the API documentation from the `README.md`, and adding appropriate docstrings to the functions of the module that are part of the public API. I think this would work well together with my suggestion in #7, thus making the `README.md` a more "high-level" overview of the package.

openjournals/joss-reviews#4129
