### 2022-02-11 cgrudz/DataAssimilationBenchmarks.jl#10 @peanutfun
Document output file structure
All operations in this module are output file based, meaning that no function returns its data, but all data is written into output files to be retrieved later on. For this to be viable also in conjunction with other modules, it would be helpful to have the structure of the output documented. Which datasets are written by the, e.g., `l96_time_series` and `filter_state` functions, and what are their data types and dimensions? This would particularly help when writing scripts to analyze the results.

The output documentation could be part of the function docstring (see #8), or added to the `README.md` where appropriate.

openjournals/joss-reviews#4129
