### 2022-02-15 cgrudz/DataAssimilationBenchmarks.jl#11 @peanutfun
Rework analysis scripts to take path to data file as argument
The analysis scripts currently seem defunct, as they try to run visualization scripts on specific, partially hard-coded paths. I think they would be very helpful to immediately visualize the results. They could be reworked to take a path to a `.jld2` data file, similar to the other functions of this package, and then analyze/visualize the data contained in it.
