### 2022-02-15 cgrudz/DataAssimilationBenchmarks.jl#12 @peanutfun
Options for custom observations or observation operators
As far as I see, functions in this package currently assume that every state dimension is always observed (with some observation error). From my experience with data assimilation, this is rarely the case, and the performance of DA algorithms can strongly differ depending on which, and how many, state dimensions are part of the observation. For a DA comparison framework, I suggest more control over observations (i.e., observation operators) at the user side of things. A first step would be to determine which state dimensions are observed, e.g., via a vector of indices that is supplied as function argument. A further improvement would be the option to add the complete observation operator matrix which calculates the observations via multiplication with the state vector.

openjournals/joss-reviews#4129
