### 2022-02-15 cgrudz/DataAssimilationBenchmarks.jl#14 @peanutfun
Add community guidelines
[Community guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines) are a review criterion of JOSS publications. Please add information to this repository on how to seek support, report issues, and how to contribute. This can be supplied in sections of the `README.md` and in a separate `CONTRIBUTING.md` file. More information on this topic can be found, e.g., in the Github docs: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors

openjournals/joss-reviews#4129
