### 2022-02-22 cgrudz/DataAssimilationBenchmarks.jl#15 @peanutfun
Add instructions on how to add new DA schemes and models
The `README.md` lacks instructions on how to extend the framework with new models and new DA schemes. Independently from my suggestion to change how DA schemes can be specified when calling functions of this module (#13), the current way of implementing additions should be documented. For example, the [DAPPER Python library](https://github.com/nansencenter/DAPPER) documents [how to add models](https://github.com/nansencenter/DAPPER/tree/master/dapper/mods#defining-your-own-model) and at least [refers to examples](https://github.com/nansencenter/DAPPER/tree/master/dapper/da_methods#defining-your-own-method) for adding DA schemes. I understand that defining interfaces for DA schemes and models is more difficult in Julia because the language features no classes like in C++ or Python. However, I think that at least a rough implementation guideline should be supplied. This can be done in the `README.md`, or in another document if the instructions become too lengthy.

openjournals/joss-reviews#4129
