### 2022-02-22 cgrudz/DataAssimilationBenchmarks.jl#16 @peanutfun
Suggestions for improving JOSS paper
The [JOSS paper](https://github.com/cgrudz/DataAssimilationBenchmarks.jl/blob/master/paper.md) gives a very good introduction into the topic of data assimilation, provides a nice overview of the project, and compares it well to similar packages. I only have a few suggestions to make it even more concise. Most of them concern sections that are part of the `README.md` and hence not required in the paper.

* Remove the *Documentation* section. A package documentation is a review criterion of a JOSS publication and hence will exist in the package repository.
* Remove the *Installation* section. Installation instructions are part of the `README.md`. The first paragraph of the *Installation* section currently gives an overview of the source code structure. This can be moved to the *Summary* section. I also think it suffices to mention in the Summary that the package is registered in the Julia General registry and hence can be downloaded and installed via REPL.
* The reference `grudzien2020numerical` appears twice in the list of references. I am not sure what causes this. It might be an issue with the JOSS paper generator.
* Please add DOIs to every reference. If no DOI is available, please add an URL to the source.
* For referencing Numba, please refer to the respective [instructions in the documentation](https://numba.readthedocs.io/en/stable/user/faq.html#how-do-i-reference-cite-acknowledge-numba-in-other-work).
* Likewise, the Julia docs [encourage a citation](https://julialang.org/research/#research) when using Julia for research.
* Personally, I would not consider the webpage https://julialang.org/benchmarks/ a viable reference. I suggest adding a footnote in this case instead.
* The reference to the [Zenodo archive of DAPPER](https://zenodo.org/record/2029296) does not have the same author list as given by the citation instructions of said archive. However, since the archive contains a presumably outdated version, I suggest referencing the Github repository https://github.com/nansencenter/DAPPER instead.

openjournals/joss-reviews#4129
