### 2022-02-23 #4129 @peanutfun
@cgrudz, I finished my review. Let me summarize:

DataAssimilationBenchmarks.jl is a well-working framework for researching data assimilation techniques. Installation is easy and I experienced no issues when executing functions of this package. The automated tests are working well and help understanding the inner workings of the module, although code coverage is still a bit low. The `README.md` gives a good overview of the framework and extensively documents the public API. Finally, the paper is well written and does an especially good job in motivating the development of the framework.

My main issue with the current state of the package is that I, as a person unfamiliar with the source code, see no clear way of extending the framework with new models and DA methods. In that regard, I feel that DataAssimilationBenchmarks.jl fails its own premise because it is intended to compare DA methods against each other and include more filters and smoothers in the future. As an open-source project, it should encourage contributions by others, and therefore needs instructions on how to add such new methods to the framework.

The other points I want to address mostly concern documentation. The `README.md` should include an *Example Usage* section, that illustrates the capabilities of the framework and the intended workflow, and supplies a useful set of default function arguments. As the module functions do not return data structures but write all data into files, the structure and content of these files should be documented as well. Additionally, the repository lacks community guidelines, which is more of a technicality. And finally, I think the paper can be shortened, and some of its references can be improved.

I suggested a lot more in the issues I created during my review, which is mainly due to the fact that I see a lot of potential in DataAssimilationBenchmarks.jl. But most of them are well beyond the scope of this review. Two things I still would like to point out: First, you already documented the API extensively in the `README.md`. Moving this documentation over to docstrings in the code would yield a documentation that is formatted in a standardized and expected way, and should not be much work. And second, I think that reworking the analysis scripts to work on any system would be a huge selling point. Ideally, the framework would then cover everything from running a model, over benchmarking DA methods in parallel, to evaluating and analyzing their performance.

## Issues for meeting review criteria

- [ ] cgrudz/DataAssimilationBenchmarks.jl#7
- [ ] cgrudz/DataAssimilationBenchmarks.jl#10
- [ ] cgrudz/DataAssimilationBenchmarks.jl#14
- [ ] cgrudz/DataAssimilationBenchmarks.jl#15
- [ ] Incorporate suggestions to JOSS paper (or conclusively reject them) cgrudz/DataAssimilationBenchmarks.jl#16

## Possible improvements I strongly recommend

*(These are personal suggestions to improve the package, and they are not required to meet the review criteria for publication in JOSS)*

- [ ] Use docstrings to document API cgrudz/DataAssimilationBenchmarks.jl#8
- [ ] Rework analysis scripts to be used with every experiment/benchmark cgrudz/DataAssimilationBenchmarks.jl#11
