### 2022-03-01 #4129 @cgrudz
@peanutfun thank you for your excellent suggestions and detailed review, I really appreciate the effort you took to go through the code as you did.  I just wanted to write a quick message to let you know that I am intending to start on revisions and reading through the posted issues in about a month or two -- I am waiting in part to see if I can fund an undergraduate research assistant to help resolve the issues, as I think these will make good exercises for an RA again.  However, if this funding does not go through, I do intend to resolve these issues myself.  I'll keep the review process updated with my status shortly.

Cheers!
Colin
