### 2022-03-06 #4129 @tmigot
Hi @cgrudz ! Sorry, I am a bit late in the review. But, I saw @peanutfun already did a great job.
I am checking the  `Functionality` section.
- Installation: Does installation proceed as outlined in the documentation?

Everything works following your suggestions. However, when adding your package the "usual" way in Julia, i.e.
```
add DataAssimilationBenchmark
```
I get the following error when testing
```
     Testing Running tests...
Test Summary:               | Pass  Total
Calculate Order Convergence |    2      2
Test Summary: | Pass  Total
Lorenz-96     |    2      2
Runtime 0.0925 minutes
┌ Warning: Opening file with JLD2.MmapIO failed, falling back to IOStream
└ @ JLD2 ~/.julia/packages/JLD2/k9Gt0/src/JLD2.jl:233
Error encountered while save FileIO.File{FileIO.DataFormat{:JLD2}, String}("/home/tmigot/.julia/packages/DataAssimilationBenchmarks/ZFBxn/src/experiments/../data/time_series/IEEE39bus_time_series_seed_0000_diff_0.000_tanl_0.01_nanl_05000_spin_1500_h_0.010.jld2").

Fatal error:
Time Series Generation: Test Failed at /home/tmigot/.julia/packages/DataAssimilationBenchmarks/ZFBxn/test/runtests.jl:41
  Expression: TestTimeSeriesGeneration.testGenIEEE39bus()
Stacktrace:
 [1] macro expansion
   @ ~/packages/julias/julia-1.7.1/share/julia/stdlib/v1.7/Test/src/Test.jl:445 [inlined]
 [2] macro expansion
   @ ~/.julia/packages/DataAssimilationBenchmarks/ZFBxn/test/runtests.jl:41 [inlined]
 [3] macro expansion
   @ ~/packages/julias/julia-1.7.1/share/julia/stdlib/v1.7/Test/src/Test.jl:1283 [inlined]
 [4] top-level scope
   @ ~/.julia/packages/DataAssimilationBenchmarks/ZFBxn/test/runtests.jl:39
Test Summary:          | Pass  Fail  Total
Time Series Generation |    3     1      4
ERROR: LoadError: Some tests did not pass: 3 passed, 1 failed, 0 errored, 0 broken.
in expression starting at /home/tmigot/.julia/packages/DataAssimilationBenchmarks/ZFBxn/test/runtests.jl:2
ERROR: Package DataAssimilationBenchmarks errored during testing
```
Maybe the error comes from Julia 1.7 or is it an issue with the `add` ?

In the latter case, this might be inconvenient, for instance if someone uses this package as a dependency. What do you think? Once we clarify this, I can open an issue on this topic.

- Functionality: Have the functional claims of the software been confirmed?

As mentioned in https://github.com/cgrudz/DataAssimilationBenchmarks.jl/issues/7  the documentation and the paper lacks an illustration of what the package is capable of doing.
