### 2022-03-06 #4129 @tmigot
@cgrudz Have you thought about making a real documentation hosted with Github Pages and using `Documenter.jl`?

See docs for Documenting in Julia: https://juliadocs.github.io/Documenter.jl/stable/

I have a package that has been recently published in JOSS as an example: https://github.com/JuliaSmoothOptimizers/DCISolver.jl (you can access the doc via the badges dev and stable in the readme).

This would allow clearer and more complete documentation. That could be very relevant, especially, in the perspective of expanding your package.
