### 2022-03-08 #4129 @cgrudz
@tmigot, no worries, thank you very much for your helpful suggestions and for catching the bug with the package add.  As I mentioned above I will start to work on these revisions in earnest in one to two months, once I determine the availability of a research assistant to help with these tasks.  I'll address these items as soon as I can. Cheers!
