+++ docs/src/index.md
@@ -1,0 +1,7 @@
+# DataAssimilationBenchmarks.jl
+
+Documentation for DataAssimilationBenchmarks.jl
+
+```@docs
+DataAssimilationBenchmarks.L96.dx_dt
+```
