+++ src/analysis/state_smoother/plt_smoother_rmse_spread_optimum_forecast_lag_versus_nens.py
@@ -106,7 +106,7 @@ for meth in method_list:
             stat_name = 'forecast'
 
         if meth == "etks_classic":
-            meth_name = "ETKS"
+            meth_name = "EnKS"
         elif meth == "etks_single_iteration":
             meth_name = "SIEnKS"
         elif meth == "enks-n-primal_classic":
