+++ src/methods/EnsembleKalmanSchemes.jl
@@ -15,39 +15,66 @@ export alternating_obs_operator, analyze_ens, analyze_ens_para, rand_orth,
 ##############################################################################################
 # Type union declarations for multiple dispatch and type aliases
 
-# covariance matrix types
+"""
+    CovM = Union{UniformScaling{Float64}, Diagonal{Float64}, Symmetric{Float64}}
+
+Type union of covariance matrix types, for optimized computation based on characteristics.
+"""
 CovM = Union{UniformScaling{Float64}, Diagonal{Float64}, Symmetric{Float64}}
 
-# conditioning matrix types
+"""
+    ConM = Union{UniformScaling{Float64}, Symmetric{Float64}}
+
+Type union of conditioning matrix types, which will be used for optimization routines in the
+transform method.
+"""
 ConM = Union{UniformScaling{Float64}, Symmetric{Float64}}
 
-# right transform types, including soley a transform, or transform, weights rotation package
+"""
+    TransM = Union{Tuple{Symmetric{Float64,Array{Float64,2}},Array{Float64,2},Array{Float64,2}},
+                   Tuple{Symmetric{Float64,Array{Float64,2}},Array{Float64,1},Array{Float64,2}},
+                   Array{Float64,2}}
+
+Type union of right transform types, including soley a transform, or a transform, weights
+and rotation package.
+"""
 TransM = Union{Tuple{Symmetric{Float64,Array{Float64,2}},Array{Float64,2},Array{Float64,2}},
                Tuple{Symmetric{Float64,Array{Float64,2}},Array{Float64,1},Array{Float64,2}},
                Array{Float64,2}}
 
-# vectors and ensemble members of sample
+"""
+    VecA = Union{Vector{Float64}, SubArray{Float64, 1}}
+
+Type union of vectors and ensemble members of sample for using in integration schemes and
+related array operations.
+"""
 VecA = Union{Vector{Float64}, SubArray{Float64, 1}}
 
-# arrays and views of arrays
+"""
+    ArView = Union{Array{Float64, 2}, SubArray{Float64, 2}}
+
+Type union of arrays and views of arrays for use within ensemble conditioning operations,
+integration schemes and other array operations.
+"""
 ArView = Union{Array{Float64, 2}, SubArray{Float64, 2}}
 
 ##############################################################################################
 ##############################################################################################
 # Main methods, debugged and validated
 ##############################################################################################
-# alternating id obs
+"""
+    alternating_obs_operator(ens::Array{Float64,2}, obs_dim::Int64, 
+                             kwargs::Dict{String,Any})
 
+Observation of alternating state vector components, possibly nonlinear transformation.
+This selects components to observe based on the observation dimension and if
+parameter estimation is being performed.  Parameters are always unobservable,
+and even states will be removed from the state vector until the observation dimension
+is appropriate.  Nonlinear observations are optional, as described for the
+Lorenz-96 model by Asch, Bocquet, Nodet pg. 181.
+"""
 function alternating_obs_operator(ens::Array{Float64,2}, obs_dim::Int64,
                                   kwargs::Dict{String,Any})
-    """Observation of alternating state vector components, possibly nonlinear transformation.
-
-    This selects components to observe based on the observation dimension and if
-    parameter estimation is being performed.  Parameters are always unobservable,
-    and even states will be removed from the state vector until the observation dimension
-    is appropriate.  Nonlinear observations are optional, as described for the
-    Lorenz-96 model by Asch, Bocquet, Nodet pg. 181 """
-
     sys_dim, N_ens = size(ens)
 
     if haskey(kwargs, "state_dim")
@@ -109,10 +136,12 @@ end
 
 
 ##############################################################################################
-# ensemble state statistics
+"""
[…]
