+++ docs/make.jl
@@ -22,9 +22,6 @@ makedocs(
             ]
 )
 
-# Documenter can also automatically deploy documentation to gh-pages.
-# See "Hosting Documentation" and deploydocs() in the Documenter manual
-# for more information.
-#=deploydocs(
-    repo = "<repository url>"
-)=#
+deploydocs(
+    repo = "github.com/cgrudz/DataAssimilationBenchmarks.jl.git",
+)
