+++ docs/make.jl
@@ -22,6 +22,3 @@ makedocs(
             ]
 )
 
-deploydocs(
-    repo = "github.com/cgrudz/DataAssimilationBenchmarks.jl.git",
-)
