+++ .travis.yml
@@ -4,7 +4,7 @@ julia:
 codecov: true
 after_success:
 - julia -e 'using Pkg; cd(Pkg.dir("DataAssimilationBenchmarks")); Pkg.add("Coverage"); using Coverage; Codecov.submit(Codecov.process_folder())'
-
+codecov: true
 os: linux
 arch:
   - arm64
