+++ src/experiments/SingleExperimentDriver.jl
@@ -1,9 +1,9 @@
 ##############################################################################################
 module SingleExperimentDriver
 ##############################################################################################
-##############################################################################################
 # imports and exports
-using ..FilterExps, ..SmootherExps, ..GenerateTimeSeries, JLD2
+using JLD2, HDF5
+using ..DataAssimilationBenchmarks, ..FilterExps, ..SmootherExps, ..GenerateTimeSeries
 export filter_state_exp, filter_param_exp, classic_smoother_state_exp,
        classic_smoother_param_exp, single_iteration_smoother_state_exp,
        single_iteration_smoother_param_exp, iterative_smoother_state_exp,
@@ -17,10 +17,7 @@ export filter_state_exp, filter_param_exp, classic_smoother_state_exp,
 # for debugging, benchmarking and profiling.  These will work simply with the standard macros,
 # while the parallel submission scripts are used for performance on servers
 ##############################################################################################
-##############################################################################################
-##############################################################################################
 ## Generate time series data
-##############################################################################################
 ##############################################################################################
 
 ##############################################################################################
@@ -48,9 +45,7 @@ end
 
 
 ##############################################################################################
-##############################################################################################
 ## DA experiments begin here
-##############################################################################################
 ##############################################################################################
 
 ##############################################################################################
@@ -81,8 +76,6 @@ time_series = path * fname
 
 
 ##############################################################################################
-
-##############################################################################################
 # Filters
 ##############################################################################################
 # filter state estimation, arguments are
@@ -106,8 +99,6 @@ end
 
 
 ##############################################################################################
-
-##############################################################################################
 # Classic smoothers
 ##############################################################################################
 # classic EnKS style smoothing for state estimation, arguments are
@@ -129,8 +120,6 @@ function classic_smoother_param_exp()
     classic_param(args)
 end
 
-
-##############################################################################################
 
 ##############################################################################################
 # Single iteration smoothers
