+++ docs/src/submodules/experiments/SmootherExps.md
@@ -1,12 +1,37 @@
 # SmootherExps 
 
+
+## SmootherExps API
+
+The `SmootherExps.jl` sub-module configures twin experiments using stored time series data as generated above for 
+efficiency when using the same base-line time series to generate possibly different experiment configurations.
+Experiment configurations are generated by function calls as with the filter experiments, but with the additional
+options of how the outer-loop is configured with a classic, single-iteration or the fully iterative Gauss-Newton style smoother.
+The parameters `lag` and `shift` specify how the data assimilation windows are translated in over the observation
+and analysis times.  The `mda` parameter is only applicable to the single-iteration and Gauss-Newton style smoothers,
+utlizing sequential multiple data assimilation.  Note, the single-iteration and fully iterative Gauss-Newton style smoothers are
+only defined for MDA compatible values of lag and shift where the lag is an integer multiple of the shift.
+
+Currently debugged and validated smoother experiment configurations include
+```
+classic_state          -- classic EnKS style state estimation
+classic_param          -- classic EnKS style state-parameter estimation
+single_iteration_state -- single-iteration EnKS state estimation
+single_iteration_param -- single-iteration EnKS state-parameter estimation
+iterative_state        -- Gauss-Newton style state estimation
+iterative_param        -- Gauss-Newton style state-parameter estimation
+```
+Other techniques are still in debugging and validation.  Each of these takes an analysis type as used in the
+`transform` function in the `EnsembleKalmanSchemes.jl` sub-module, like the filter analyses in the filter experiments.
+
+
 All experiments are funcitonalized so that they can be called from an array of parameter
 values which will typically be varied with naive parallelism.  Relevant arguments and
 experimental results are dumped as a side effect to a dictionary in a JLD2.
 Experiments return their runtime in minutes.
 
 
-
+## Docstrings
 ```@autodocs
 Modules = [DataAssimilationBenchmarks.SmootherExps]
 ```

[…]
