+++ docs/make.jl
@@ -35,3 +35,7 @@ makedocs(
             ]
 )
 
+
+deploydocs(
+    repo = "github.com/cgrudz/DataAssimilationBenchmarks.jl.git",
+)
