### 2022-04-15 5feaf6e
Updating to named tuple experiment inputs
All main experiments have changed over to a named tuple form of the
argument list, so that these can be packed with names and interchanged
as parameter reproduction sets.  SingleExperimenDriver is changed in
paradigm to simply form a dictionary of standard experiment
configurations to be called by name and loaded as a named tuple.  Minor
changes to naming conventions, making consistent across filter /
smoother inflation naming.  All tests have been updated to call
SingleExperimentDriver to load experiments, this forms a basic template
for developing th
[…]
