+++ src/models/L96.jl
@@ -3,11 +3,14 @@ module L96
 ##############################################################################################
 # imports and exports
 using ..DataAssimilationBenchmarks 
-export dx_dt, jacobian, l96s_tay_2_step!, ρ, α
-
+export dx_dt, jacobian, l96s_tay_2_step!, compute_α_ρ 
 ##############################################################################################
-# auxiliary function to return modular indices for the lorenz model
+"""
+    mod_indx!(indx::Int64, dim::Int64) 
 
+Auxiliary function to return state vector indices for the Lorenz-96 model, where `indx` is
+taken mod `dim`.  Mod zero is replaced with `dim` for indexing in Julia state vectors.
+"""
 function mod_indx!(indx::Int64, dim::Int64)
     indx = mod(indx, dim)
     if indx==0
@@ -18,15 +21,12 @@ end
 
 
 ##############################################################################################
-# time derivative
-
-
 """
     dx_dt(x::VecA, t::Float64, dx_params::ParamDict)
 
-    Time derivative for Lorenz-96 model, x is a single model state of size state_dim, t
-    is a dummy time argument for consistency with integration methods, dx_params is a 
-    parameter dictionary which is called for the forcing parameter.
+Time derivative for Lorenz-96 model, `x` is a single model state of size `state_dim`, `t`
+is a dummy time argument for consistency with integration methods, `dx_params` is of type
+`ParaDict` which is called for the forcing parameter.
 """
 function dx_dt(x::VecA, t::Float64, dx_params::ParamDict)
     # unpack the (only) derivative parameter for l96
@@ -51,16 +51,14 @@ end
 
 
 ##############################################################################################
-# linearized time derivative
-
 """
     jacobian(x::Vector{Float64}, t::Float64, dx_params::ParamDict) 
     
-    Computes the Jacobian of Lorenz-96 about the state x. The time variable t is a dummy
-    variable for consistency with integration methods, dx_params is a parameter dictionary
-    which is called for the forcing parameter. Note that this is designed to load entries in
-    a zeros array and return a sparse array to make a compromise between memory
-    and computational resources.
+Computes the Jacobian of Lorenz-96 about the state `x`. The time variable `t` is a dummy
+variable for consistency with integration methods, `dx_params` is of type `ParamDict`
+which is called for the forcing parameter. Note that this is designed to load entries in
+a zeros array and return a sparse array to make a compromise between memory
+and computational resources.
 """
 function jacobian(x::Vector{Float64}, t::Float64, dx_params::ParamDict)
 
@@ -91,39 +89,46 @@ function jacobian(x::Vector{Float64}, t::Float64, dx_params::ParamDict)
     return sparse(dxF)
 end
 
-##############################################################################################
-# Auxiliary functions for the 2nd order Taylor-Stratonovich expansion below. These need
-# to be computed once, only as a function of the order of truncation of the Fourier
-# series, p for full timeseries.
-
-function ρ(p::Int64)
-    1.0/12.0 - 0.5 * π^(-2.0) * sum(1.0 ./ Vector{Float64}(1:p).^2.0)
-end
-
-function α(p::Int64)
-    (π^2.0) / 180.0 - 0.5 * π^(-2.0) * sum(1.0 ./ Vector{Float64}(1:p).^4.0)
-end
 
 ##############################################################################################
-# 2nd order strong taylor SDE step
+"""
+    compute_α_ρ(p::Int64) 
 
+Computes auxiliary functions for the 2nd order Taylor-Stratonovich expansion. The constants
+`α` and `ρ` need to be computed once, only as a function of the order of truncation of the
[…]
