+++ test/runtests.jl
@@ -1,12 +1,9 @@
 ##############################################################################################
 module runtests
 ##############################################################################################
-##############################################################################################
 # imports and exports
 using Test
 using JLD2
-
-##############################################################################################
 ##############################################################################################
 # include test sub-modules
 include("TestDeSolvers.jl")
@@ -17,8 +14,6 @@ include("TestFilterExps.jl")
 include("TestClassicSmootherExps.jl")
 include("TestIterativeSmootherExps.jl")
 include("TestSingleIterationSmootherExps.jl")
-
-##############################################################################################
 ##############################################################################################
 # Run tests
 
@@ -88,6 +83,7 @@ end
     @test TestSingleIterationSmootherExps.run_mda_smoother_param_L96()
     @test TestSingleIterationSmootherExps.analyze_mda_smoother_param_L96()
 end
+
 
 ##############################################################################################
 # end module
