+++ src/models/IEEE39bus.jl
@@ -4,14 +4,13 @@ module IEEE39bus
 # imports and exports
 using ..DataAssimilationBenchmarks
 export dx_dt
-
 ##############################################################################################
 """
     dx_dt(x::VecA, t::Float64, dx_params::ParamDict) 
     
 Time derivative of the phase and fequency of the effective-network swing equation model.
-Input x is a 2 n_g vector of the phase and fequency at each of the n_g generator buses.
-The input dx_params is a ParamDict of all system parameters to be passed to the
+Input x is a 2 `n_g` vector of the phase and fequency at each of the `n_g` generator buses.
+The input `dx_params` of type `ParamDict` containing system parameters to be passed to the
 integration scheme.  The system is currenty defined autonomously to be run as an SDE,
 noise perturbed steady state.
 """
