+++ test/TestIEEE39bus.jl
@@ -1,19 +1,21 @@
 ##############################################################################################
 module TestIEEE39bus
 ##############################################################################################
-##############################################################################################
 # imports and exports
-using DataAssimilationBenchmarks.DeSolvers, DataAssimilationBenchmarks.L96
+using DataAssimilationBenchmarks
 using JLD, Statistics
-
 ##############################################################################################
-##############################################################################################
-# Test to see if the model without noise reaches the synchronous steady state
+"""
+    test_synchrony() 
 
+This function tests to see if the swing equation model without noise reaches the synchronous
+steady state for the system, by evaluating the standard deviation of the state components
+after the spin up period.
+"""
 function test_synchrony()
     try
         # load the observations
-        path = joinpath(@__DIR__, "../src/data/time_series/")
+        path = pkgdir(DataAssimilationBenchmarks) * "/src/data/time_series/"
         obs = load(path * "IEEE39bus_time_series_seed_0000_diff_0.000_tanl_0.01" *
                    "_nanl_05000_spin_1500_h_0.010.jld2")
         obs = obs["obs"][:, 3001:end]
@@ -29,6 +31,7 @@ function test_synchrony()
         false
     end
 end
+
 
 ##############################################################################################
 # end module
