+++ docs/make.jl
@@ -36,6 +36,6 @@ makedocs(
 )
 
 
-deploydocs(
-    repo = "github.com/cgrudz/DataAssimilationBenchmarks.jl.git",
-)
+#deploydocs(
+#    repo = "github.com/cgrudz/DataAssimilationBenchmarks.jl.git",
+#)

[…]
