### 2022-05-10 d49f7f8
Major docs revision and streamlining
