+++ src/experiments/SingleExperimentDriver.jl
@@ -4,16 +4,22 @@ module SingleExperimentDriver
 # imports and exports
 using JLD2, HDF5
 using ..DataAssimilationBenchmarks, ..FilterExps, ..SmootherExps, ..GenerateTimeSeries
-export exps
-##############################################################################################
-# The following are standard inputs for experiments, written as named tuples and stored in a
-# dictionary.  These standard inputs are used in the package for for debugging, testing, 
-# benchmarking and profiling. Parallel submission scripts are used for performance on servers.
+export exps, foo
 ##############################################################################################
 path = pkgdir(DataAssimilationBenchmarks) * "/src/data/time_series/"
 
+"""
+    exps["Experiment_name"]["Parameter_settings"]
+
+This dictionary contains standard inputs for experiments, written as named tuples and stored
+hierarchically by experiment type.  These standard inputs are used in the package
+for for debugging, testing, benchmarking and profiling. Parallel submission scripts are used
+for performance on servers.
+"""
 exps = Dict{String, Any}(
+        # Generate time series experiment configurations
         "Generate_time_series" => Dict{String, Any}(
+          # Generates a short time series of the L96 model for testing
           "L96_deterministic_test" => (
             seed      = 0,
             state_dim = 40,
@@ -23,6 +29,7 @@ exps = Dict{String, Any}(
             diffusion = 0.00,
             F         = 8.0,
            ),
+          # Generates a short time series of the IEEE39bus model for testing
           "IEEE39bus_deterministic_test" => (
             seed      = 0,
             tanl      = 0.01,
@@ -31,7 +38,9 @@ exps = Dict{String, Any}(
             diffusion = 0.0,
            ),
           ),
+        # Filter twin experiment configurations
         "Filter" => Dict{String, Any}(
+          # Lorenz-96 ETKF state estimation standard configuration
           "L96_ETKF_state_test" => (
             time_series = path * 
             "L96_time_series_seed_0000_dim_40_diff_0.000_F_08.0_tanl_0.05_nanl_05000_" *
@@ -45,6 +54,7 @@ exps = Dict{String, Any}(
             N_ens       = 21,
             s_infl      = 1.02,
            ),
+          # Lorenz-96 ETKF joint state-parameter estimation standard configuration
           "L96_ETKF_param_test" => (
             time_series = path * 
             "L96_time_series_seed_0000_dim_40_diff_0.000_F_08.0_tanl_0.05_nanl_05000_" *
@@ -61,21 +71,24 @@ exps = Dict{String, Any}(
             s_infl  = 1.02,
             p_infl  = 1.0,
            ),
+          # IEEE39bus ETKF state estimation standard configuration
           "IEEE39bus_ETKF_state_test" => (
             time_series = path *
             "IEEE39bus_time_series_seed_0000_diff_0.000_tanl_0.01_nanl_05000_spin_1500_" *
             "h_0.010.jld2",
-            method="etkf",
-            seed=0,
-            nanl=3500,
-            obs_un=0.1,
-            obs_dim=20,
-            γ=1.00,
-            N_ens=21,
-            s_infl=1.02
+            method  = "etkf",
+            seed    = 0,
+            nanl    = 3500,
+            obs_un  = 0.1,
+            obs_dim = 20,
+            γ       = 1.00,
+            N_ens   = 21,
+            s_infl  = 1.02
            ),
           ),
[…]
