+++ src/models/L96.jl
@@ -24,9 +24,9 @@ end
 """
     dx_dt(x::VecA, t::Float64, dx_params::ParamDict)
 
-Time derivative for Lorenz-96 model, `x` is a single model state of size `state_dim`, `t`
-is a dummy time argument for consistency with integration methods, `dx_params` is of type
-`ParaDict` which is called for the forcing parameter.
+Time derivative for Lorenz-96 model, `x` is a  model state of size `state_dim` and type
+[`VecA`](@ref), `t` is a dummy time argument for consistency with integration methods,
+`dx_params` is of type [`ParamDict`](@ref) which is called for the forcing parameter.
 """
 function dx_dt(x::VecA, t::Float64, dx_params::ParamDict)
     # unpack the (only) derivative parameter for l96
@@ -52,15 +52,15 @@ end
 
 ##############################################################################################
 """
-    jacobian(x::Vector{Float64}, t::Float64, dx_params::ParamDict) 
+    jacobian(x::VecA, t::Float64, dx_params::ParamDict) 
     
-Computes the Jacobian of Lorenz-96 about the state `x`. The time variable `t` is a dummy
-variable for consistency with integration methods, `dx_params` is of type `ParamDict`
-which is called for the forcing parameter. Note that this is designed to load entries in
-a zeros array and return a sparse array to make a compromise between memory
-and computational resources.
+Computes the Jacobian of Lorenz-96 about the state `x` of type [`VecA`](@ref). The time
+variable `t` is a dummy variable for consistency with integration methods,
+`dx_params` is of type [`ParamDict`](@ref) which is called for the forcing parameter.
+Note that this is designed to load entries in a zeros array and return a sparse array to
+make a compromise between memory and computational resources.
 """
-function jacobian(x::Vector{Float64}, t::Float64, dx_params::ParamDict)
+function jacobian(x::VecA, t::Float64, dx_params::ParamDict)
 
     x_dim = length(x)
     dxF = zeros(x_dim, x_dim)
@@ -97,8 +97,8 @@ end
 Computes auxiliary functions for the 2nd order Taylor-Stratonovich expansion. The constants
 `α` and `ρ` need to be computed once, only as a function of the order of truncation of the
 Fourier series, the argument `p`, for the integration method.  These constants are then
-supplied as arguments to `l96s_tay2_step!` in `kwargs`. See `l96s_tay2_step!` for the
-interpretation and usage of these constants.
+supplied as arguments to `l96s_tay2_step!` in `kwargs`. See [`l96s_tay2_step!`](@ref) for
+the interpretation and usage of these constants.
 """
 function compute_α_ρ(p::Int64)
     function α(p::Int64)
@@ -113,24 +113,23 @@ end
 
 ##############################################################################################
 """
-    l96s_tay2_step!(x::Vector{Float64}, t::Float64, kwargs::Dict{String,Any}) 
+    l96s_tay2_step!(x::VecA, t::Float64, kwargs::Dict{String,Any}) 
     
 One step of integration rule for l96 second order taylor rule
 The constants `ρ` and `α` are to be computed with `compute_α_ρ`, depending
 only on `p`, and supplied for all steps. This is the general formulation which includes,
 e.g., dependence on the truncation of terms in the auxilliary function `C` with
 respect to the parameter `p`.  In general, truncation at `p=1` is all that is
-necessary for order 2.0 convergence, and in this case `C` below is identically
-equal to zero.  This auxilliary function can be removed (and is removed) in other
-implementations for simplicity.
+necessary for order 2.0 convergence.
 
[…]
