+++ docs/make.jl
@@ -25,6 +25,7 @@ makedocs(
                     "GenerateTimeSeries" => "submodules/experiments/GenerateTimeSeries.md",
                     "FilterExps" => "submodules/experiments/FilterExps.md",
                     "SmootherExps" => "submodules/experiments/SmootherExps.md",
+                    "SingleExperimentDriver" => "submodules/experiments/SingleExperimentDriver.md",
                     "Slurm" => "submodules/experiments/Slurm.md",
                    ],
                  "Analysis" => Any[
@@ -36,6 +37,6 @@ makedocs(
 )
 
 
-#deploydocs(
-#    repo = "github.com/cgrudz/DataAssimilationBenchmarks.jl.git",
-#)
+deploydocs(
+    repo = "github.com/cgrudz/DataAssimilationBenchmarks.jl.git",
+)
