+++ docs/src/submodules/experiments/ParallelExperimentDriver.md
@@ -8,6 +8,6 @@ collection will cancel.  A fix for this is being explored, but the recommendatio
 scripts below as templates for generating large parameter grid configurations and running them on servers.
 
 ## Docstrings
-```@autodocs
-Modules = [DataAssimilationBenchmarks.SingleExperimentDriver]
-```
+#```@autodocs
+#Modules = [DataAssimilationBenchmarks.ParallelExperimentDriver]
+#```
