+++ docs/src/submodules/experiments/FilterExps.md
@@ -1,7 +1,7 @@
 # FilterExps 
 
 
-### FilterExps API
+## FilterExps API
 The `FilterExps.jl` sub-module configures twin experiments using stored time series data as generated above for 
 efficiency when using the same base-line time series to generate possibly different experiment configurations.
 Experiment configurations are generated by

[…]
