+++ README.md
@@ -8,17 +8,34 @@
 [![Build Status](https://app.travis-ci.com/cgrudz/DataAssimilationBenchmarks.jl.svg?branch=master)](https://app.travis-ci.com/cgrudz/DataAssimilationBenchmarks.jl)
 [![codecov](https://codecov.io/gh/cgrudz/DataAssimilationBenchmarks.jl/branch/master/graph/badge.svg?token=3XLYTH8YSZ)](https://codecov.io/gh/cgrudz/DataAssimilationBenchmarks.jl)
 
+Lines of code counter (without comments or blank lines) courtesy of [Tokei](https://github.com/XAMPPRocky/tokei).
+
 ## Welcome to DataAssimilationBenchmarks.jl!
 
 ### Description
-This is my personal data assimilation benchmark research code with an emphasis on testing and validation
-of ensemble-based filters and sequential smoothers in chaotic toy models.  The code is meant to be performant, 
-in the sense that large hyper-parameter discretizations can be explored to determine structural sensitivity 
-and reliability of results across different experimental regimes, with parallel implementations in Slurm.
-This includes code for developing and testing data assimilation schemes in the 
-[L96-s model](https://gmd.copernicus.org/articles/13/1903/2020/) currently, with further models in development.
-This project supported the development of all numerical results and benchmark simulations considered in the pre-print
-[A fast, single-iteration ensemble Kalman smoother for sequential data assimilation](https://gmd.copernicus.org/preprints/gmd-2021-306/)
+
+This is a data assimilation research code base with an emphasis on prototyping, testing and
+validating sequential filters and smoothers in toy model twin experiments.
+This code is meant to be performant in the sense that large hyper-parameter discretizations
+can be explored to determine hyper-parameter sensitivity and reliability of results across
+different experimental regimes, with parallel implementations in native Julia distributed
+computing and using workload managers such as Slurm.
+
+This package currently includes code for developing and testing data assimilation schemes in
+the [L96-s model](https://gmd.copernicus.org/articles/13/1903/2020/) and the IEEE 39 bus test
+case in the form of the [effective network
+model](https://iopscience.iop.org/article/10.1088/1367-2630/17/1/015012)
+model equations. New toy models and data assimilation schemes are in continuous development
+in the development branch.  Currently validated techniques are available in the master
+branch.
+
+This package supported the development of all numerical results and benchmark simulations
+in the pre-print
+[A fast, single-iteration ensemble Kalman smoother for sequential data
+assimilation](https://gmd.copernicus.org/preprints/gmd-2021-306/)
 available currently in open review in Geoscientific Model Development.
 
-Lines of code counter (without comments or blank lines) courtesy of [Tokei](https://github.com/XAMPPRocky/tokei).
+### Documentation
+
+Please see the [github pages site](https://cgrudz.github.io/DataAssimilationBenchmarks.jl/dev/).
+
