+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"submodules/methods/EnsembleKalmanSchemes/#Ensemble-Kalman-Schemes","page":"EnsembleKalmanSchemes","title":"Ensemble Kalman Schemes","text":"","category":"section"},{"location":"submodules/methods/EnsembleKalmanSchemes/#API-for-data-assimilation-solvers","page":"EnsembleKalmanSchemes","title":"API for data assimilation solvers","text":"","category":"section"},{"location":"submodules/methods/EnsembleKalmanSchemes/","page":"EnsembleKalmanSchemes","title":"EnsembleKalmanSchemes","text":"Different filter and smoothing schemes are run through the routines including","category":"page"},{"location":"submodules/methods/EnsembleKalmanSchemes/","page":"EnsembleKalmanSchemes","title":"EnsembleKalmanSchemes","text":"ensemble_filter(analysis::String, ens::Array{Float64,2}, obs::Vector{Float64},\n                         obs_cov::CovM, state_infl::Float64, kwargs::Dict{String,Any})\n\nls_smoother_classic(analysis::String, ens::Array{Float64,2}, obs::Array{Float64,2},\n                             obs_cov::CovM, state_infl::Float64, kwargs::Dict{String,Any})\n\nls_smoother_single_iteration(analysis::String, ens::Array{Float64,2}, obs::Array{Float64,2},\n                             obs_cov::CovM, state_infl::Float64, kwargs::Dict{String,Any})\n\n\nls_smoother_gauss_newton(analysis::String, ens::Array{Float64,2}, obs::Array{Float64,2},\n                             obs_cov::CovM, state_infl::Float64, kwargs::Dict{String,Any};\n                             ϵ::Float64=0.0001, tol::Float64=0.001, max_iter::Int64=10)\n\n# type union for multiple dispatch over specific types of covariance matrices\nCovM = Union{UniformScaling{Float64}, Diagonal{Float64}, Symmetric{Float64}}\n\nanalysis   -- string of the DA scheme string name, given to the transform sub-routine in\n              methods\nens        -- ensemble matrix defined by the array with columns given by the replicates of\n              the model state\nobs        -- observation vector for the current analysis in ensemble_filter / array with\n              columns given by the observation vectors for the ordered sequence of analysis\n\t\t\t\t\t\t\ttimes in the current smoothing window\nobs_cov    -- observation error covariance matrix, must be positive definite of type CovM\nstate_infl -- multiplicative covariance inflation factor for the state variable covariance\n              matrix, set to one this is the standard Kalman-like update\nkwargs     -- keyword arguments for parameter estimation or other functionality, including\n              integration parameters for the state model in smoothing schemes","category":"page"},{"location":"submodules/methods/EnsembleKalmanSchemes/","page":"EnsembleKalmanSchemes","title":"EnsembleKalmanSchemes","text":"The type of analysis to be passed to the transform step is specified with the analysis string.  Observations for the filter schemes correspond to information available at a single analysis time while the ls (lag-shift) smoothers require an array of observations corresponding to all analysis times within the DAW.  Observation covariances should be typed according to the type union above for efficiency.  The state_infl is a required tuneable parameter for multiplicative covariance inflation.   Extended parameter state covariance inflation can be specified
[…]
