### 2022-05-13 #4129 @cgrudz
@taless474 , thanks for checking in.  Things are currently in-progress where I am about ready to start clearing some of the issues ([new doc pages](https://cgrudz.github.io/DataAssimilationBenchmarks.jl/dev/)!), and I am expecting to have a summer research assistant who will be working with the software and will help with the revision and new features for the next release.  I am anticipating resolving the major items and submitting a revision after the research internship has completed around the end of summer, but I'll post updates as I feel like issues / tickets have been cleared.  Cheers!
