### 2022-06-20 459ed34
Uploading further documention, smoothing out import conventions
