+++ docs/make.jl
@@ -37,6 +37,6 @@ makedocs(
 )
 
 
-deploydocs(
-    repo = "github.com:cgrudz/DataAssimilationBenchmarks.jl.git",
-)
+#deploydocs(
+#    repo = "github.com:cgrudz/DataAssimilationBenchmarks.jl.git",
+#)
