+++ src/models/IEEE39bus.jl
@@ -8,7 +8,7 @@ export dx_dt
 """
     dx_dt(x::VecA, t::Float64, dx_params::ParamDict) 
     
-Time derivative of the phase and fequency of the effective-network swing equation model.
+Time derivative of the phase and fequency of the [effective-network swing equation model](https://iopscience.iop.org/article/10.1088/1367-2630/17/1/015012).
 Input x is a 2 `n_g` [`VecA`](@ref) of the phase and fequency at each of the `n_g`
 generator buses. The input `dx_params` of type [`ParamDict`](@ref) containing system
 parameters to be passed to the integration scheme.  The system is currenty defined
