+++ src/methods/EnsembleKalmanSchemes.jl
@@ -15,17 +15,38 @@ export alternating_obs_operator, analyze_ens, analyze_ens_para, rand_orth,
 ##############################################################################################
 """
     alternating_obs_operator(ens::Array{Float64,2}, obs_dim::Int64, 
-                             kwargs::Dict{String,Any})
+                             kwargs::StepKwargs)
 
-Observation of alternating state vector components, possibly nonlinear transformation.
-This selects components to observe based on the observation dimension and if
-parameter estimation is being performed.  Parameters are always unobservable,
-and even states will be removed from the state vector until the observation dimension
-is appropriate.  Nonlinear observations are optional, as described for the
-Lorenz-96 model by Asch, Bocquet, Nodet pg. 181.
+This produces observations of alternating state vector components for generating pseudo-data.
+
+This operator can take either a truth twin, single model state or an ensemble of states, and
+map this data to the observation space.  The truth twin is assumed to be 2D, where the
+first index corresponds to the state dimension and the second index corresponds to the time
+dimension.  The ensemble is assumed to be 2D where the first index corresponds to the state
+dimension and the second index corresponds to the ensemble dimension.  The operator selects
+components of the state dimension to observe based on the observation dimension, adjusting
+if parameter estimation is being performed.  Model parameters are always assumed
+unobservable, and statistical replicates of model parameters in the ensemble will be first
+truncated out of the ensemble matrix before mapping the state vectors to the observation
+space. States correpsonding to even state dimension indices are removed from the state
+vector until the observation dimension is appropriate.  If the observation dimension is
+less than half the state dimension, states corresponding to odd state dimension idices
+are subsequently removed until the observation dimension is appropriate.
+
+The `γ` parameter required in `kwargs` of type  [`StepKwargs`](@ref) controls the
+component-wise transformation of the remaining state vector components mapped to the
+observation space.  For `γ=1`, there is no transformation applied, and the observation
+operator acts as a linear projection onto the remaining components of the state vector.
+For `γ>1.0`, the nonlinear observation operator of 
+[Asch, M., et al. (2016).](https://epubs.siam.org/doi/book/10.1137/1.9781611974546),
+pg. 181 is applied, which limits to the identity for `γ=1.0`.  If `γ=0.0`, the quadratic
+observation operator of [Hoteit, I., et al. (2012).](https://journals.ametsoc.org/view/journals/mwre/140/2/2011mwr3640.1.xml)
+is applied to the remaining state components.  If `γ<0.0`, the exponential observation
+operator of [Wu, G., et al. (2014).](https://npg.copernicus.org/articles/21/955/2014/)
+is applied to the remaining state vector components.
 """
 function alternating_obs_operator(ens::Array{Float64,2}, obs_dim::Int64,
-                                  kwargs::Dict{String,Any})
+                                  kwargs::StepKwargs)
     sys_dim, N_ens = size(ens)
 
     if haskey(kwargs, "state_dim")
@@ -66,17 +87,15 @@ function alternating_obs_operator(ens::Array{Float64,2}, obs_dim::Int64,
         
     γ = kwargs["gamma"]::Float64
     if γ > 1.0
-        # sets nonlinear observation as given on page 181, Asch, Bocquet, Nodet
         for i in 1:N_ens
             x = obs[:, i]
             obs[:, i]  = (x / 2.0) .* ( 1.0 .+ ( abs.(x) / 10.0 ).^(γ - 1.0) )
         end
+
[…]
