### 2022-06-21 a793b7a
New documentation on the software and smoothing of arguments
