+++ src/methods/EnsembleKalmanSchemes.jl
@@ -38,11 +38,11 @@ component-wise transformation of the remaining state vector components mapped to
 observation space.  For `γ=1`, there is no transformation applied, and the observation
 operator acts as a linear projection onto the remaining components of the state vector.
 For `γ>1.0`, the nonlinear observation operator of 
-[Asch, M., et al. (2016).](https://epubs.siam.org/doi/book/10.1137/1.9781611974546),
+[Asch, et al. (2016).](https://epubs.siam.org/doi/book/10.1137/1.9781611974546),
 pg. 181 is applied, which limits to the identity for `γ=1.0`.  If `γ=0.0`, the quadratic
-observation operator of [Hoteit, I., et al. (2012).](https://journals.ametsoc.org/view/journals/mwre/140/2/2011mwr3640.1.xml)
+observation operator of [Hoteit, et al. (2012).](https://journals.ametsoc.org/view/journals/mwre/140/2/2011mwr3640.1.xml)
 is applied to the remaining state components.  If `γ<0.0`, the exponential observation
-operator of [Wu, G., et al. (2014).](https://npg.copernicus.org/articles/21/955/2014/)
+operator of [Wu, et al. (2014).](https://npg.copernicus.org/articles/21/955/2014/)
 is applied to the remaining state vector components.
 """
 function alternating_obs_operator(ens::Array{Float64,2}, obs_dim::Int64,
@@ -370,36 +370,36 @@ Currently validated `analysis` options:
  * `analysis=="enkf" || analysis=="enks"` computes the stochastic transform for the EnKF/S
    as in [Carrassi, et al. 2018](https://wires.onlinelibrary.wiley.com/doi/10.1002/wcc.535).
  * `analysis=="etkf" || analysis=="etks"` computes the deterministic ensemble transform 
-   as in the ETKF described in [Grudzien et al.
+   as in the ETKF described in [Grudzien, et al.
    2021](https://gmd.copernicus.org/preprints/gmd-2021-306/).
  * `analysis[1:7]=="mlef-ls" || analysis[1:7]=="mles-ls"` computes the maximum likelihood
-   ensemble filter transform described in [Grudzien et al. 
+   ensemble filter transform described in [Grudzien, et al. 
    2021](https://gmd.copernicus.org/preprints/gmd-2021-306/), optimizing the nonlinear
    cost function with Newton-based 
    [line searches](https://julianlsolvers.github.io/LineSearches.jl/stable/).
  * `analysis[1:4]=="mlef" || analysis[1:4]=="mles"` computes the maximum likelihood     
    ensemble filter transform described in
-   [Grudzien et al. 2021](https://gmd.copernicus.org/preprints/gmd-2021-306/),
+   [Grudzien, et al. 2021](https://gmd.copernicus.org/preprints/gmd-2021-306/),
    optimizing the nonlinear
    cost function with simple Newton-based scheme. 
  * `analysis=="enkf-n-dual" || analysis=="enks-n-dual"` 
-   computes the dual form of the EnKF-N transform as in [Bocquet et al.
+   computes the dual form of the EnKF-N transform as in [Bocquet, et al.
    2015](https://npg.copernicus.org/articles/22/645/2015/)
    Note: this cannot be used with the nonlinear observation operator.
    This uses the Brent method for the argmin problem as this
    has been more reliable at finding a global minimum than Newton optimization.
  * `analysis=="enkf-n-primal" || analysis=="enks-n-primal"`
-   computes the primal form of the EnKF-N transform as in [Bocquet et al.
+   computes the primal form of the EnKF-N transform as in [Bocquet, et al.
    2015](https://npg.copernicus.org/articles/22/645/2015/),
[…]
