### 2022-06-27 1dce432
Updating project dependencies to include auto differentiation
