### 2022-06-28 2cb2df4
Smoothing variable types for more automatic diff
