+++ src/methods/EnsembleKalmanSchemes.jl
@@ -14,8 +14,7 @@ export alternating_obs_operator, analyze_ens, analyze_ens_para, rand_orth,
 # Main methods, debugged and validated
 ##############################################################################################
 """
-    alternating_obs_operator(ens::Array{Float64,2}, obs_dim::Int64, 
-                             kwargs::StepKwargs)
+    alternating_obs_operator(ens::ArView, obs_dim::Int64, kwargs::StepKwargs)
 
 This produces observations of alternating state vector components for generating pseudo-data.
 
@@ -45,8 +44,7 @@ is applied to the remaining state components.  If `γ<0.0`, the exponential obse
 operator of [Wu, et al. (2014).](https://npg.copernicus.org/articles/21/955/2014/)
 is applied to the remaining state vector components.
 """
-function alternating_obs_operator(ens::Array{Float64,2}, obs_dim::Int64,
-                                  kwargs::StepKwargs)
+function alternating_obs_operator(ens::ArView, obs_dim::Int64, kwargs::StepKwargs)
     sys_dim, N_ens = size(ens)
 
     if haskey(kwargs, "state_dim")
@@ -107,7 +105,7 @@ end
 
 ##############################################################################################
 """
-    analyze_ens(ens::ArView, truth::Vector{Float64}) 
+    analyze_ens(ens::ArView, truth::VecA) 
 
 Computes the ensemble state RMSE as compared with truth twin, and the ensemble spread.
 
@@ -115,7 +113,7 @@ Note: the ensemble `ens` should only include the state vector components to comp
 truth twin state vector `truth`, without replicates of the model parameters.  These can be
 passed as an [`ArView`](@ref) for efficient memory usage.
 """
-function analyze_ens(ens::ArView, truth::Vector{Float64})
+function analyze_ens(ens::ArView, truth::VecA)
 
     # infer the shapes
     sys_dim, N_ens = size(ens)
@@ -137,7 +135,7 @@ end
 
 ##############################################################################################
 """
-    analyze_ens_para(ens::ArView, truth::Vector{Float64}) 
+    analyze_ens_para(ens::ArView, truth::VecA) 
 
 Computes the ensemble parameter RMSE as compared with truth twin, and the ensemble spread.
 
@@ -145,7 +143,7 @@ Note: the ensemble `ens` should only include the extended state vector component
 parameter replicates to compare with the truth twin's governing model parameters `truth`. 
 These can be passed as an [`ArView`](@ref) for efficient memory usage.
 """
-function analyze_ens_para(ens::ArView, truth::Vector{Float64})
+function analyze_ens_para(ens::ArView, truth::VecA)
 
     # infer the shapes
     param_dim, N_ens = size(ens)
@@ -199,8 +197,7 @@ end
 
 ##############################################################################################
 """
-    inflate_state!(ens::Array{Float64,2}, inflation::Float64, sys_dim::Int64, 
-                   state_dim::Int64)
+    inflate_state!(ens::ArView, inflation::Float64, sys_dim::Int64, state_dim::Int64)
 
 Applies multiplicative covariance inflation to the state components of the ensemble matrix.
 
@@ -210,8 +207,7 @@ variables are assumed to be in the leading `state_dim` rows of `ens`, while exte
 parameter replicates are after. Multiplicative inflation is performed only in the leading
 components of the ensemble anomalies from the ensemble mean, in-place in memory.
 """
-function inflate_state!(ens::Array{Float64,2}, inflation::Float64, sys_dim::Int64,
-                        state_dim::Int64)
+function inflate_state!(ens::ArView, inflation::Float64, sys_dim::Int64, state_dim::Int64)
     if inflation != 1.0
[…]
