### 2022-07-01 d36b141
Multi-dispact of observation operator
Making multi-dispatch of the observation operator to accomodate
vector imputs instead of arrays -- also made the function robust
to missing gamma parameter (renamed \gamma) for when  a linear operator,
equivalent to \gamma=1.0, is desired but \gamma is unspecified.
