### 2022-07-02 e0f8cd1
Testing new type structure to see if this gets beter performance
Going to test several different forms to see where we can balance
speed and compatibility with auto differentiation.
