+++ src/methods/EnsembleKalmanSchemes.jl
@@ -40,7 +40,7 @@ is applied to the remaining state components.  If `γ<0.0`, the exponential obse
 operator of [Wu, et al. (2014).](https://npg.copernicus.org/articles/21/955/2014/)
 is applied to the remaining state vector components.
 """
-function alternating_obs_operator(x::VecA, obs_dim::Int64, kwargs::StepKwargs)
+function alternating_obs_operator(x::VecA(T), obs_dim::Int64, kwargs::StepKwargs) where T <: Real
     sys_dim = length(x)
 
     if haskey(kwargs, "state_dim")
@@ -131,7 +131,7 @@ is applied to the remaining state components.  If `γ<0.0`, the exponential obse
 operator of [Wu, et al. (2014).](https://npg.copernicus.org/articles/21/955/2014/)
 is applied to the remaining state vector components.
 """
-function alternating_obs_operator(ens::ArView, obs_dim::Int64, kwargs::StepKwargs)
+function alternating_obs_operator(ens::ArView(T), obs_dim::Int64, kwargs::StepKwargs) where T <: Real
     sys_dim, N_ens = size(ens)
 
     if haskey(kwargs, "state_dim")
@@ -202,7 +202,7 @@ Note: the ensemble `ens` should only include the state vector components to comp
 truth twin state vector `truth`, without replicates of the model parameters.  These can be
 passed as an [`ArView`](@ref) for efficient memory usage.
 """
-function analyze_ens(ens::ArView, truth::VecA)
+function analyze_ens(ens::ArView(T), truth::VecA(T)) where T <: Real
 
     # infer the shapes
     sys_dim, N_ens = size(ens)
@@ -232,7 +232,7 @@ Note: the ensemble `ens` should only include the extended state vector component
 parameter replicates to compare with the truth twin's governing model parameters `truth`. 
 These can be passed as an [`ArView`](@ref) for efficient memory usage.
 """
-function analyze_ens_para(ens::ArView, truth::VecA)
+function analyze_ens_para(ens::ArView(T), truth::VecA(T)) where T <: Real
 
     # infer the shapes
     param_dim, N_ens = size(ens)
@@ -296,7 +296,8 @@ variables are assumed to be in the leading `state_dim` rows of `ens`, while exte
 parameter replicates are after. Multiplicative inflation is performed only in the leading
 components of the ensemble anomalies from the ensemble mean, in-place in memory.
 """
-function inflate_state!(ens::ArView, inflation::Float64, sys_dim::Int64, state_dim::Int64)
+function inflate_state!(ens::ArView(T), inflation::Float64, sys_dim::Int64,
+                        state_dim::Int64) where T <: Real
     if inflation != 1.0
         x_mean = mean(ens, dims=2)
         X = ens .- x_mean
@@ -320,7 +321,8 @@ parameter replicates are after. Multiplicative inflation is performed only in th
 `state_dim + 1: state_dim` components of the ensemble anomalies from the ensemble mean,
 in-place in memory.
 """
-function inflate_param!(ens::ArView, inflation::Float64, sys_dim::Int64, state_dim::Int64)
+function inflate_param!(ens::ArView(T), inflation::Float64, sys_dim::Int64,
+                        state_dim::Int64) where T <: Real
     if inflation == 1.0
         return ens
     else
@@ -344,11 +346,11 @@ of `UniformScaling` and `Diagonal` covariance matrices are computed directly, wh
 square roots of  the more general class of `Symmetric` covariance matrices are computed via
 the singular value decomposition, for stability and accuracy for close-to-singular matrices.
 """
-function square_root(M::T) where {T <: CovM}
[…]
