### 2022-07-02 69e1e38
Fixing type constructors for better multiple dispatch / single dispatch
