### 2022-07-02 d70584d
Fixing issues with typing for auto diff compatability
