+++ .travis.yml
@@ -4,7 +4,6 @@ julia:
 codecov: true
 after_success:
 - julia -e 'using Pkg; cd(Pkg.dir("DataAssimilationBenchmarks")); Pkg.add("Coverage"); using Coverage; Codecov.submit(Codecov.process_folder())'
-
 jobs:
   include:
     - stage: "Documentation"
