+++ .travis.yml
@@ -2,8 +2,7 @@ language: julia
 julia:
   - 1.7
 codecov: true
-after_success:
-- julia -e 'using Pkg; cd(Pkg.dir("DataAssimilationBenchmarks")); Pkg.add("Coverage"); using Coverage; Codecov.submit(Codecov.process_folder())'
+
 jobs:
   include:
     - stage: "Documentation"
@@ -14,3 +13,5 @@ jobs:
                                                Pkg.instantiate()'
         - julia --project=docs/ docs/make.jl
       after_success: skip
+        - julia -e 'using Pkg; cd(Pkg.dir("DataAssimilationBenchmarks")); Pkg.add("Coverage"); using Coverage; Codecov.submit(Codecov.process_folder())'
+
