+++ docs/make.jl
@@ -36,7 +36,7 @@ makedocs(
             ]
 )
 
-#
-#deploydocs(
-#    repo = "github.com:cgrudz/DataAssimilationBenchmarks.jl.git",
-#)
+
+deploydocs(
+    repo = "github.com:cgrudz/DataAssimilationBenchmarks.jl.git",
+)
