+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"submodules/methods/EnsembleKalmanSchemes/#Ensemble-Kalman-Schemes","page":"EnsembleKalmanSchemes","title":"Ensemble Kalman Schemes","text":"","category":"section"},{"location":"submodules/methods/EnsembleKalmanSchemes/#API-for-data-assimilation-solvers","page":"EnsembleKalmanSchemes","title":"API for data assimilation solvers","text":"","category":"section"},{"location":"submodules/methods/EnsembleKalmanSchemes/","page":"EnsembleKalmanSchemes","title":"EnsembleKalmanSchemes","text":"There are currently four families of data assimilation solvers available in this package, which define the outer-loop of the data assimilation cycle.  Particularly, these define how the sequential data assimilation cycle will pass over a time series of observations.  Ensemble filters run only forward-in-time.  The classic lag-shift smoother runs identically to the filter in its forecast and filter steps, but includes an additional retrospective analysis to past ensemble states stored in memory.  The single iteration smoother follows the same convention as the classic smoother, except in that new cycles are initiated from a past, reanlyzed ensemble state.  The Gauss-Newton iterative smoothers are 4D smoothers, which iteratively optimize the initial condition at the beginning of a data assimilation cycle, and propagate this initial condition to initialize the subsequent cycle.  A full discussion of these methods can be found in Grudzien, et al. 2021.","category":"page"},{"location":"submodules/methods/EnsembleKalmanSchemes/","page":"EnsembleKalmanSchemes","title":"EnsembleKalmanSchemes","text":"For each outer-loop method defining the data assimilation cycle, different types of analyses can be specified within their arguments.  Likewise, these outer-loop methods require arguments such as the ensemble state or the range of ensemble states to analyze, an observation to assimilate or a range of observations to assimilate, as the observation error covariances, the ensemble covariance inflation parameter and key word arguments for running the underlying dynamical state model. Examples of the syntax are below:","category":"page"},{"location":"submodules/methods/EnsembleKalmanSchemes/","page":"EnsembleKalmanSchemes","title":"EnsembleKalmanSchemes","text":"ensemble_filter(analysis::String, ens::Array{Float64,2}, obs::Vector{Float64},\n                         obs_cov::CovM, state_infl::Float64, kwargs::StepKwargs)\n\nls_smoother_classic(analysis::String, ens::Array{Float64,2}, obs::Array{Float64,2},\n                             obs_cov::CovM, state_infl::Float64, kwargs::StepKwargs)\n\nls_smoother_single_iteration(analysis::String, ens::Array{Float64,2}, obs::Array{Float64,2},\n                             obs_cov::CovM, state_infl::Float64, kwargs::StepKwargs)\n\nls_smoother_gauss_newton(analysis::String, ens::Array{Float64,2}, obs::Array{Float64,2},\n                             obs_cov::CovM, state_infl::Float64, kwargs::StepKwargs;\n                             ϵ::Float64=0.0001, tol::Float64=0.001, max_iter::Int64=10)\n\n\"\"\"\nanalysis   -- string name analysis scheme given to the transform sub-routine\nens        -- ensemble matrix defined by the array with columns given by the replicates of\n              the model state\nobs        -- observation vector for the current analysis in ensemble_filter / array with\n              columns given by the observation vectors for the ordered sequence of analysis\n\t\t\t\t\t\
[…]
