+++ .travis.yml
@@ -2,17 +2,14 @@ language: julia
 julia:
   - 1.7
 codecov: true
-  - julia -e 'using Pkg; cd(Pkg.dir("DataAssimilationBenchmarks")); Pkg.add("Coverage"); Pkg.test("MyPkg", coverage=true)'
-  after_success:
-    - julia -e 'using Codecov; Codecov.submit(Codecov.process_folder())'
-
 jobs:
   include:
     - stage: "Documentation"
       julia: 1.7
       os: linux
       script:
-        - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd()));
-                                               Pkg.instantiate()'
+        - julia --project=docs/ -e 'using Pkg; Pkg.instantiate(); Pkg.add(PackageSpec(path=pwd()))'
         - julia --project=docs/ docs/make.jl
       after_success: skip
+after_success:
+    - julia -e 'using Pkg; Pkg.add("Coverage"); using Coverage; Codecov.submit(process_folder())'
