### 2022-07-06 52cb925
Updating docs / types / multiple dispatch
Removed EnKF perturbed obs, not currently supporting this method,
updated types for better consistency and use with auto differentiation
and multiple dispatch.
