### 2022-07-07 dc92e2c
Bug fixes / removing closing type assertions / further refactor
Fixed bug in \gamma < 0.0 case for vector observation code, had artifact
from the multi array code.  Making the vector slice / projection a
devoted method for better maintainability, refactoring this for multi
dispatch.  Removed closing type assertions, causes unexpected behavior
with auto differentiation and slowdown of code in places.  Tuning
broadcasting operations in observation operator.
