+++ src/methods/EnsembleKalmanSchemes.jl
@@ -14,24 +14,109 @@ export alternating_obs_operator, analyze_ens, analyze_ens_param, rand_orth,
 # Main methods, debugged and validated
 ##############################################################################################
 """
+    alternating_projector!(x::VecA(T), obs_dim::Int64) where T <: Real
+
+Utility method produces a projection of alternating vector components via slicing.
+```
+return x
+```
+
+This operator takes a single model state `x` of type [`VecA`](@ref) and maps this data to
+alternating entries.  The operator selects components of the vector
+based on the observation dimension.  States correpsonding to even state dimension indices
+are removed from the state vector until the observation dimension is appropriate.
+If the observation dimension is less than half the state dimension, states corresponding
+to odd state dimension idices are subsequently removed until the observation dimension
+is appropriate.
+"""
+function alternating_projector!(x::VecA(T), obs_dim::Int64) where T <: Real
+    sys_dim = length(x)
+    if obs_dim == sys_dim
+
+    elseif (obs_dim / sys_dim) > 0.5
+        # the observation dimension is greater than half the state dimension, so we
+        # remove only the trailing odd-index rows equal to the difference
+        # of the state and observation dimension
+        R = sys_dim - obs_dim
+        indx = 1:(sys_dim - 2 * R)
+        indx = [indx; sys_dim - 2 * R + 2: 2: sys_dim]
+        x = x[indx]
+
+    elseif (obs_dim / sys_dim) == 0.5
+        # the observation dimension is equal to half the state dimension so we remove exactly
+        # half the rows, corresponding to those with even-index
+        x = x[1:2:sys_dim, :]
+
+    else
+        # the observation dimension is less than half of the state dimension so that we
+        # remove all even rows and then all but the remaining, leading obs_dim rows
+        x = x[1:2:sys_dim]
+        x = x[1:obs_dim]
+    end
+    return x
+end
+
+
+##############################################################################################
+"""
+    alternating_projector!(ens::ArView(T), obs_dim::Int64) where T <: Real
+
+Utility method produces a projection of alternating ensemble components in-place via slicing.
+```
+return ens
+```
+
+This operator takes either a truth twin time series or an ensemble of states of type
+[`ArView`](@ref), and maps this data to alternating row components.  The truth twin in
+this version is assumed to be 2D, where the first index corresponds to the state dimension
+and the second index corresponds to the time dimension.  The ensemble is assumed to be
+2D where the first index corresponds to the state dimension and the second index
+corresponds to the ensemble dimension.
+States correpsonding to even state dimension indices are removed from the state
+vector until the observation dimension is appropriate.  If the observation dimension is
+less than half the state dimension, states corresponding to odd state dimension idices
+are subsequently removed until the observation dimension is appropriate.
+"""
+function alternating_projector!(ens::ArView(T), obs_dim::Int64) where T <: Real
+    sys_dim, N_ens = size(ens)
+    if obs_dim == sys_dim
+
+    elseif (obs_dim / sys_dim) > 0.5
+        # the observation dimension is greater than half the state dimension, so we
+        # remove only the trailing odd-index rows equal to the difference
+        # of the state and observation dimension
+        R = sys_dim - obs_dim
[…]
