### 2022-07-07 c8bf28d
Bugfix for observation operator
Former version wasn't overwriting input in place, this fixes bug by
changing function style to operate entirely on copy of data,
 may consider reworking this later.
