### 2022-07-07 3dce6bc
Alternating Observation Operator Jacobian Function, Test, and New Model for Observation Operators
I created the function for explicitly calculating the jacobian of the Alternating Observation Operator, created an explicit observation operator module with all of the observation operator, and tested/verified its behavior using automatic differentiation in a test using ForwardDiff. Subfiles in the documentation have been updated accordingly.
