### 2022-07-07 3e5bd82
Break-Up Observation Operator Test
This update breaks up the observational operator test into 3 different tests, one for each gamma, to achieve adequate coverage.
