### 2022-07-11 115772c
Integrating observation models into API
Updating conventions in EnsembleKalmanSchemes for a unified observation
model module for both ensemble and VAR schemes.  Consolidating multiple
dispatch doc strings.  Making inflation arguments optional in Ensemble
Kalman schemes and based on keyword arguments in kwargs.  Preparing for
improved experiments API, when finally getting to a total revamp.
