+++ src/models/ObsOperators.jl
@@ -2,25 +2,32 @@
 module ObsOperators
 ##############################################################################################
 # imports and exports
-using Random, Distributions, Statistics
 using LinearAlgebra, SparseArrays
 using ..DataAssimilationBenchmarks
-using Optim, LineSearches, LinearAlgebra
 export alternating_projector, alternating_obs_operator, alternating_obs_operator_jacobian
 ##############################################################################################
 # Main methods
 ##############################################################################################
 """
     alternating_projector(x::VecA(T), obs_dim::Int64) where T <: Real
+    alternating_projector(ens::ArView(T), obs_dim::Int64) where T <: Real
 
-Utility method produces a projection of alternating vector components via slicing.
+Utility method produces a projection of alternating vector or ensemble components via slicing.
+
 ```
 return x
+return ens
 ```
 
-This operator takes a single model state `x` of type [`VecA`](@ref) and maps this data to
-alternating entries.  The operator selects components of the vector
-based on the observation dimension.  States correpsonding to even state dimension indices
+This operator takes a single model state `x` of type [`VecA`](@ref), a truth twin time series
+or an ensemble of states of type [`ArView`](@ref), and maps this data to alternating
+row components.  If truth twin is 2D, then the first index corresponds to the state dimension
+and the second index corresponds to the time dimension.  The ensemble is assumed to be
+2D where the first index corresponds to the state dimension and the second index
+corresponds to the ensemble dimension.
+
+The operator selects row components of the input to keep based on the `obs_dim`.
+States correpsonding to even state dimension indices
 are removed from the state vector until the observation dimension is appropriate.
 If the observation dimension is less than half the state dimension, states corresponding
 to odd state dimension idices are subsequently removed until the observation dimension
@@ -54,26 +61,6 @@ function alternating_projector(x::VecA(T), obs_dim::Int64) where T <: Real
 end
 
 
-##############################################################################################
-"""
-    alternating_projector(ens::ArView(T), obs_dim::Int64) where T <: Real
-
-Utility method produces a projection of alternating ensemble components in-place via slicing.
-```
-return ens
-```
-
-This operator takes either a truth twin time series or an ensemble of states of type
-[`ArView`](@ref), and maps this data to alternating row components.  The truth twin in
-this version is assumed to be 2D, where the first index corresponds to the state dimension
-and the second index corresponds to the time dimension.  The ensemble is assumed to be
-2D where the first index corresponds to the state dimension and the second index
-corresponds to the ensemble dimension.
-States correpsonding to even state dimension indices are removed from the state
-vector until the observation dimension is appropriate.  If the observation dimension is
-less than half the state dimension, states corresponding to odd state dimension idices
-are subsequently removed until the observation dimension is appropriate.
-"""
 function alternating_projector(ens::ArView(T), obs_dim::Int64) where T <: Real
     sys_dim, N_ens = size(ens)
     if obs_dim == sys_dim
@@ -105,15 +92,21 @@ end
 ##############################################################################################
 """
     alternating_obs_operator(x::VecA(T), obs_dim::Int64, kwargs::StepKwargs) where T <: Real
+    alternating_obs_operator(ens::ArView(T), obs_dim::Int64,
+                             kwargs::StepKwargs) where T <: Real
 
 This produces observations of alternating state vector components for generating pseudo-data.
 ```
 return obs
 ```
 
-This operator takes a single model state `x` of type [`VecA`](@ref) and maps this data to
[…]
