+++ docs/src/submodules/models/L96.md
@@ -1,5 +1,35 @@
 # Lorenz-96 model
+The classical form for the Lorenz-96 equations are defined as 
+$$\frac{\mathrm{d}\pmb{x}}{\mathrm{d} t} = \pmb{f}(\pmb{x}),$$
+where for each state component $i\in\{1,\cdots,n\}$,
 
+$$\begin{align}
+ f^i(\pmb{x}) &=-x^{i-2}x^{i-1} + x^{i-1}x^{i+1} - x^i + F
+\end{align}$$
+
+such that the components of the vector $\pmb{x}$ are given by the variables $x^i$ with
+periodic boundary conditions, $x^0=x^n$, $x^{-1}=x^{n-1}$ and $x^{n+1}=x^{1}$.  The term
+$F$ in the Lorenz-96 system is the forcing parameter that injects energy to the model.
+With the above definition for the classical Lorenz-96 equations, we define the L96-s model
+with additive noise as
+
+$$\begin{align}
+\frac{\mathrm{d} \pmb{x}}{\mathrm{d} t} = \pmb{f}(\pmb{x}) +
+ s(t)\mathbf{I}_{n}\pmb{W}(t),
+\end{align}$$
+
+where $\pmb{f}$ is defined as in the classical equations, $\mathbf{I}_n$ is the
+$n\times n$ identity matrix, $\pmb{W}(t)$ is an $n$-dimensional Wiener process and
+$s(t):\mathbb{R}\rightarrow \mathbb{R}$ is a measurable function of (possibly)
+time-varying diffusion coefficients. This model is analyzed in-depth for data assimilation
+twin experiments in the manuscript
+[Grudzien, C. et al. (2020).](https://gmd.copernicus.org/articles/13/1903/2020/gmd-13-1903-2020.html)
+and further details of using the system for data assimilation benchmarks in stochastic
+dynamics are discussed there.  The methods in the below define the model equations,
+higher-order dynamics such as the Jacobian, and the order 2.0 Taylor-Stratonovich
+scheme for the L96-s system.
+
+## Methods
 ```@autodocs
 Modules = [DataAssimilationBenchmarks.L96]
 ```
