+++ docs/src/submodules/models/IEEE39bus.md
@@ -1,5 +1,87 @@
 # IEEE39 Bus Swing Equation Model
 
+This is a version of the IEEE-39 bus test case as described by 
+[Nishikawa, T. et al.](https://iopscience.iop.org/article/10.1088/1367-2630/17/1/015012)
+The model, denoted the "effective network", consists of the ten generator buses in the
+network with all other buses eliminated by the classical Kron reduction.
+The power flow is described in steady state by a fixed point of the nonlinear swing equations
+
+$$\frac{2H_i}{\omega_\mathrm{R}} \ddot{\delta}_i + \frac{D_i}{\omega_\mathrm{R}} \dot{\delta}_i = A_{i}^\mathrm{EN} - \sum_{j =1, j\neq i}^{n_g} K_{ij}\sin\left(\delta_i - \delta_j -\gamma_{ij}^\mathrm{EN}\right),$$
+
+where we define each of the following:
+  * the angular reference frequency (in radians) about which the steady state synchronizes is defined as $\omega_\mathrm{R}$;
+  * the angle of rotation of the generator rotor at bus $i$, relative to the frame rotating at the reference frequency, is defined as $\delta_i(t)$;
+  * the difference between the reference frequency and the frequency of the rotor at bus $i$ is defined $\dot{\delta}_i(t)$;
+  * the rate of acceleration of the difference between the angle of the rotor at bus $i$ and the frame rotating at the reference frequency is defined as $\ddot{\delta}_i(t)$;
+  * the values of the inertia and damping at bus $i$ are defined as $H_i$ and $D_i$ respectively;
+  * the strength of the dynamical coupling of the buses $i$ and $j$ is defined as $K_{ij}$, while $\gamma_{ij}$ represents the phase shift involved in the coupling of these buses;
+  * the active power injected into the network by the generator at bus $i$ is represented by $A^\mathrm{EN}_i$; and 
+  * the number of generators in the network is defined as $n_g =10$.
+
+This model assumes constant, passive loads at each bus that draws power.
+The actual parameters used in the model are defined by files in the
+```
+DataAssimilationBenchmarks/src/models/IEEE39bus_inputs/
+```
+directory, taken from the configuration studied by 
+[Nishikawa, T. et al.](https://iopscience.iop.org/article/10.1088/1367-2630/17/1/015012),
+with details on their interpretation in section 4.1.
+
+The stochastic form in this code loosens the assumption of constant loads in this model by
+assuming that, at the time scale of interest, the draw of power fluctuates randomly about
+the constant level that defines the steady state. We introduce a Wiener process to 
+the above equations of the form  $s W_i(t)$, where $s$ is a parameter in the model
+controlling the relative diffusion level.  We assume that the fluctuations in the net
+power are uncorrelated across buses and that the diffusion in all buses is proportional to
+$s$. 
+
+Making a change of variables $\psi_i =  \dot{\delta}_i$, we recover the system of nonlinear
+SDEs,
+
+$$\dot{\delta}_i = \psi_i,$$
+
+$$\dot{\psi}_i = \frac{A^\mathrm{EN}_i \omega_\mathrm{R}}{2H_i} - \frac{D_i}{2H_i} \psi_i 
+ - \sum_{j=1,j\neq i}^{n_g} \frac{K_{ij}^\mathrm{EN}\omega_\mathrm{R}}{2H_i} \sin\left(\delta_i - \delta_j -\gamma_{ij}^\mathrm{EN}\right) + \frac{ s \omega_R}{2 H_i} \mathrm{d}W_i(t).$$
+
[…]
