### 2022-07-12 c47ae22
3d-VAR cost function and a test
