+++ docs/src/submodules/methods/XdVAR.md
@@ -1,0 +1,11 @@
+# XdVAR
+
+```{julia}
+D3_var_cost(x::VecA(T), obs::VecA(T), x_background::VecA(T), state_cov::CovM(T), 
+    H_obs::Function, obs_cov::CovM(T), kwargs::StepKwargs) where T <: Real
+```
+## Methods
+
+```@autodocs
+Modules = [DataAssimilationBenchmarks.XdVAR]
+```
