### 2022-07-12 2888bf5
Simple Gradient Test for 3D-VAR
