+++ src/methods/XdVAR.jl
@@ -11,7 +11,6 @@ using ..DataAssimilationBenchmarks
     D3_var_cost(x::VecA(T), obs::VecA(T), x_background::VecA(T), state_cov::CovM(T), 
     obs_cov::CovM(T), kwargs::StepKwargs) where T <: Real
 """
-
 function D3_var_cost(x::VecA(T), obs::VecA(T), x_background::VecA(T), state_cov::CovM(T), 
     H_obs::Function, obs_cov::CovM(T), kwargs::StepKwargs) where T <: Real
     # initializations
