### 2022-07-14 dcf7899
Updating experiments API, working on integration of EnKF and VAR exps
Pushing forward changes to make experiment API more general / consistent
to accomodate VAR experiments in the same framework as EnKF experiments.
More generalization is needed, but overall approach is to utilize
NamedTuples to configure hyperparameters for established experiment
routines.
