+++ src/experiments/SmootherExps.jl
@@ -7,15 +7,16 @@ using LinearAlgebra
 using JLD2, HDF5
 using ..DataAssimilationBenchmarks, ..ObsOperators, ..EnsembleKalmanSchemes, ..DeSolvers,
       ..L96, ..IEEE39bus
-export classic_state, classic_param, single_iteration_state, single_iteration_param,
-       iterative_state, iterative_param
+export classic_ensemble_state, classic_ensemble_param, single_iteration_ensemble_state,
+       single_iteration_ensemble_param,
+       iterative_ensemble_state, iterative_ensemble_param
 ##############################################################################################
 # Main smoothing experiments, debugged and validated for use with schemes in methods directory
 ##############################################################################################
 """
-    classic_state((time_series::String, method::String, seed::Int64, nanl::Int64, lag::Int64,
-                   shift::Int64, obs_un::Float64, obs_dim::Int64, γ::Float64, N_ens::Int64, 
-                   s_infl::Float64)::NamedTuple)
+    classic_ensemble_state((time_series::String, method::String, seed::Int64, nanl::Int64,
+                            lag::Int64, shift::Int64, obs_un::Float64, obs_dim::Int64,
+                            γ::Float64, N_ens::Int64, s_infl::Float64)::NamedTuple)
 
 Classic ensemble Kalman smoother state estimation twin experiment.  Twin experiment parameters
 such as the observation dimension, observation uncertainty, data assimilation method, number
@@ -77,12 +78,12 @@ where the file name is written dynamically according to the selected parameters 
              "_stateInfl_" * rpad(round(s_infl, digits=2), 4, "0") * 
              ".jld2"
 """
-function classic_state((time_series, method, seed, nanl, lag, shift, obs_un, obs_dim,
-                        γ, N_ens, s_infl)::NamedTuple{
-                      (:time_series,:method,:seed,:nanl,:lag,:shift,:obs_un,:obs_dim,  
-                       :γ,:N_ens,:s_infl),
-                      <:Tuple{String,String,Int64,Int64,Int64,Int64,Float64,Int64,
-                              Float64,Int64,Float64}})
+function classic_ensemble_state((time_series, method, seed, nanl, lag, shift, obs_un, obs_dim,
+                                 γ, N_ens, s_infl)::NamedTuple{
+                               (:time_series,:method,:seed,:nanl,:lag,:shift,:obs_un,:obs_dim,
+                                :γ,:N_ens,:s_infl),
+                               <:Tuple{String,String,Int64,Int64,Int64,Int64,Float64,Int64,
+                                       Float64,Int64,Float64}})
     # time the experiment
     t1 = time()
 
@@ -293,10 +294,10 @@ end
 
 ##############################################################################################
 """
-    classic_param((time_series::String, method::String, seed::Int64, nanl::Int64, lag::Int64,
-                   shift::Int64, obs_un::Float64, obs_dim::Int64, γ::Float64, p_err::Float64,
-                   p_wlk::Float64, N_ens::Int64, s_infl::Float64,
-                   s_infl::Float64})::NamedTuple)
+    classic_ensemble_param((time_series::String, method::String, seed::Int64, nanl::Int64,
+                            lag::Int64, shift::Int64, obs_un::Float64, obs_dim::Int64,
+                            γ::Float64, p_err::Float64, p_wlk::Float64, N_ens::Int64,
+                            s_infl::Float64, s_infl::Float64})::NamedTuple)
 
 Classic ensemble Kalman smoother joint state-parameter estimation twin experiment.  Twin
 experiment parameters such as the observation dimension, observation uncertainty, data
[…]
