### 2022-07-15 685668b
Further revising docs / statement of purpose to include VAR
