+++ docs/src/submodules/experiments/FilterExps.md
@@ -9,7 +9,7 @@ when reading in the data.
 
 Filter experiment configurations are generated by supplying a `NamedTuple` with the required fields
 as specified in the experiment method.  Conventions for these arguments are as follows:
-  * `time_series` - the path to the `.jld2` truth twin data set;
+  * `time_series` - the path and file name of the `.jld2` truth twin data set;
   * `method` - the sub-method analysis scheme string name;
   * `seed` - the pseudo-random seed that will define, e.g., the observation noise sequence;
   * `nanl` - the number of observation / analysis times to produce a posterior estimate;

[…]
