### 2022-07-19 036552d
Gauss-Newton Optimization
I implemented Gauss-Newton optimization for XdVAR in the 3d case and created a test.
