+++ src/methods/XdVAR.jl
@@ -4,6 +4,7 @@ module XdVAR
 # imports and exports
 using LinearAlgebra, SparseArrays
 using ..DataAssimilationBenchmarks
+using ForwardDiff, ReverseDiff
 ##############################################################################################
 # Main methods
 ##############################################################################################
@@ -25,6 +26,88 @@ function D3_var_cost(x::VecA(T), obs::VecA(T), x_background::VecA(T), state_cov:
     return J
 end
 
+
+##############################################################################################
+"""
+    D3_var_grad(x::VecA(T), obs::VecA(T), x_background::VecA(T), state_cov::CovM(T), 
+    obs_cov::CovM(T), kwargs::StepKwargs) where T <: Real
+"""
+
+function D3_var_grad(x::VecA(T), obs::VecA(T), x_background::VecA(T), state_cov::CovM(T), 
+    H_obs::Function, obs_cov::CovM(T), kwargs::StepKwargs) where T <: Real
+    # initializations
+    function wrap_cost(x)
+        XdVAR.D3_var_cost(x, obs, x_background, state_cov, H_obs, obs_cov, kwargs)
+    end
+
+    grad = ForwardDiff.gradient(wrap_cost, x)
+    return grad
+end
+
+
+##############################################################################################
+"""
+    D3_var_hessian(x::VecA(T), obs::VecA(T), x_background::VecA(T), state_cov::CovM(T), 
+    obs_cov::CovM(T), kwargs::StepKwargs) where T <: Real
+"""
+
+function D3_var_hessian(x::VecA(T), obs::VecA(T), x_background::VecA(T), state_cov::CovM(T), 
+    H_obs::Function, obs_cov::CovM(T), kwargs::StepKwargs) where T <: Real
+    
+    function wrap_cost(x)
+        XdVAR.D3_var_cost(x, obs, x_background, state_cov, H_obs, obs_cov, kwargs)
+    end
+
+    hess = ForwardDiff.hessian(wrap_cost, x)
+
+    return hess
+end
+
+##############################################################################################
+"""
+    D3_var_hessian(x::VecA(T), obs::VecA(T), x_background::VecA(T), state_cov::CovM(T), 
+    obs_cov::CovM(T), kwargs::StepKwargs) where T <: Real
+"""
+
+function D3_var_NewtonOp(x::VecA(T), obs::VecA(T), x_background::VecA(T), state_cov::CovM(T), 
+    H_obs::Function, obs_cov::CovM(T), kwargs::StepKwargs) where T <: Real
+    # initializations
+    j_max = 40
+    tol = 0.001
+    j = 0
+    sys_dim = length(x)
+
+    # gradient preallocation over-write 
+    function grad!(g::VecA(T), x::VecA(T)) where T <: Real
+        g[:] = D3_var_grad(x, obs, x_background, state_cov, H_obs, obs_cov, kwargs)
+    end
+
+    # hessian preallocation over-write
+    function hess!(h::ArView(T), x::VecA(T)) where T <: Real
+        h .= inv(D3_var_hessian(x, obs, x_background, state_cov, H_obs, obs_cov, kwargs))
+    end
+
+    # step 6: perform the optimization by simple Newton
+    grad_x = Array{Float64}(undef, (sys_dim))
+    hess_x = Array{Float64}(undef, (sys_dim, sys_dim))
+
+    while j < j_max
+        # compute the gradient and hessian
+        grad!(grad_x, x)
+        hess!(hess_x, x)
+        
+        # perform Newton approximation
+        Δx = hess_x*grad_x
+        x = x - Δx
+
+        if norm(Δx) < tol
+            break
+        else
+            j+=1
+        end
+    end
+    return x
+end
 
 ##############################################################################################
 # end module
