### 2022-07-19 18ecf8a
Updated Gauss-Newton Optimization
