### 2022-07-19 f4911bc
Update runtest.jl to reflect correct order given comments
