+++ docs/src/submodules/experiments/FilterExps.md
@@ -7,9 +7,11 @@ observation generating process.  The frequency of observations in continuous tim
 by the frequency of data saved in the time series and is inferred by the experiment
 when reading in the data.
 
-Filter experiment configurations are generated by supplying a `NamedTuple` with the required fields
-as specified in the experiment method.  Conventions for these arguments are as follows:
-  * `time_series` - the path and file name of the `.jld2` truth twin data set;
+Filter experiment configurations are generated by supplying a 
+[NamedTuple](https://docs.julialang.org/en/v1/base/base/#Core.NamedTuple)
+with the required fields as specified in the experiment method.  Conventions for
+these arguments are as follows:
+  * `time_series` - the path and file name of the [.jld2](https://juliaio.github.io/JLD2.jl/dev/) truth twin data set;
   * `method` - the sub-method analysis scheme string name;
   * `seed` - the pseudo-random seed that will define, e.g., the observation noise sequence;
   * `nanl` - the number of observation / analysis times to produce a posterior estimate;
