### 2022-07-19 3929a0c
Basic code cleanup and deploying docs
