+++ docs/make.jl
@@ -39,6 +39,6 @@ makedocs(
 )
 
 
-#deploydocs(
-#    repo = "github.com:cgrudz/DataAssimilationBenchmarks.jl.git",
-#)
+deploydocs(
+    repo = "github.com:cgrudz/DataAssimilationBenchmarks.jl.git",
+)
