### 2022-07-19 80ab46a
Code cleanup
