+++ test/TestObsOperators.jl
@@ -3,7 +3,7 @@ module TestObsOperators
 ##############################################################################################
 # imports and exports
 using DataAssimilationBenchmarks.ObsOperators
-using ForwardDiff, ReverseDiff
+using ForwardDiff
 ##############################################################################################
 """
     alternating_obs_jacobian_pos() 
