+++ test/TestSingleIterationSmootherExps.jl
@@ -23,8 +23,8 @@ function analyze_sda_ensemble_smoother_state_L96()
     try
         # test if the filter RMSE for standard simulation falls below adequate threshold
         path = pkgdir(DataAssimilationBenchmarks) * "/src/data/etks-single-iteration/"
-        data = load(path * "etks-single-iteration_L96_state_seed_0000_diff_0.000_sysD_40_" * 
-                    "obsD_40_obsU_1.00_gamma_001.0_nanl_03500_tanl_0.05_h_0.05_lag_010_" * 
+        data = load(path * "etks-single-iteration_L96_state_seed_0000_diff_0.000_sysD_40_" *
+                    "obsD_40_obsU_1.00_gamma_001.0_nanl_03500_tanl_0.05_h_0.05_lag_010_" *
                     "shift_001_mda_false_nens_021_stateInfl_1.02.jld2")
         filt_rmse = data["filt_rmse"]
         post_rmse = data["post_rmse"]
@@ -55,7 +55,7 @@ function analyze_mda_ensemble_smoother_state_L96()
         # test if the filter RMSE for standard simulation falls below adequate threshold
         path = pkgdir(DataAssimilationBenchmarks) * "/src/data/etks-single-iteration/"
         data = load(path * "etks-single-iteration_L96_state_seed_0000_diff_0.000_sysD_40_" *
-                    "obsD_40_obsU_1.00_gamma_001.0_nanl_03500_tanl_0.05_h_0.05_lag_010_" * 
+                    "obsD_40_obsU_1.00_gamma_001.0_nanl_03500_tanl_0.05_h_0.05_lag_010_" *
                     "shift_001_mda_true_nens_021_stateInfl_1.02.jld2")
         filt_rmse = data["filt_rmse"]
         post_rmse = data["post_rmse"]
@@ -91,7 +91,7 @@ function analyze_sda_ensemble_smoother_param_L96()
         path = pkgdir(DataAssimilationBenchmarks) * "/src/data/etks-single-iteration/"
         data = load(path * "etks-single-iteration_L96_param_seed_0000_diff_0.000_sysD_41" *
                     "_obsD_40_obsU_1.00_gamma_001.0_paramE_0.10_paramW_0.0010_nanl_03500" *
-                    "_tanl_0.05_h_0.05_lag_010_shift_001_mda_false_nens_021_" * 
+                    "_tanl_0.05_h_0.05_lag_010_shift_001_mda_false_nens_021_" *
                     "stateInfl_1.02_paramInfl_1.00.jld2")
         filt_rmse = data["filt_rmse"]
         post_rmse = data["post_rmse"]
