+++ src/experiments/FilterExps.jl
@@ -26,19 +26,19 @@ Output from the experiment is saved in a dictionary of the form,
                             "fore_spread" => fore_spread,
                             "filt_spread" => filt_spread,
                             "method" => method,
-                            "seed" => seed, 
+                            "seed" => seed,
                             "diffusion" => diffusion,
                             "dx_params" => dx_params,
                             "sys_dim" => sys_dim,
-                            "obs_dim" => obs_dim, 
+                            "obs_dim" => obs_dim,
                             "obs_un" => obs_un,
                             "gamma" => γ,
                             "nanl" => nanl,
                             "tanl" => tanl,
                             "h" =>  h,
-                            "N_ens" => N_ens, 
+                            "N_ens" => N_ens,
                             "s_infl" => round(s_infl, digits=2)
-                           ) 
+                           )
 
 Experiment output is written to a directory defined by
 
@@ -46,25 +46,25 @@ Experiment output is written to a directory defined by
 
 where the file name is written dynamically according to the selected parameters as follows:
 
-    method * 
+    method *
     "_" * model *
-    "_state_seed_" * lpad(seed, 4, "0") * 
-    "_diff_" * rpad(diffusion, 5, "0") * 
-    "_sysD_" * lpad(sys_dim, 2, "0") * 
-    "_obsD_" * lpad(obs_dim, 2, "0") * 
+    "_state_seed_" * lpad(seed, 4, "0") *
+    "_diff_" * rpad(diffusion, 5, "0") *
+    "_sysD_" * lpad(sys_dim, 2, "0") *
+    "_obsD_" * lpad(obs_dim, 2, "0") *
     "_obsU_" * rpad(obs_un, 4, "0") *
     "_gamma_" * lpad(γ, 5, "0") *
-    "_nanl_" * lpad(nanl, 5, "0") * 
-    "_tanl_" * rpad(tanl, 4, "0") * 
+    "_nanl_" * lpad(nanl, 5, "0") *
+    "_tanl_" * rpad(tanl, 4, "0") *
     "_h_" * rpad(h, 4, "0") *
-    "_nens_" * lpad(N_ens, 3,"0") * 
-    "_stateInfl_" * rpad(round(s_infl, digits=2), 4, "0") * 
+    "_nens_" * lpad(N_ens, 3,"0") *
+    "_stateInfl_" * rpad(round(s_infl, digits=2), 4, "0") *
     ".jld2"
 
 """
 function ensemble_filter_state((time_series, method, seed, nanl, obs_un, obs_dim,
                                  γ, N_ens, s_infl)::NamedTuple{
-                               (:time_series,:method,:seed,:nanl,:obs_un,:obs_dim, 
+                               (:time_series,:method,:seed,:nanl,:obs_un,:obs_dim,
                                 :γ,:N_ens,:s_infl),
                                <:Tuple{String,String,Int64,Int64,Float64,Int64,
                                        Float64,Int64,Float64}})
@@ -81,7 +81,7 @@ function ensemble_filter_state((time_series, method, seed, nanl, obs_un, obs_dim
 
     # define the observation operator HARD-CODED in this line
     H_obs = alternating_obs_operator
-    
+
     # set the integration step size for the ensemble at 0.01 if an SDE, if deterministic
     # simply use the same step size as the observation model
     if diffusion > 0.0
@@ -89,7 +89,7 @@ function ensemble_filter_state((time_series, method, seed, nanl, obs_un, obs_dim
     else
         h = ts["h"]
     end
-    
+
     # define the dynamical model derivative for this experiment from the name
     # supplied in the time series
     if model == "L96"
[…]
