### 2022-07-26 6b1eee2
3D-VAR Experiment Sketch and Updated Docs
