+++ src/experiments/ParallelExperimentDriver.jl
@@ -3,66 +3,87 @@ module ParallelExperimentDriver
 ##############################################################################################
 # imports and exports
 using Distributed
-@everywhere push!(LOAD_PATH, "/DataAssimilationBenchmarks/src/")
-@everywhere push!(LOAD_PATH, "/DataAssimilationBenchmarks/src/methods")
-@everywhere push!(LOAD_PATH, "/DataAssimilationBenchmarks/src/models")
-@everywhere push!(LOAD_PATH, "/DataAssimilationBenchmarks/src/experiments")
-@everywhere using JLD2, ..DataAssimilationBenchmarks, ..FilterExps, ..SmootherExps,
-                  ..EnsembleKalmanSchemes, ..DeSolvers, ..L96, ..ParallelExperimentDriver
-@everywhere export experiment
-#@everywhere export wrap_exp
+@everywhere using DataAssimilationBenchmarks
+@everywhere export wrap_exp
 
 ##############################################################################################
-## Timeseries data
-##############################################################################################
-# observation timeseries to load into the experiment as truth twin, timeseries are named by
-# the model, seed to initialize, the integration scheme used to produce, number of analyses,
-# the spinup length, and the time length between observation points
-
-ts1 = "../data/time_series/L96_time_series_seed_0000_dim_40_diff_0.00_F_08.0_tanl_0.05" *
-      "_nanl_50000_spin_5000_h_0.050.jld2"
-ts2 = "../data/time_series/L96_time_series_seed_0000_dim_40_diff_0.00_F_08.0_tanl_0.10" *
-      "_nanl_50000_spin_5000_h_0.050.jld2"
-ts3 = "../data/time_series/L96_time_series_seed_0000_dim_40_diff_0.00_F_08.0_tanl_0.15" *
-      "_nanl_50000_spin_5000_h_0.050.jld2"
-ts4 = "../data/time_series/L96_time_series_seed_0000_dim_40_diff_0.00_F_08.0_tanl_0.20" *
-      "_nanl_50000_spin_5000_h_0.050.jld2"
-ts5 = "../data/time_series/L96_time_series_seed_0000_dim_40_diff_0.00_F_08.0)tanl_0.25" *
-      "_nanl_50000_spin_5000_h_0.050.jld2"
-
-##############################################################################################
-## Experiment parameter generation
+# Utility methods
 ##############################################################################################
 
+path = pkgdir(DataAssimilationBenchmarks) * "/src/data/time_series/"
+
 ##############################################################################################
-# Filters
+# FIlters
 ##############################################################################################
-## filter_state
-#
-## [time_series, scheme, seed, nanl, obs_un, obs_dim, N_ens, infl] = args
-#
-#schemes = ["enkf-n-primal", "enkf-n-primal-ls", "enkf-n-dual"]
-#seed = 0
-#obs_un = 1.0
-#obs_dim = 40
-#N_ens = 15:43
-#infl = [1.0]#LinRange(1.0, 1.20, 21)
-#nanl = 2500
-#
-## load the experiments
-#args = Tuple[]
-#for scheme in schemes
-#    for N in N_ens
-#        for α in infl
-#            tmp = (time_series, scheme, seed, nanl, obs_un, obs_dim, N, α)
-#            push!(args, tmp)
-#        end
-#    end
-#end
-#
-#experiment = FilterExps.filter_state
-#
-#
+
+function adaptive_inflation_comp()
+    # set time series parameters
+    seed      = 123,
+    h         = 0.05,
+    state_dim = 40,
+    tanl      = 0.05,
+    nanl      = 6500,
+    spin      = 1500,
+    diffusion = 0.00,
+    F         = 8.0,
+
+    # generate truth twin time series
+    L96_time_series(
+                    (
[…]
