+++ src/experiments/SmootherExps.jl
@@ -18,11 +18,10 @@ export classic_ensemble_state, classic_ensemble_param, single_iteration_ensemble
                             lag::Int64, shift::Int64, obs_un::Float64, obs_dim::Int64,
                             γ::Float64, N_ens::Int64, s_infl::Float64)::NamedTuple)
 
-Classic ensemble Kalman smoother state estimation twin experiment.  Twin experiment parameters
-such as the observation dimension, observation uncertainty, data assimilation method, number
-of cycles, ensemble size etc. are specified in the arguments. NOTE: the classic scheme does
-not use multiple data assimilation and we hard code `mda=false` in the function for
-consistency with the API of other methods.
+Classic ensemble Kalman smoother state estimation twin experiment.
+
+NOTE: the classic scheme does not use multiple data assimilation and we hard code
+`mda=false` in the function for consistency with the API of other methods.
 
 Output from the experiment is saved in a dictionary of the form,
 
@@ -57,7 +56,7 @@ Output from the experiment is saved in a dictionary of the form,
 
 Experiment output is written to a directory defined by
 
-    pkgdir(DataAssimilationBenchmarks) * "/src/data/" * method * "-classic/"
+    path = pkgdir(DataAssimilationBenchmarks) * "/src/data/" * method * "-classic/"
 
 where the file name is written dynamically according to the selected parameters as follows:
 
@@ -299,11 +298,71 @@ end
                             γ::Float64, p_err::Float64, p_wlk::Float64, N_ens::Int64,
                             s_infl::Float64, s_infl::Float64})::NamedTuple)
 
-Classic ensemble Kalman smoother joint state-parameter estimation twin experiment.  Twin
-experiment parameters such as the observation dimension, observation uncertainty, data
-assimilation method, number of cycles, ensemble size etc. are specified in the arguments.
+Classic ensemble Kalman smoother joint state-parameter estimation twin experiment.
+
 NOTE: the classic scheme does not use multiple data assimilation and we hard code `mda=false`
 in the function for consistency with other methods.
+
+Output from the experiment is saved in a dictionary of the form,
+
+    data = Dict{String,Any}(
+                            "fore_rmse" => fore_rmse,
+                            "filt_rmse" => filt_rmse,
+                            "post_rmse" => post_rmse,
+                            "param_rmse" => para_rmse,
+                            "fore_spread" => fore_spread,
+                            "filt_spread" => filt_spread,
+                            "post_spread" => post_spread,
+                            "param_spread" => para_spread,
+                            "method" => method,
+                            "seed" => seed,
+                            "diffusion" => diffusion,
+                            "dx_params" => dx_params,
+                            "param_truth" => param_truth,
+                            "sys_dim" => sys_dim,
+                            "state_dim" => state_dim,
+                            "obs_dim" => obs_dim,
+                            "obs_un" => obs_un,
+                            "gamma" => γ,
+                            "p_err" => p_err,
+                            "p_wlk" => p_wlk,
+                            "nanl" => nanl,
+                            "tanl" => tanl,
+                            "lag" => lag,
+                            "shift" => shift,
+                            "mda" => mda,
+                            "h" => h,
+                            "N_ens" => N_ens,
+                            "s_infl" => round(s_infl, digits=2),
+                            "p_infl"  => round(p_infl, digits=2)
+                           )
+
[…]
