### 2022-07-27 ba160a2
Missing gamma argument in the NamedTuple
