+++ src/experiments/ParallelExperimentDriver.jl
@@ -86,7 +86,7 @@ function adaptive_inflation_comp()
                        nanl = nanl,
                        obs_un = obs_un,
                        obs_dim = obs_dim,
-                       γ = γ
+                       γ = γ,
                        N_ens = N_ens,
                        s_infl = s_infl
                       )
