### 2022-07-28 5d68143
D3VARExp Implementation of Cycles and Basic Plot Generation
