+++ src/experiments/ParallelExperimentDriver.jl
@@ -3,10 +3,8 @@ module ParallelExperimentDriver
 ##############################################################################################
 # imports and exports
 using ..DataAssimilationBenchmarks
-using ..DataAssimilationBenchmarks.GenerateTimeSeries
-using ..DataAssimilationBenchmarks.FilterExps
-using ..DataAssimilationBenchmarks.SmootherExps
-export adaptive_inflation_comp
+export ensemble_filter_adaptive_inflation, ensemble_filter_param, classic_ensemble_state,
+       classic_ensemble_param, single_iteration_ensemble_state, iterative_ensemble_state 
 
 ##############################################################################################
 # Utility methods and definitions
@@ -15,15 +13,16 @@ export adaptive_inflation_comp
 path = pkgdir(DataAssimilationBenchmarks) * "/src/data/time_series/"
 
 ##############################################################################################
-# FIlters
+# Filters
 ##############################################################################################
+# compare adaptive inflation methods
 
-function adaptive_inflation_comp()
+function ensemble_filter_adaptive_inflation()
 
     exp = DataAssimilationBenchmarks.FilterExps.ensemble_filter_state
     function wrap_exp(arguments)
         try
-            print(exp(arguments))
+            exp(arguments)
         catch
             print("Error on " * string(arguments) * "\n")
         end
@@ -40,18 +39,18 @@ function adaptive_inflation_comp()
     F         = 8.0
 
     # generate truth twin time series
-    L96_time_series(
-                    (
-                      seed      = seed,
-                      h         = h,
-                      state_dim = state_dim,
-                      tanl      = tanl,
-                      nanl      = nanl,
-                      spin      = spin,
-                      diffusion = diffusion,
-                      F         = F,
-                     )
-                   )
+    GenerateTimeSeries.L96_time_series(
+                                       (
+                                         seed      = seed,
+                                         h         = h,
+                                         state_dim = state_dim,
+                                         tanl      = tanl,
+                                         nanl      = nanl,
+                                         spin      = spin,
+                                         diffusion = diffusion,
+                                         F         = F,
+                                        )
+                                      )
 
     # define load path to time series
     time_series = path * "L96_time_series_seed_" * lpad(seed, 4, "0") *
@@ -69,7 +68,7 @@ function adaptive_inflation_comp()
     seed = 1234
     obs_un = 1.0
     obs_dim = 40
-    N_enss = 15:2:43
+    N_enss = 15:3:42
     s_infls = [1.0]
     nanl = 4000
     γ = 1.0
@@ -99,192 +98,489 @@ end
 
 
 ##############################################################################################
-# filter_param
-## [time_series, scheme, seed, nanl, obs_un, obs_dim, param_err, param_wlk, N_ens,
-##  state_infl, param_infl] = args
-#
-#schemes = ["enkf", "etkf"]
-#seed = 0
-#obs_un = 1.0
-#obs_dim = 40
-#param_err = 0.03
-#param_wlk = [0.0000, 0.0001, 0.0010, 0.0100]
-#N_ens = 14:41
-#state_infl = LinRange(1.0, 1.20, 21)
-#param_infl = LinRange(1.0, 1.00, 1)
-#nanl = 2500
-#
-## load the experiments
-#args = Tuple[]
-#for scheme in schemes
-#    for wlk in param_wlk
-#        for N in N_ens
-#            for s_infl in state_infl
-#                for p_infl in param_infl
-#                    tmp = (time_series, scheme, seed, nanl, obs_un, obs_dim,
-#                           param_err, wlk, N, s_infl, p_infl)
-#                    push!(args, tmp)
-#                end
-#            end
-#        end
-#    end
-#end
-#
[…]
