+++ docs/src/submodules/experiments/ParallelExperimentDriver.md
@@ -2,10 +2,12 @@
 
 ## ParallelExperimentDriver API
 
-The `ParallelExperimentDriver.jl` is a simple parallel implementation, though currently lacks a soft-fail when numerical
-instability is encountered.  This means that if a single experiment configuration in the collection fails due to overflow, the entire
-collection will cancel.  A fix for this is being explored, but the recommendation is to use the slurm submit
-scripts below as templates for generating large parameter grid configurations and running them on servers.
+The `ParallelExperimentDriver.jl` is a simple parallel implementation of calling experiment parameter arrays
+with `pmap` and Julia's native distributed computing.  This defines argumentless functions to construct
+the parameter array and input data necessary to generate a sensitivity test, and implements a soft-fail for
+experiments instability is encountered causing a crash of an experiment. This means that if a single experiment
+configuration in the parameter array fails due to overflow, the remaining configurations will continue their
+own course unaffected.
 
 ## Docstrings
 #```@autodocs
