### 2022-07-28 b4ea006
Uploading basic run script
