+++ docs/make.jl
@@ -38,6 +38,6 @@ makedocs(
             ]
 )
 
-#deploydocs(
-#    repo = "github.com:cgrudz/DataAssimilationBenchmarks.jl.git",
-#)
+deploydocs(
+    repo = "github.com:cgrudz/DataAssimilationBenchmarks.jl.git",
+)
