+++ src/experiments/D3VARExps.jl
@@ -2,20 +2,14 @@
 module D3VARExps
 ##############################################################################################
 # imports and exports
-using Random, Distributions, LinearAlgebra, StatsBase
+using Random, Distributions, LinearAlgebra, StatsBase, Statistics
 using JLD2, HDF5, Plots
 using ..DataAssimilationBenchmarks, ..ObsOperators, ..DeSolvers, ..XdVAR
 ##############################################################################################
 # Main 3DVAR experiments
 ##############################################################################################
 
-#=function D3_var_filter_analysis((time_series, method, seed, nanl, lag, shift, obs_un, obs_dim,
-                        γ, N_ens, s_infl)::NamedTuple{
-                        (:time_series,:method,:seed,:nanl,:lag,:shift,:obs_un,:obs_dim,
-                        :γ,:N_ens,:s_infl),
-                        <:Tuple{String,String,Int64,Int64,Int64,Int64,Float64,Int64,
-                            Float64,Int64,Float64}})=#
-function D3_var_filter_analysis()
+function D3_var_filter_analysis_simple()
     
     # time the experiment
     t1 = time()
@@ -120,7 +114,7 @@ function D3_var_filter_analysis()
     filt_rmse[i] = rmse_filter
 
     # reinitializing x_b and x_t for next cycle
-    x_b = x_opt
+     x_b = x_opt
     end
 
     data = Dict{String,Any}(
@@ -149,10 +143,328 @@ function D3_var_filter_analysis()
     print("Runtime " * string(round((time() - t1)  / 60.0, digits=4))  * " minutes\n")
     # make plot
     t = 1:nanl
-    plot(t, fore_rmse, marker=(:circle,5), label = "Forecast")
+    plot(t, fore_rmse, marker=(:circle,5), label = "Forecast", title="Root-Mean-Square Error vs. Time")
     plot!(t, filt_rmse, marker=(:circle,5), label = "Filter")
     xlabel!("Time [Cycles]")
     ylabel!("Root-Mean-Square Error [RMSE]")
+end
+
+
+##############################################################################################
+
+function D3_var_filter_analysis((time_series, γ, is_informed, tuning_factor, is_updated)::NamedTuple{
+                        (:time_series,:γ,:is_informed,:tuning_factor,:is_updated),<:Tuple{String,
+                            Float64,Bool,Float64,Bool}})
+    
+    # time the experiment
+    t1 = time()
+
+    # load the timeseries and associated parameters
+    ts = load(time_series)::Dict{String,Any}
+    diffusion = ts["diffusion"]::Float64
+    dx_params = ts["dx_params"]::ParamDict(Float64)
+    tanl = ts["tanl"]::Float64
+    nanl = ts["nanl"]::Int64
+    # set the integration step size for the ensemble 
+    h = ts["h"]::Float64
+
+    # define the observation operator HARD-CODED in this line
+    H_obs = alternating_obs_operator
+
+    # define the dynamical model derivative for this experiment
+    # we are assuming Lorenz-96 model
+    dx_dt = L96.dx_dt
+
+    # define integration method
+    step_model! = rk4_step!
+
+    # number of discrete forecast steps
+    f_steps = convert(Int64, tanl / h)
+
+    # set seed
+    seed = ts["seed"]::Int64
+    Random.seed!(seed)
+
+    # define the initialization
+    o = ts["obs"]::Array{Float64, 2}
+    obs_un = 1
+    obs_cov = obs_un^2.0 * I
[…]
