+++ src/DataAssimilationBenchmarks.jl
@@ -135,6 +135,7 @@ include("experiments/FilterExps.jl")
 include("experiments/SmootherExps.jl")
 include("experiments/SingleExperimentDriver.jl")
 include("experiments/D3VARExps.jl")
+include("experiments/VarAnalysisExperimentDriver.jl")
 using .DeSolvers
 using .EnsembleKalmanSchemes
 using .L96
