+++ docs/src/submodules/experiments/VarAnalysisExperimentDriver.md
@@ -1,0 +1,8 @@
+# VarAnalysisExperimentDriver
+
+
+## Methods
+
+```@autodocs
+Modules = [DataAssimilationBenchmarks.VarAnalysisExperimentDriver]
+```
