+++ src/run_var_analysis_test.jl
@@ -3,7 +3,7 @@ module run_var_analysis_test
 ##############################################################################################
 # imports and exports
 using Distributed
-@everywhere using ..DataAssimilationBenchmarks
+@everywhere using DataAssimilationBenchmarks
 
 ##############################################################################################
 
