### 2022-08-16 1cc5e80
3D-VAR Experiments Clean-Up
The code for the 3D-VAR filter experiments, parallel analysis driver, and plotting has been cleaned up as well as file placement in the directory.
