+++ src/experiments/D3VARExps.jl
@@ -2,29 +2,24 @@
 module D3VARExps
 ##############################################################################################
 # imports and exports
-using Random, Distributions, LinearAlgebra, StatsBase, Statistics
+using Random, Distributions, LinearAlgebra, StatsBase, Statistics, Measures
 using JLD2, HDF5, Plots
 using ..DataAssimilationBenchmarks, ..ObsOperators, ..DeSolvers, ..XdVAR
 ##############################################################################################
 # Main 3DVAR experiments
 ##############################################################################################
-
+"""
+    D3_var_filter_analysis_simple()
+"""
 function D3_var_filter_analysis_simple()
-    
     # time the experiment
     t1 = time()
 
     # Define experiment parameters
     # number of cycles in experiment
     nanl = 40
-    # load the timeseries and associated parameters
-    # ts = load(time_series)::Dict{String,Any}
-    # diffusion = ts["diffusion"]::Float64
     diffusion = 0.0
-    # dx_params = ts["dx_params"]::ParamDict(Float64)
-    # tanl = ts["tanl"]::Float64
     tanl = 0.05
-    # model = ts["model"]::String
     γ = [8.0]
 
     # define the observation operator HARD-CODED in this line
@@ -36,8 +31,8 @@ function D3_var_filter_analysis_simple()
     # define derivative parameter
     dx_params = Dict{String, Vector{Float64}}("F" => [8.0])
 
-    # define the dynamical model derivative for this experiment from the name
-    # supplied in the time series - we are assuming Lorenz-96 model
+    # define the dynamical model derivative for this experiment - we are assuming 
+    # Lorenz-96 model
     dx_dt = L96.dx_dt
 
     # define integration method
@@ -53,7 +48,6 @@ function D3_var_filter_analysis_simple()
     # define the initialization
     # observation noise
     v = rand(Normal(0, 1), 40)
-
     # define the initial observation range and truth reference solution
     x_b = zeros(40)
     x_t = x_b + v
@@ -84,37 +78,26 @@ function D3_var_filter_analysis_simple()
             step_model!(x_t, 0.0, kwargs)
         end
 
-    # multivariate - rand(MvNormal(zeros(40), I))
-    w = rand(Normal(0, 1), 40)
-    obs = x_t + w
+        # multivariate - rand(MvNormal(zeros(40), I))
+        w = rand(Normal(0, 1), 40)
+        obs = x_t + w
 
-    state_cov = I
-    obs_cov = I
+        state_cov = I
+        obs_cov = I
 
-    # generate initial forecast cost
-    # J_i = XdVAR.D3_var_cost(x_b, obs, x_b, state_cov, H_obs, obs_cov, kwargs)
-    # print("Cost Function Output: \n")
-    # display(J)
-    # optimized cost function input and value
-    x_opt = XdVAR.D3_var_NewtonOp(x_b, obs, x_b, state_cov, H_obs, obs_cov, kwargs)
-    # J_opt = XdVAR.D3_var_cost(x_opt, obs, x_b, state_cov, H_obs, obs_cov, kwargs)
-    # print("Optimized Cost Function Output: \n")
-    # display(J_opt)
+        # optimized cost function input and value
+        x_opt = XdVAR.D3_var_NewtonOp(x_b, obs, x_b, state_cov, H_obs, obs_cov, kwargs)
 
-    # compare model forecast and truth twin via RMSE
-    rmse_forecast = sqrt(msd(x_b, x_t))
[…]
