### 2022-08-16 f4b36cd
Resolving conflicts
Resolved make.jl and slurm conflicts.
