+++ src/slurm_submit_scripts/SlurmParallelSubmit.jl
@@ -1,469 +1,0 @@
-##############################################################################################
-module SlurmParallelSubmit 
-##############################################################################################
-# imports and exports
-using FilterExps, SmootherExps, EnsembleKalmanSchemes, DeSolvers, L96, JLD2, Debugger
-
-##############################################################################################
-##############################################################################################
-## Time series data 
-##############################################################################################
-# observation time series to load into the experiment as truth twin
-# time series are named by the model, seed to initialize, the integration scheme
-# used to produce, number of analyses, the spinup length, and the time
-# length between observation points
-
-path = joinpath(@__DIR__, "../data/")
-
-ts01 = path * "time_series/L96_time_series_seed_0000_dim_40_diff_0.00_" *
-       "F_08.0_tanl_0.05_nanl_50000_spin_5000_h_0.010.jld2"
-ts02 = path * "time_series/L96_time_series_seed_0000_dim_40_diff_0.00_" *
-       "F_08.0_tanl_0.10_nanl_50000_spin_5000_h_0.010.jld2"
-ts03 = path * "time_series/L96_time_series_seed_0000_dim_40_diff_0.00_" *
-       "F_08.0_tanl_0.15_nanl_50000_spin_5000_h_0.010.jld2"
-ts04 = path * "time_series/L96_time_series_seed_0000_dim_40_diff_0.00_" *
-       "F_08.0_tanl_0.20_nanl_50000_spin_5000_h_0.010.jld2"
-ts05 = path * "time_series/L96_time_series_seed_0000_dim_40_diff_0.00_" *
-       "F_08.0_tanl_0.25_nanl_50000_spin_5000_h_0.010.jld2"
-ts06 = path * "time_series/L96_time_series_seed_0000_dim_40_diff_0.00_" *
-       "F_08.0_tanl_0.30_nanl_50000_spin_5000_h_0.010.jld2"
-ts07 = path * "time_series/L96_time_series_seed_0000_dim_40_diff_0.00_" *
-       "F_08.0_tanl_0.35_nanl_50000_spin_5000_h_0.010.jld2"
-ts08 = path * "time_series/L96_time_series_seed_0000_dim_40_diff_0.00_" *
-       "F_08.0_tanl_0.40_nanl_50000_spin_5000_h_0.010.jld2"
-ts09 = path * "time_series/L96_time_series_seed_0000_dim_40_diff_0.00_" *
-       "F_08.0_tanl_0.45_nanl_50000_spin_5000_h_0.010.jld2"
-ts10 = path * "time_series/L96_time_series_seed_0000_dim_40_diff_0.00_" *
-       "F_08.0_tanl_0.50_nanl_50000_spin_5000_h_0.010.jld2"
-ts11 = path * "time_series/L96_time_series_seed_0000_dim_40_diff_0.00_" *
-       "F_08.0_tanl_0.55_nanl_50000_spin_5000_h_0.010.jld2"
-ts12 = path * "time_series/L96_time_series_seed_0000_dim_40_diff_0.00_" *
-       "F_08.0_tanl_0.60_nanl_50000_spin_5000_h_0.010.jld2"
-
-
-##############################################################################################
-## Experiment parameter generation 
-##############################################################################################
-##############################################################################################
-# Filters
-##############################################################################################
-# arguments are 
-# time_series, method, seed, nanl, obs_un, obs_dim, γ, N_ens, infl = args
-#
-#methods = ["mlef-transform", "mlef-ls-transform"]
-#seed = 0
-#obs_un = 1.0
-#obs_dim = 40
-#gammas = Array{Float64}(1:10)
-#N_ens = 15:2:43 
-#infl = LinRange(1.0, 1.10, 11)
[…]
