+++ src/DataAssimilationBenchmarks.jl
@@ -135,6 +135,8 @@ include("experiments/FilterExps.jl")
 include("experiments/SmootherExps.jl")
 include("experiments/SingleExperimentDriver.jl")
 include("experiments/ParallelExperimentDriver.jl")
+include("experiments/D3VARExps.jl")
+include("experiments/VarAnalysisExperimentDriver.jl")
 using .DeSolvers
 using .EnsembleKalmanSchemes
 using .L96
@@ -144,8 +146,11 @@ using .GenerateTimeSeries
 using .FilterExps
 using .SmootherExps
 using .SingleExperimentDriver
+using .ParallelExperimentDriver
+using .D3VARExps
 export DeSolvers, EnsembleKalmanSchemes, XdVAR, L96, IEEE39bus, ObsOperators,
-    GenerateTimeSeries, FilterExps, SingleExperimentDriver, ParallelExperimentDriver
+       GenerateTimeSeries, FilterExps, SingleExperimentDriver, ParallelExperimentDriver,
+       D3VARExps
 
 ##############################################################################################
 # info
