+++ src/DataAssimilationBenchmarks.jl
@@ -149,7 +149,8 @@ using .SingleExperimentDriver
 using .ParallelExperimentDriver
 using .D3VARExps
 export DeSolvers, EnsembleKalmanSchemes, XdVAR, L96, IEEE39bus, ObsOperators,
-    GenerateTimeSeries, FilterExps, SingleExperimentDriver, D3VARExps
+    GenerateTimeSeries, FilterExps, SingleExperimentDriver, ParallelExperimentDriver, 
+    D3VARExps
 
 ##############################################################################################
 # info
