+++ src/DataAssimilationBenchmarks.jl
@@ -149,9 +149,8 @@ using .SingleExperimentDriver
 using .ParallelExperimentDriver
 using .D3VARExps
 export DeSolvers, EnsembleKalmanSchemes, XdVAR, L96, IEEE39bus, ObsOperators,
-    GenerateTimeSeries, FilterExps, SingleExperimentDriver, ParallelExperimentDriver, 
-    D3VARExps
-
+       GenerateTimeSeries, FilterExps, SingleExperimentDriver, ParallelExperimentDriver,
+       D3VARExps
 ##############################################################################################
 # info
 
