+++ src/DataAssimilationBenchmarks.jl
@@ -151,7 +151,6 @@ using .D3VARExps
 export DeSolvers, EnsembleKalmanSchemes, XdVAR, L96, IEEE39bus, ObsOperators,
        GenerateTimeSeries, FilterExps, SingleExperimentDriver, ParallelExperimentDriver,
        D3VARExps
-
 ##############################################################################################
 # info
 
