### 2022-09-01 669f025
Removing slurm references, too specific to be generally useful
