### 2022-09-02 60f0b76
Merging Var methods more formally for consistency and scalability
Merged Var filter experiments into the existing module / framework
with FIlterExps.jl, SingleExperimentDriver.jl with a basic
reproducible configuration, ParallelExperimentDriver.jl for a
parameter map constructor (to be tested) and in test cases with
propagated argument changes for method consistency.
