+++ src/experiments/SmootherExps.jl
@@ -39,7 +39,7 @@ Output from the experiment is saved in a dictionary of the form,
                             "sys_dim" => sys_dim,
                             "obs_dim" => obs_dim,
                             "obs_un" => obs_un,
-                            "gamma" => γ,
+                            "γ" => γ,
                             "nanl" => nanl,
                             "tanl" => tanl,
                             "lag" => lag,
@@ -145,7 +145,7 @@ function classic_ensemble_state((time_series, method, seed, nanl, lag, shift, ob
                               "h" => h,
                               "diffusion" => diffusion,
                               "s_infl" => s_infl,
-                              "gamma" => γ,
+                              "γ" => γ,
                               "shift" => shift,
                               "mda" => mda
                              )
@@ -252,7 +252,7 @@ function classic_ensemble_state((time_series, method, seed, nanl, lag, shift, ob
                             "sys_dim" => sys_dim,
                             "obs_dim" => obs_dim,
                             "obs_un" => obs_un,
-                            "gamma" => γ,
+                            "γ" => γ,
                             "nanl" => nanl,
                             "tanl" => tanl,
                             "lag" => lag,
@@ -323,7 +323,7 @@ Output from the experiment is saved in a dictionary of the form,
                             "state_dim" => state_dim,
                             "obs_dim" => obs_dim,
                             "obs_un" => obs_un,
-                            "gamma" => γ,
+                            "γ" => γ,
                             "p_err" => p_err,
                             "p_wlk" => p_wlk,
                             "nanl" => nanl,
@@ -451,7 +451,7 @@ function classic_ensemble_param((time_series, method, seed, nanl, lag, shift, ob
                               "h" => h,
                               "diffusion" => diffusion,
                               "dx_params" => dx_params,
-                              "gamma" => γ,
+                              "γ" => γ,
                               "state_dim" => state_dim,
                               "p_wlk" => p_wlk,
                               "s_infl" => s_infl,
@@ -588,7 +588,7 @@ function classic_ensemble_param((time_series, method, seed, nanl, lag, shift, ob
                             "state_dim" => state_dim,
                             "obs_dim" => obs_dim,
                             "obs_un" => obs_un,
-                            "gamma" => γ,
+                            "γ" => γ,
                             "p_err" => p_err,
                             "p_wlk" => p_wlk,
                             "nanl" => nanl,
@@ -658,7 +658,7 @@ Output from the experiment is saved in a dictionary of the form,
                             "sys_dim" => sys_dim,
                             "obs_dim" => obs_dim,
                             "obs_un" => obs_un,
-                            "gamma" => γ,
+                            "γ" => γ,
                             "nanl" => nanl,
                             "tanl" => tanl,
                             "lag" => lag,
@@ -763,7 +763,7 @@ function single_iteration_ensemble_state((time_series, method, seed, nanl, lag, 
                               "dx_params" => dx_params,
                               "h" => h,
                               "diffusion" => diffusion,
-                              "gamma" => γ,
+                              "γ" => γ,
                               "shift" => shift,
                               "s_infl" => s_infl,
                               "mda" => mda
@@ -951,7 +951,7 @@ function single_iteration_ensemble_state((time_series, method, seed, nanl, lag, 
                             "sys_dim" => sys_dim,
                             "obs_dim" => obs_dim,
                             "obs_un" => obs_un,
-                            "gamma" => γ,
+                            "γ" => γ,
                             "nanl" => nanl,
                             "tanl" => tanl,
                             "lag" => lag,
@@ -1019,7 +1019,7 @@ Output from the experiment is saved in a dictionary of the form,
[…]
