+++ src/experiments/D3VARExps.jl
@@ -8,308 +8,6 @@ using ..DataAssimilationBenchmarks, ..ObsOperators, ..DeSolvers, ..XdVAR
 ##############################################################################################
 # Main 3DVAR experiments
 ##############################################################################################
-"""
-    D3_var_filter_analysis_simple()
-"""
-function D3_var_filter_analysis_simple()
-    # time the experiment
-    t1 = time()
-
-    # Define experiment parameters
-    # number of cycles in experiment
-    nanl = 40
-    diffusion = 0.0
-    tanl = 0.05
-    γ = [8.0]
-
-    # define the observation operator HARD-CODED in this line
-    H_obs = alternating_obs_operator
-
-    # set the integration step size for the ensemble at 0.01 - we are assuming SDE
-    h = 0.01
-
-    # define derivative parameter
-    dx_params = Dict{String, Vector{Float64}}("F" => [8.0])
-
-    # define the dynamical model derivative for this experiment - we are assuming 
-    # Lorenz-96 model
-    dx_dt = L96.dx_dt
-
-    # define integration method
-    step_model! = rk4_step!
-
-    # number of discrete forecast steps
-    f_steps = convert(Int64, tanl / h)
-
-    # set seed
-    seed = 234
-    Random.seed!(seed)
-
-    # define the initialization
-    # observation noise
-    v = rand(Normal(0, 1), 40)
-    # define the initial observation range and truth reference solution
-    x_b = zeros(40)
-    x_t = x_b + v
-    
-    # define kwargs for the analysis method
-    # and the underlying dynamical model
-    kwargs = Dict{String,Any}(
-                              "dx_dt" => dx_dt,
-                              "f_steps" => f_steps,
-                              "step_model" => step_model!,
-                              "dx_params" => dx_params,
-                              "h" => h,
-                              "diffusion" => diffusion,
-                              "gamma" => γ,
-                             )
-
-    # create storage for the forecast and analysis statistics
-    fore_rmse = Vector{Float64}(undef, nanl)
-    filt_rmse = Vector{Float64}(undef, nanl)
-    
-    for i in 1:nanl
-        #print("Iteration: ")
-        #display(i)
-        for j in 1:f_steps
-            # M(x^b)
-            step_model!(x_b, 0.0, kwargs)
-            # M(x^t)
-            step_model!(x_t, 0.0, kwargs)
-        end
-
-        # multivariate - rand(MvNormal(zeros(40), I))
-        w = rand(Normal(0, 1), 40)
-        obs = x_t + w
-
-        state_cov = I
-        obs_cov = I
-
-        # optimized cost function input and value
-        x_opt = XdVAR.D3_var_NewtonOp(x_b, obs, x_b, state_cov, H_obs, obs_cov, kwargs)
-
-        # compare model forecast and truth twin via RMSE
-        rmse_forecast = sqrt(msd(x_b, x_t))
-        fore_rmse[i] = rmse_forecast
-
-        # compare optimal forecast and truth twin via RMSE
-        rmse_filter = sqrt(msd(x_opt, x_t))
-        filt_rmse[i] = rmse_filter
-
-        # reinitializing x_b and x_t for next cycle
-        x_b = x_opt
-    end
-
-    data = Dict{String,Any}(
-                            "seed" => seed,
-                            "diffusion" => diffusion,
-                            "dx_params" => dx_params,
-                            "gamma" => γ,
-                            "tanl" => tanl,
-                            "nanl" => nanl,
[…]
