### 2022-09-02 65ed22b
Removing debugger, bugfix
