### 2022-09-06 7d71852
Merging in parallel experiment driver for 3D-VAR
Made successful run of parallel experiment for the tuned inflation
with 3D-VAR using either the ID or climatology as the background,
with hook to generate the climatology or to read from file in the
experiment.  Needs scripts for generating plots and analysis, but
VAR codes are now effectively merged and smoothed with the conventions
of the main code base.
