### 2022-09-06 a3a1234
Bugfix with dimension of output as 1x1 versus scalar
