### 2022-09-06 b3a4bd7
Removing debugger bug
