+++ src/experiments/D3VARExps.jl
@@ -1,13 +1,0 @@
-##############################################################################################
-module D3VARExps
-##############################################################################################
-# imports and exports
-using Random, Distributions, LinearAlgebra, StatsBase, Statistics, Measures
-using JLD2, HDF5, Plots
-using ..DataAssimilationBenchmarks, ..ObsOperators, ..DeSolvers, ..XdVAR
-##############################################################################################
-# Main 3DVAR experiments
-##############################################################################################
-# end module
-
-end
