### 2022-09-07 f700b34
Docs update, mild refactoring
