+++ src/experiments/run_sensitivity_test.jl
@@ -4,7 +4,6 @@ module run_sensitivity_test
 # imports and exports
 using Distributed
 @everywhere using DataAssimilationBenchmarks
-
 ##############################################################################################
 
 config = ParallelExperimentDriver.ensemble_filter_adaptive_inflation
@@ -12,7 +11,7 @@ config = ParallelExperimentDriver.ensemble_filter_adaptive_inflation
 print("Generating experiment configurations from " * string(config) * "\n")
 print("Generate truth twin\n")
 
-args, exp = config()
+args, wrap_exp = config()
 num_exps = length(args)
 
 print("Configuration ready\n")
@@ -20,7 +19,7 @@ print("\n")
 print("Running " * string(num_exps) * " configurations on " * string(nworkers()) *
       " total workers\n")
 print("Begin pmap\n")
-pmap(exp, args)
+pmap(wrap_exp, args)
 print("Experiments completed, verify outputs in the appropriate directory under:\n")
 print(pkgdir(DataAssimilationBenchmarks) * "/src/data\n")
 
