+++ src/experiments/SingleExperimentDriver.jl
@@ -4,293 +4,361 @@ module SingleExperimentDriver
 # imports and exports
 using JLD2, HDF5
 using ..DataAssimilationBenchmarks, ..FilterExps, ..SmootherExps, ..GenerateTimeSeries
-export exps
+export time_series_exps, enkf_exps, d3_var_exps, classic_enks_exps, sienks_exps, ienks_exps
 ##############################################################################################
+# Utility methods and definitions
+##############################################################################################
+
 path = pkgdir(DataAssimilationBenchmarks) * "/src/data/time_series/"
 
+##############################################################################################
+# Time series
+##############################################################################################
 """
-    exps["Experiment_name"]["Parameter_settings"]
+    time_series_exps = Dict{String, Any}
 
-This dictionary contains standard inputs for experiments, written as named tuples and stored
-hierarchically by experiment type.  These standard inputs are used in the package
-for for debugging, testing, benchmarking and profiling. Parallel submission scripts are used
-for performance on servers.
+Pre-configured time series experiments for generating test data for twin experiments.
 """
-exps = Dict{String, Any}(
-        # Generate time series experiment configurations
-        "Generate_time_series" => Dict{String, Any}(
-          # Generates a short time series of the L96 model for testing
-          "L96_deterministic_test" => (
-            seed      = 0,
-            h         = 0.05,
-            state_dim = 40,
-            tanl      = 0.05,
-            nanl      = 5000,
-            spin      = 1500,
-            diffusion = 0.00,
-            F         = 8.0,
-           ),
-          # Generates a short time series of the IEEE39bus model for testing
-          "IEEE39bus_deterministic_test" => (
-            seed      = 0,
-            h         = 0.01,
-            tanl      = 0.01,
-            nanl      = 5000,
-            spin      = 1500,
-            diffusion = 0.0,
-           ),
-          ),
-        # Filter twin experiment configurations
-        "Filter" => Dict{String, Any}(
-          # Lorenz-96 ETKF state estimation standard configuration
-          "L96_ETKF_state_test" => (
-            time_series = path *
-            "L96_time_series_seed_0000_dim_40_diff_0.000_F_08.0_tanl_0.05_nanl_05000_" *
-            "spin_1500_h_0.050.jld2",
-            method      = "etkf",
-            seed        = 0,
-            nanl        = 3500,
-            obs_un      = 1.0,
-            obs_dim     = 40,
-            γ           = 1.00,
-            N_ens       = 21,
-            s_infl      = 1.02,
-           ),
-          # Lorenz-96 3D-VAR state estimation standard configuration
-          "L96_D3_var_state_test" => (
-            time_series = path *
-            "L96_time_series_seed_0000_dim_40_diff_0.000_F_08.0_tanl_0.05_nanl_05000_" *
-            "spin_1500_h_0.050.jld2",
-            bkg_cov     = "ID",
-            seed        = 0,
-            nanl        = 3500,
-            obs_un      = 1.0,
-            obs_dim     = 40,
-            γ           = 1.00,
-            s_infl      = 0.20,
-           ),
[…]
