+++ docs/src/submodules/experiments/SmootherExps.md
@@ -1,6 +1,5 @@
 # SmootherExps 
 
-## SmootherExps API
 The `SmootherExps.jl` sub-module contains methods to configure filter twin experiments,
 using a stored time series as generated by [GenerateTimeSeries](@ref) as the underlying
 observation generating process.  The frequency of observations in continuous time is defined

[…]
