### 2022-09-07 cgrudz/DataAssimilationBenchmarks.jl#12 @cgrudz
This issue is resolved, there were options for sparse and nonlinear observations with different types of nonlinearity already built into the system, but this was not sufficiently documented before.  It should now be clear how to use these options based on the documentation on the new ObsOperators module:
https://cgrudz.github.io/DataAssimilationBenchmarks.jl/dev/submodules/models/ObsOperators/
