### 2022-09-07 cgrudz/DataAssimilationBenchmarks.jl#11 @cgrudz
At the moment, this is going to be on the back burner as the various ways one might configure an experiment makes general post-processing a difficult task.  This will be under consideration for future versions, but only after some more general tuning and integration of analysis and visualization codes in native Julia are implemented.
