### 2022-09-07 cgrudz/DataAssimilationBenchmarks.jl#10 @cgrudz
Examples of function outputs are now included in the doc strings for the various methods of the included submodules, as in e.g., the GenerateTimeSeries submodule:
https://cgrudz.github.io/DataAssimilationBenchmarks.jl/dev/submodules/experiments/GenerateTimeSeries/
