### 2022-09-07 cgrudz/DataAssimilationBenchmarks.jl#9 @cgrudz
This was a very important point, and many thanks for bringing this up.  The code base has now been restructured in favor of named tuples.  In particular, this solves the issue of nameless arguments, but also allows one to use the parameter map constructors as defined in ParallelExperimentDriver:
https://cgrudz.github.io/DataAssimilationBenchmarks.jl/dev/submodules/experiments/ParallelExperimentDriver/
This allows one to abstract hyper-parameter grids into named tuples with a simple loop constructor and apply the native Distributed parallel mapping to run a large grid of experiment configurations in naive parallelism.
