### 2022-09-07 cgrudz/DataAssimilationBenchmarks.jl#8 @cgrudz
Many thanks, this has been completed adopted in the new documentation pages:
https://cgrudz.github.io/DataAssimilationBenchmarks.jl/dev/
