### 2022-09-07 e17cd0d
Update CODE_OF_CONDUCT.md
Updating contact
