+++ .github/ISSUE_TEMPLATE/feature-requests-and-new-use-cases.md
@@ -1,0 +1,23 @@
+---
+name: Feature requests and new use-cases
+about: Suggest an idea for this project
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+**Please introduce yourself and your use-case for the software**
+Are you developing a new method that you want to test versus other similar learning / estimation schemes in standard benchmarks or do you want to examine the behavior of an existing method within the main code base in a novel context?  Do you want to implement a new state or observation model to benchmark methods in the code base?  Please note that there is no support for operational data assimilation or data assimilation with real data, please see alternatives that exist for that scope.
+
+**Is your feature request related to a problem? Please describe.**
+A clear and concise description of what the problem is. Ex. I'm always frustrated when [...]
+
+**Describe the solution you'd like**
+A clear and concise description of what you want to happen.  Provide relevant references for models, methods, benchmarks, etc.
+
+**Describe a plan forward for your use-case and any alternatives you've considered**
+A clear and concise description of how you see this new functionality being implemented and any alternative solutions or features you've considered.
+
+**Additional context**
+Add any other context or screenshots about the feature request here.
