### 2022-09-07 d58e301
Updating docs, making community guidelines
