### 2022-09-07 cgrudz/DataAssimilationBenchmarks.jl#15 @cgrudz
Rough guidelines are now included in the documentation:

https://cgrudz.github.io/DataAssimilationBenchmarks.jl/dev/community/Contributing/#Contributing

This is currently being developed piece-wise, due to the small scale of the project at the moment, but with future planning for a more unified refactoring of the codes.
