### 2022-09-07 cgrudz/DataAssimilationBenchmarks.jl#14 @cgrudz
Community guidelines are now included in the main documentation:

https://cgrudz.github.io/DataAssimilationBenchmarks.jl/dev/community/Contributing/

As well as the code of conduct:

https://github.com/cgrudz/DataAssimilationBenchmarks.jl/blob/master/CODE_OF_CONDUCT.md
