### 2022-09-07 b033e09
Updating docs and outputs for consistency
