+++ docs/src/submodules/experiments/SmootherExps.md
@@ -8,7 +8,7 @@ when reading in the data.
 
 ```@raw html
 <div sytle="float:left; width:100%;">
-<img style="width:95%" src="../cyclingSDA.png" alt="Observation analysis forecast cycle over multiple data assimilation windows">
+<img style="width:95%" src="https://raw.githubusercontent.com/cgrudz/DataAssimilationBenchmarks.jl/master/assets/cyclingSDA.png" alt="Observation analysis forecast cycle over multiple data assimilation windows">
 </div>
 ```
 Smoother experiment configurations are generated by supplying a 
