### 2022-09-08 240a473
Updating docs and paper
