+++ paper.bib
@@ -84,27 +84,33 @@
   publisher={Courier Corporation}
 }
 
-@article{grudzien2020numerical,
-  title={On the numerical integration of the Lorenz-96 model, with scalar additive noise, for benchmark twin experiments},
-  author={Grudzien, C. and Bocquet, M. and Carrassi, A.},
-  journal={Geoscientific Model Development},
-  volume={13},
-  number={4},
-  pages={1903--1924},
-  year={2020},
-  publisher={Copernicus GmbH}
+@inproceedings{lam2015numba,
+  title={Numba: A llvm-based python jit compiler},
+  author={Lam, S.K. and Pitrou, A. and Seibert, S.},
+  booktitle={Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC},
+  pages={1--6},
+  year={2015}
 }
 
-@misc{Numba,
-  title = {Numba Documentation},
-  year = {2021},
-  url = {https://numba.readthedocs.io/en/stable/},
-  howpublished = {Accessed: 2021-11-29}
+@article{kalnay20074denkf,
+  title={4-D-Var or ensemble Kalman filter?},
+  author={Kalnay, E and Li, H and Miyoshi, T and Yang, S.C. and Ballabrera-Poy, J.},
+  journal={Tellus A: Dynamic Meteorology and Oceanography},
+  volume={59},
+  number={5},
+  pages={758--773},
+  year={2007},
+  publisher={Taylor \& Francis}
 }
 
-@misc{julia, 
-    title = {Julia Benchmarks},
-    year = {2021},
-    url = {https://julialang.org/benchmarks/ },
-    howpublished = {Accessed: 2021-11-29}
+@article{bezanson2017julia,
+  title={Julia: A fresh approach to numerical computing},
+  author={Bezanson, J. and Edelman, A. and Karpinski, S. and Shah, V.B.},
+  journal={SIAM review},
+  volume={59},
+  number={1},
+  pages={65--98},
+  year={2017},
+  publisher={SIAM}
+  doi = {10.1137/141000671},
 }
