+++ paper.md
@@ -4,20 +4,24 @@ tags:
   - Julia
   - Data Assimilation
   - Optimization
-  - ensemble Kalman filter
-  - ensemble Kalman smoother
-  - ensemble-variational
+  - Machine learning
 authors:
   - name: Colin Grudzien
     orcid: 0000-0002-3084-3178
-    affiliation: 1 
+    affiliation: 1,2
+  - name: Charlotte Merchant
+    affiliation: 1,3 
   - name: Sukhreen Sandhu 
-    affiliation: 2
+    affiliation: 4
 affiliations:
- - name: Department of Mathematics and Statistics, University of Nevada, Reno
+ - name: CW3E - Scripps Instition of Oceanography, University of California, San Diego
    index: 1
- - name: Department of Computer Science and Engineering, University of Nevada, Reno
+ - name: Department of Mathematics and Statistics, University of Nevada, Reno
    index: 2
+ - name: Department of Computer Science, Princeton University 
+   index: 3
+ - name: Department of Computer Science and Engineering, University of Nevada, Reno
+   index: 4
 date: 30 November 2021
 bibliography: paper.bib
 ---
@@ -41,37 +45,33 @@ development and empirical validation of novel data assimilation techniques.
 While analytical proofs can be derived for classical methods such as the Kalman filter
 in linear-Gaussian dynamics [@jazwinski2007stochastic], most currently developed DA
 techniques are designed for estimation in nonlinear, non-Gaussian models where no
-analytical solution may exist.  Similar to nonlinear optimization, DA methods,
+analytical solution may exist.  DA methods,
 therefore, must be studied with rigorous numerical simulation in standard test-cases
 to demonstrate the effectiveness and computational performance of novel algorithms.
 Pursuant to proposing a novel DA method, one should likewise compare the performance
 of a proposed scheme with other standard methods within the same class of estimators.
 
-This package implements several standard data assimilation algorithms, including
-widely used performance modifications that are used in practice to tune these estimators.
-This software framework was written specifically to support the development and intercomparison
-of the novel single-iteration ensemble Kalman smoother (SIEnKS) [@grudzien2021fast].
-Details of the primary DA schemes, including pseudo-code for the methods detailing their implementation,
-and DA experiment benchmark configurations, with root mean square error and ensemble spread diagonostics
-for estimator validation, can be found in the above principal reference.  Additional details on
-numerical integration schemes used in this work for simulating the Lorenz-96 model are found
-in the secondary reference [@grudzien2020numerical].
+This package implements a variety of standard data assimilation algorithms,
+including some of the widely used performance modifications that are used in
+practice to tune these estimators. This software framework was written originally
+to support the development and intercomparison of methods studied in [@grudzien2021fast].
+Details of the ensemble DA schemes, including pseudo-code for the methods detailing
+their implementation, and DA experiment benchmark configurations, can be found in
+the above principal reference.  Additional details on numerical integration schemes
+used in this work for simulating the Lorenz-96 model are found in the secondary
+reference [@grudzien2020numerical].
 
 # Statement of need
 
 Standard libraries exist for full-scale DA system research and development, e.g.,
 the Data Assimilation Research Testbed (DART)[@anderson2009data], but
 there are fewer standard options for theoretical research and algorithm development in
-simple test systems. DataAssimilationBenchmarks.jl provides one framework for studying
-ensemble-based filters and sequential smoothers that are commonly used in online,
-geoscientific prediction settings.  Validated methods, and methods in development, focus
-on evaluating the performance and the structural stability of techniques over wide ranges
-of hyper-parameters that are commonly used to tune estimators in practice.  Specifically,
-this is designed to run naively parallel experiment configurations over independent
-parameters such as ensemble size, static covariance inflation, observation
-operator / network designs that affect the estimator stability and performance.
[…]
