+++ paper.bib
@@ -111,6 +111,6 @@
   number={1},
   pages={65--98},
   year={2017},
-  publisher={SIAM}
+  publisher={SIAM},
   doi = {10.1137/141000671},
 }
