+++ paper.bib
@@ -7,6 +7,7 @@
   pages={1283--1296},
   year={2009},
   publisher={American Meteorological Society},
+	doi={https://doi.org/10.1175/2009BAMS2618.1},
 }
 
 @article{grudzien2021fast,
@@ -16,6 +17,7 @@
   pages={1--62},
   year={2021},
   publisher={Copernicus GmbH},
+	doi={https://doi.org/10.5194/gmd-2021-306},
 }
 
 @article{vetra2018state,
@@ -27,25 +29,28 @@
   pages={1--43},
   year={2018},
   publisher={Taylor \& Francis},
+	doi={http://doi.org/10.1080/16000870.2018.1445364},
 }
 
-@software{patrick_n_raanes_2018_2029296,
-  author       = {Patrick N. Raanes and others},
-  title        = {nansencenter/DAPPER: Version 0.8},
-  month        = {Dec},
-  year         = {2018},
+@software{dapper,
+  author       = {Raanes, P.N. and others},
+  title        = {nansencenter/DAPPER: Version 1.2.1,},
+  month        = {September},
+  year         = {2022},
   publisher    = {Zenodo},
-  version      = {v0.8},
   doi          = {10.5281/zenodo.2029296},
-  url          = {https://doi.org/10.5281/zenodo.2029296},
+  url          = {https://github.com/nansencenter/DAPPER},
 }
 
-@online{enkprocesses,
-    author = {Constantinou, N.C., and others},
-    title = {EnsembleKalmanProcesses.jl},
-    year = {2021},
-    url = {https://github.com/CliMA/EnsembleKalmanProcesses.jl},
-    urldate = {2021-10-26},
+@software{enkprocesses,
+  author       = {Dunbar, O. and others},
+  title        = {CliMA/EnsembleKalmanProcesses.jl: v0.10.0},
+  month        = {aug},
+  year         = {2022},
+  publisher    = {Zenodo},
+  version      = {v0.10.0},
+  doi          = {10.5281/zenodo.7036069},
+  url          = {https://doi.org/10.5281/zenodo.7036069},
 }
 
 @article{grudzien2020numerical,
@@ -57,13 +62,16 @@
   pages={1903--1924},
   year={2020},
   publisher={Copernicus GmbH},
+  URL = {https://gmd.copernicus.org/articles/13/1903/2020/},
+  DOI = {10.5194/gmd-13-1903-2020},
 }
 
 @book{asch2016data,
   title={Data assimilation: methods, algorithms, and applications},
   author={Asch, M., and Bocquet, M., and Nodet, M.},
   year={2016},
-  publisher={SIAM}
+  publisher={SIAM},
+	doi={https://doi.org/10.1137/1.9781611974546},
 }
 
 @article{carrassi2018data,
@@ -74,14 +82,17 @@
   number={5},
   pages={e535},
   year={2018},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
 }
 
 @book{jazwinski2007stochastic,
-  title={Stochastic processes and filtering theory},
+  title={Stochastic Processes and Filtering Theory},
   author={Jazwinski, A.H.},
+  isbn={9780486785349},
+  series={Dover Books on Electrical Engineering Series},
+  url={https://books.google.com/books?id=cZzKswEACAAJ},
   year={2007},
-  publisher={Courier Corporation}
+  publisher={Dover Publications, Incorporated},
 }
 
 @inproceedings{lam2015numba,
@@ -90,6 +101,7 @@
   booktitle={Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC},
   pages={1--6},
   year={2015}
[…]
