+++ paper.md
@@ -3,6 +3,7 @@ title: 'DataAssimilationBenchmarks.jl: a data assimilation research framework.'
 tags:
   - Julia
   - Data Assimilation
+	- Bayesian Inference
   - Optimization
   - Machine learning
 authors:
@@ -14,7 +15,7 @@ authors:
   - name: Sukhreen Sandhu 
     affiliation: 4
 affiliations:
- - name: CW3E - Scripps Instition of Oceanography, University of California, San Diego
+ - name: CW3E - Scripps Institution of Oceanography, University of California, San Diego
    index: 1
  - name: Department of Mathematics and Statistics, University of Nevada, Reno
    index: 2
@@ -22,7 +23,7 @@ affiliations:
    index: 3
  - name: Department of Computer Science and Engineering, University of Nevada, Reno
    index: 4
-date: 30 November 2021
+date: 30 September 2022
 bibliography: paper.bib
 ---
 
@@ -55,37 +56,37 @@ This package implements a variety of standard data assimilation algorithms,
 including some of the widely used performance modifications that are used in
 practice to tune these estimators. This software framework was written originally
 to support the development and intercomparison of methods studied in [@grudzien2021fast].
-Details of the ensemble DA schemes, including pseudo-code for the methods detailing
+Details of the studied ensemble DA schemes, including pseudo-code detailing
 their implementation, and DA experiment benchmark configurations, can be found in
 the above principal reference.  Additional details on numerical integration schemes
-used in this work for simulating the Lorenz-96 model are found in the secondary
-reference [@grudzien2020numerical].
+utilized in this package can be found in the secondary reference [@grudzien2020numerical].
 
 # Statement of need
 
 Standard libraries exist for full-scale DA system research and development, e.g.,
-the Data Assimilation Research Testbed (DART)[@anderson2009data], but
+the Data Assimilation Research Testbed (DART) [@anderson2009data], but
 there are fewer standard options for theoretical research and algorithm development in
 simple test systems. Many basic research frameworks, furthermore, do not include
 standard operational techniques developed from classical variational methods,
 due to the difficulty in constructing tangent linear and adjoint codes [@kalnay20074denkf].
 DataAssimilationBenchmarks.jl provides one framework for studying squential filters
 and smoothers that are commonly used in online, geoscientific prediction settings,
-including ensemble methods and variational schemes.
+including both ensemble methods and variational schemes, with hybrid method planned for
+future development.
 
 ## Comparison with similar projects
 
 Similar projects to DataAssimilationBenchmarks.jl include the DAPPER Python library
-[@patrick_n_raanes_2018_2029296], DataAssim.jl used by [@vetra2018state], and
-EnsembleKalmanProcesses.jl [@enkprocesses] of the Climate Modeling Alliance.  These alternatives
-are differentiated primarily in that:
+[@dapper], DataAssim.jl used by [@vetra2018state], and
+EnsembleKalmanProcesses.jl [@enkprocesses] of the Climate Modeling Alliance.
+These alternatives are differentiated primarily in that:
 
   * DAPPER is a Python-based library which is well-established, and includes many of the same
-	estimators and models. However, numerical simulations in Python run notably slower than simulations in Julia
-	when numerical routines cannot be vectorized in Numpy [@bezanson2017julia].
-	Particularly, this can make the wide hyper-parameter search intended above computationally challenging
-	without utilizing additional packages such as Numba [@lam2015numba] for code acceleration such as faster
-	for-loops.
+	estimators and models. However, numerical simulations in Python run notably slower than
+	simulations in Julia when numerical routines cannot be vectorized in Numpy
+	[@bezanson2017julia]. Particularly, this can make the wide hyper-parameter search
[…]
