+++ paper.bib
@@ -100,7 +100,7 @@
   author={Lam, S.K. and Pitrou, A. and Seibert, S.},
   booktitle={Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC},
   pages={1--6},
-  year={2015}
+  year={2015},
 	doi={https://doi.org/10.1145/2833157.2833162},
 }
 
