+++ .travis.yml
@@ -1,11 +1,11 @@
 language: julia
 julia:
-  - 1.7
+  - 1.8
 codecov: true
 jobs:
   include:
     - stage: "Documentation"
-      julia: 1.7
+      julia: 1.8
       os: linux
       script:
         - julia --project=docs/ -e 'using Pkg; Pkg.instantiate(); Pkg.add(PackageSpec(path=pwd()))'
