+++ paper.bib
@@ -83,6 +83,7 @@
   pages={e535},
   year={2018},
   publisher={Wiley Online Library},
+	doi={https://doi.org/10.1002/wcc.535},
 }
 
 @book{jazwinski2007stochastic,
