### 2022-09-09 cgrudz/DataAssimilationBenchmarks.jl#7 @cgrudz
I decided to add an additional section that sketches the overall workflow, where the documentation is organized a bit to emulate this workflow precisely:

https://cgrudz.github.io/DataAssimilationBenchmarks.jl/dev/submodules/experiments/Workflow/

This section in the documentation now discusses how one will go from an installation to the point of running a twin experiment following the submodules in the docs.  I think this is now a better solution, though I will certainly take further suggestions on how to clarify the use example of the software.
