### 2022-09-12 c33ec7b
Smoothing docs and references
