+++ src/methods/EnsembleKalmanSchemes.jl
@@ -86,7 +86,7 @@ end
 """
     rand_orth(N_ens::Int64)
 
-This generates a random, mean-preserving, orthogonal matrix as in [Sakov & Oke
+This generates a random, mean-preserving, orthogonal matrix as in [Sakov et al. 
 2008](https://journals.ametsoc.org/view/journals/mwre/136/3/2007mwr2021.1.xml), depending on
 the esemble size `N_ens`.
 ```
@@ -342,36 +342,36 @@ is of type [`ConM`](@ref), the keyword arguments dictionary `kwargs` is of type
 
 Currently validated `analysis` options:
  * `analysis=="etkf" || analysis=="etks"` computes the deterministic ensemble transform
-   as in the ETKF described in [Grudzien, et al.
+   as in the ETKF described in [Grudzien et al.
    2021](https://gmd.copernicus.org/preprints/gmd-2021-306/).
  * `analysis[1:7]=="mlef-ls" || analysis[1:7]=="mles-ls"` computes the maximum likelihood
-   ensemble filter transform described in [Grudzien, et al.
+   ensemble filter transform described in [Grudzien et al.
    2021](https://gmd.copernicus.org/preprints/gmd-2021-306/), optimizing the nonlinear
    cost function with Newton-based
    [line searches](https://julianlsolvers.github.io/LineSearches.jl/stable/).
  * `analysis[1:4]=="mlef" || analysis[1:4]=="mles"` computes the maximum likelihood
    ensemble filter transform described in
-   [Grudzien, et al. 2021](https://gmd.copernicus.org/preprints/gmd-2021-306/),
+   [Grudzien et al. 2021](https://gmd.copernicus.org/preprints/gmd-2021-306/),
    optimizing the nonlinear
    cost function with simple Newton-based scheme.
  * `analysis=="enkf-n-dual" || analysis=="enks-n-dual"`
-   computes the dual form of the EnKF-N transform as in [Bocquet, et al.
+   computes the dual form of the EnKF-N transform as in [Bocquet et al.
    2015](https://npg.copernicus.org/articles/22/645/2015/)
    Note: this cannot be used with the nonlinear observation operator.
    This uses the Brent method for the argmin problem as this
    has been more reliable at finding a global minimum than Newton optimization.
  * `analysis=="enkf-n-primal" || analysis=="enks-n-primal"`
-   computes the primal form of the EnKF-N transform as in [Bocquet, et al.
+   computes the primal form of the EnKF-N transform as in [Bocquet et al.
    2015](https://npg.copernicus.org/articles/22/645/2015/),
-   [Grudzien, et al. 2021](https://gmd.copernicus.org/preprints/gmd-2021-306/).
+   [Grudzien et al. 2021](https://gmd.copernicus.org/preprints/gmd-2021-306/).
    This differs from the MLEF/S-N in that there is no approximate linearization of
    the observation operator in the EnKF-N, this only handles the approximation error
    with respect to the adaptive inflation. This uses a simple Newton-based
    minimization of the cost function for the adaptive inflation.
  * `analysis=="enkf-n-primal-ls" || analysis=="enks-n-primal-ls"`
-   computes the primal form of the EnKF-N transform as in [Bocquet, et al.
+   computes the primal form of the EnKF-N transform as in [Bocquet et al.
    2015](https://npg.copernicus.org/articles/22/645/2015/),
-   [Grudzien, et al. 2021](https://gmd.copernicus.org/preprints/gmd-2021-306/).
+   [Grudzien et al. 2021](https://gmd.copernicus.org/preprints/gmd-2021-306/).
    This differs from the MLEF/S-N in that there is no approximate linearization of
[…]
