+++ src/models/ObsOperators.jl
@@ -112,7 +112,7 @@ component-wise transformation of the remaining state vector components mapped to
 observation space.  For `γ=1.0`, there is no transformation applied, and the observation
 operator acts as a linear projection onto the remaining components of the state vector,
 equivalent to not specifying `γ`. For `γ>1.0`, the nonlinear observation operator of
-[Asch, et al. (2016).](https://epubs.siam.org/doi/book/10.1137/1.9781611974546),
+[Asch et al. 2016](https://epubs.siam.org/doi/book/10.1137/1.9781611974546),
 pg. 181 is applied,
 ```math
 \begin{align}
@@ -121,7 +121,7 @@ pg. 181 is applied,
 ```
 where ``\circ`` is the Schur product, and which limits to the identity for `γ=1.0`.
 If `γ=0.0`, the quadratic observation operator of
-[Hoteit, et al. (2012).](https://journals.ametsoc.org/view/journals/mwre/140/2/2011mwr3640.1.xml),
+[Hoteit et al. 2012](https://journals.ametsoc.org/view/journals/mwre/140/2/2011mwr3640.1.xml),
 ```math
 \begin{align}
 \mathcal{H}(\pmb{x}) =0.05 \pmb{x} \circ \pmb{x}
@@ -129,7 +129,7 @@ If `γ=0.0`, the quadratic observation operator of
 ```
 is applied to the remaining state components (note, this is not a continuous limit).
 If `γ<0.0`, the exponential observation
-operator of [Wu, et al. (2014).](https://npg.copernicus.org/articles/21/955/2014/)
+operator of [Wu et al. 2014](https://npg.copernicus.org/articles/21/955/2014/)
 ```math
 \begin{align}
 \mathcal{H}(\pmb{x}) = \pmb{x} \circ \exp\{- \gamma \pmb{x} \}
