+++ docs/src/submodules/models/L96.md
@@ -26,7 +26,7 @@ $n\times n$ identity matrix, $\pmb{W}(t)$ is an $n$-dimensional Wiener process a
 $s(t):\mathbb{R}\rightarrow \mathbb{R}$ is a measurable function of (possibly)
 time-varying diffusion coefficients. This model is analyzed in-depth for data assimilation
 twin experiments in the manuscript
-[Grudzien, C. et al. (2020).](https://gmd.copernicus.org/articles/13/1903/2020/gmd-13-1903-2020.html)
+[Grudzien et al. 2020](https://gmd.copernicus.org/articles/13/1903/2020/gmd-13-1903-2020.html)
 and further details of using the system for data assimilation benchmarks in stochastic
 dynamics are discussed there.  The methods in the below define the model equations,
 the Jacobian, and the order 2.0 Taylor-Stratonovich scheme derived especially
