+++ docs/src/submodules/methods/EnsembleKalmanSchemes.md
@@ -13,7 +13,7 @@ a past, reanlyzed ensemble state.  The Gauss-Newton iterative smoothers are 4D s
 which iteratively optimize the initial condition at the beginning of a data assimilation
 cycle, and propagate this initial condition to initialize the subsequent cycle. A full
 discussion of these methods can be found in
-[Grudzien, et al. 2021](https://gmd.copernicus.org/preprints/gmd-2021-306/).
+[Grudzien et al. 2021](https://gmd.copernicus.org/preprints/gmd-2021-306/).
 
 For each outer-loop method defining the data assimilation cycle, different types of analyses
 can be specified within their arguments.  Likewise, these outer-loop methods require
