### 2022-09-16 3832213
Update docstrings and documentation
Added docstrings for XdVAR descriptions and proof-read/made minor corrections the documentation
