+++ src/methods/XdVAR.jl
@@ -2,111 +2,195 @@
 module XdVAR
 ##############################################################################################
 # imports and exports
-using LinearAlgebra, ForwardDiff
+using LinearAlgebra, SparseArrays
 using ..DataAssimilationBenchmarks
-export D3_var_cost, D3_var_grad, D3_var_hessian, D3_var_NewtonOp  
+using ForwardDiff
 ##############################################################################################
 # Main methods
 ##############################################################################################
 """
+<<<<<<< Updated upstream
+    D3_var_cost(x::VecA(T), obs::VecA(T), x_background::VecA(T), state_cov::CovM(T),
+=======
+<<<<<<< Updated upstream
+    D3_var_cost(x::VecA(T), obs::VecA(T), x_background::VecA(T), state_cov::CovM(T), 
+>>>>>>> Stashed changes
+    obs_cov::CovM(T), kwargs::StepKwargs) where T <: Real
+=======
     D3_var_cost(x::VecA(T), obs::VecA(T), x_bkg::VecA(T), state_cov::CovM(T),
                 obs_cov::CovM(T), kwargs::StepKwargs) where T <: Real
 
-A description of purpose, behavior and outputs
-"""
-function D3_var_cost(x::VecA(T), obs::VecA(T), x_bkg::VecA(T), state_cov::CovM(T),
-                     H_obs::Function, obs_cov::CovM(T), kwargs::StepKwargs) where T <: Real
+Computes the cost of the three-dimensional variational analysis increment from an initial state 
+proposal with a static background covariance
 
+'x' is the initial state proposal vector, 'obs' is to the observation vector, 'x_bkg'
+is a free argument used to evaluate the cost of the given state proposal versus other proposal 
+states, state_cov is the background error covariance matrix, H_obs is a model mapping operator 
+for observations, and obs_cov is the observation error covariance matrix. 'kwargs' refers to 
+any additional arguments needed for the operation computation.
+
+Note: The observational and background components are equally weighted. 
+
+```
+return  0.5*back_component + 0.5*obs_component
+```
+>>>>>>> Stashed changes
+"""
+function D3_var_cost(x::VecA(T), obs::VecA(T), x_background::VecA(T), state_cov::CovM(T),
+    H_obs::Function, obs_cov::CovM(T), kwargs::StepKwargs) where T <: Real
     # initializations
     obs_dim = length(obs)
-
     # obs operator
+<<<<<<< Updated upstream
+    H = H_obs(x,obs_dim,kwargs)
+=======
     H = H_obs(x, obs_dim, kwargs)
 
-    # background discepancy
+    # background discrepancy
     δ_b = x - x_bkg
 
     # observation discrepancy
     δ_o = obs - H
+>>>>>>> Stashed changes
 
-    # cost function
-    back_component = dot(δ_b, inv(state_cov) * δ_b)
-    obs_component = dot(δ_o, inv(obs_cov) * δ_o)
-
-    0.5*back_component + 0.5*obs_component
+    back_component = transpose((x - x_background))*(inv(state_cov))*(x - x_background)
+    obs_component = transpose((obs - H))*(inv(obs_cov))*(obs - H)
+    J = 0.5*back_component + 0.5*obs_component
+    return J
 end
 
 
 ##############################################################################################
 """
+<<<<<<< Updated upstream
+    D3_var_grad(x::VecA(T), obs::VecA(T), x_background::VecA(T), state_cov::CovM(T),
+=======
+<<<<<<< Updated upstream
[…]
