+++ docs/src/community/Contributing.md
@@ -2,7 +2,7 @@
 
 ## What to expect
 
-This is currenlty a small project maintained by a small community of researchers and
+This is currently a small project maintained by a small community of researchers and
 developers, focused on academic research in data assimilation and machine learning
 methodology. This software is not intended for use in an operational data assimilation
 environment, or even for use with real data, as there are a variety of existing alternatives
@@ -36,23 +36,23 @@ bug you have encountered that will allow the community to reproduce and solve th
 Reproducibility is key, and if there are not sufficient details to reproduce an issue,
 the issue will be sent back for more details.
 
-## How to contribute new methods, models or other core core
+## How to contribute new methods, models or other core code
 
-The best way to contribute new code is to reach out to the community first, as this code
-base is in an early and active state of development and will occassionally face breaking
-changes in order to accomodate more generality and new features.  Please start with an
+The best way to contribute new code is to reach out to the community first as this code
+base is in an early and active state of development and will occasionally face breaking
+changes in order to accommodate more generality and new features.  Please start with an
 introduction of yourself in the 
 [Github discussions](https://github.com/cgrudz/DataAssimilationBenchmarks.jl/discussions)
 followed by a detailed feature request in the
 [Github issues](https://github.com/cgrudz/DataAssimilationBenchmarks.jl/issues),
 covering your use-case and what new functionality you are proposing. This will help
 the community anticipate your needs and the backend changes that might need to be implemented
-in order to accomodate new functionality. There is not currently a general system for
-how new data assmiliation methods or models are to be implemented, and it is therefore
+in order to accommodate new functionality. There is not currently a general system for
+implementing new data assimilation methods or models, and it is therefore
 critical to bring up your use-case to the community so that how this new feature is
-incorporated can be planned into the development.  Once the issue can be evaluated and
+incorporated can be planned into the development. Once the issue can be evaluated and
 discussed by the development community, the strategy is usually to create a fork of the
-main code base where new modules and features can be prototyped.  Once the new code
+main code base where new modules and features can be prototyped. Once the new code
 development is ready for a review, a pull request can be made where the new functionality
 may be merged, and possibly further refactored for general consistency and consolidation
 of codes.
