+++ docs/src/home/DataAssimilationBenchmarks.md
@@ -2,8 +2,8 @@
 The following type and type constructors are declared for optimizing numerical routines,
 for using multiple dispatch of functions with different specified input forms and for
 passing arguments between inner / outer loop steps of the DA twin experiment. Type
-constructors are designed to be flexible enough to handle multiple dispactch for
-automatic code differentiation, though seek to ensure type consistency within mehtods
+constructors are designed to be flexible enough to handle multiple dispatch for
+automatic code differentiation, though seek to ensure type consistency within methods
 for improved performance.
 
 ```@autodocs

[…]
