### 2022-09-21 d96f493
Update XdVAR.jl
Preserving rewritten changes and description
