+++ src/methods/XdVAR.jl
@@ -2,23 +2,15 @@
 module XdVAR
 ##############################################################################################
 # imports and exports
-using LinearAlgebra, SparseArrays
+using LinearAlgebra, ForwardDiff
 using ..DataAssimilationBenchmarks
-using ForwardDiff
+export D3_var_cost, D3_var_grad, D3_var_hessian, D3_var_NewtonOp  
 ##############################################################################################
 # Main methods
 ##############################################################################################
 """
-<<<<<<< Updated upstream
-    D3_var_cost(x::VecA(T), obs::VecA(T), x_background::VecA(T), state_cov::CovM(T),
-=======
-<<<<<<< Updated upstream
-    D3_var_cost(x::VecA(T), obs::VecA(T), x_background::VecA(T), state_cov::CovM(T), 
->>>>>>> Stashed changes
-    obs_cov::CovM(T), kwargs::StepKwargs) where T <: Real
-=======
     D3_var_cost(x::VecA(T), obs::VecA(T), x_bkg::VecA(T), state_cov::CovM(T),
-                obs_cov::CovM(T), kwargs::StepKwargs) where T <: Real
+        H_obs::Function, obs_cov::CovM(T), kwargs::StepKwargs) where T <: Real
 
 Computes the cost of the three-dimensional variational analysis increment from an initial state 
 proposal with a static background covariance
@@ -34,44 +26,34 @@ Note: The observational and background components are equally weighted.
 ```
 return  0.5*back_component + 0.5*obs_component
 ```
->>>>>>> Stashed changes
 """
-function D3_var_cost(x::VecA(T), obs::VecA(T), x_background::VecA(T), state_cov::CovM(T),
-    H_obs::Function, obs_cov::CovM(T), kwargs::StepKwargs) where T <: Real
+function D3_var_cost(x::VecA(T), obs::VecA(T), x_bkg::VecA(T), state_cov::CovM(T),
+                     H_obs::Function, obs_cov::CovM(T), kwargs::StepKwargs) where T <: Real
+
     # initializations
     obs_dim = length(obs)
+
     # obs operator
-<<<<<<< Updated upstream
-    H = H_obs(x,obs_dim,kwargs)
-=======
     H = H_obs(x, obs_dim, kwargs)
 
-    # background discrepancy
+    # background discepancy
     δ_b = x - x_bkg
 
     # observation discrepancy
     δ_o = obs - H
->>>>>>> Stashed changes
 
-    back_component = transpose((x - x_background))*(inv(state_cov))*(x - x_background)
-    obs_component = transpose((obs - H))*(inv(obs_cov))*(obs - H)
-    J = 0.5*back_component + 0.5*obs_component
-    return J
+    # cost function
+    back_component = dot(δ_b, inv(state_cov) * δ_b)
+    obs_component = dot(δ_o, inv(obs_cov) * δ_o)
+
+    0.5*back_component + 0.5*obs_component
 end
 
 
 ##############################################################################################
 """
-<<<<<<< Updated upstream
-    D3_var_grad(x::VecA(T), obs::VecA(T), x_background::VecA(T), state_cov::CovM(T),
-=======
-<<<<<<< Updated upstream
-    D3_var_grad(x::VecA(T), obs::VecA(T), x_background::VecA(T), state_cov::CovM(T), 
->>>>>>> Stashed changes
-    obs_cov::CovM(T), kwargs::StepKwargs) where T <: Real
-=======
-    D3_var_grad(x::VecA(T), obs::VecA(T), x_bkg::VecA(T), state_cov::CovM(T), 
[…]
