### 2022-09-21 8847241
Minor corrections to descriptions
