+++ src/methods/XdVAR.jl
@@ -15,13 +15,11 @@ export D3_var_cost, D3_var_grad, D3_var_hessian, D3_var_NewtonOp
 Computes the cost of the three-dimensional variational analysis increment from an initial state 
 proposal with a static background covariance
 
-'x' is the initial state proposal vector, 'obs' is to the observation vector, 'x_bkg'
-is a free argument used to evaluate the cost of the given state proposal versus other proposal 
-states, state_cov is the background error covariance matrix, H_obs is a model mapping operator 
+'x' is a free argument used to evaluate the cost of the given state proposal versus other 
+proposal states, 'obs' is to the observation vector, 'x_bkg' is the initial state proposal 
+vector, state_cov is the background error covariance matrix, H_obs is a model mapping operator 
 for observations, and obs_cov is the observation error covariance matrix. 'kwargs' refers to 
 any additional arguments needed for the operation computation.
-
-Note: The observational and background components are equally weighted. 
 
 ```
 return  0.5*back_component + 0.5*obs_component
@@ -59,11 +57,14 @@ Computes the gradient of the three-dimensional variational analysis increment fr
 state proposal with a static background covariance using a wrapper function for automatic 
 differentiation
 
-'x' is the initial state proposal vector, 'obs' is to the observation vector, 'x_bkg'
-is a free argument used to evaluate the cost of the given state proposal versus other proposal 
-states, state_cov is the background error covariance matrix, H_obs is a model mapping operator 
+'x' is a free argument used to evaluate the cost of the given state proposal versus other 
+proposal states, 'obs' is to the observation vector, 'x_bkg' is the initial state proposal 
+vector, state_cov is the background error covariance matrix, H_obs is a model mapping operator 
 for observations, and obs_cov is the observation error covariance matrix. 'kwargs' refers to 
 any additional arguments needed for the operation computation.
+
+'wrap_cost' is a function that allows differentiation with respect to the free argument 'x'
+while treating all other hyperparameters of the cost function as constant.
 
 ```
 return  ForwardDiff.gradient(wrap_cost, x)
@@ -90,11 +91,14 @@ Computes the hessian of the three-dimensional variational analysis increment fro
 state proposal with a static background covariance using a wrapper function for automatic 
 differentiation
 
-'x' is the initial state proposal vector, 'obs' is to the observation vector, 'x_bkg'
-is a free argument used to evaluate the cost of the given state proposal versus other proposal 
-states, state_cov is the background error covariance matrix, H_obs is a model mapping operator 
-for observations, and obs_cov is the observation error covariance matrix. 'kwargs' refers to 
-any additional arguments needed for the operation computation.
+'x' is a free argument used to evaluate the cost of the given state proposal versus other 
+proposal states, 'obs' is to the observation vector, 'x_bkg' is the initial state proposal 
+vector, state_cov is the background error covariance matrix, H_obs is a model mapping 
+operator for observations, and obs_cov is the observation error covariance matrix. 'kwargs' 
+refers to any additional arguments needed for the operation computation.
+
+'wrap_cost' is a function that allows differentiation with respect to the free argument 'x'
+while treating all other hyperparameters of the cost function as constant.
 
 ```
 return  ForwardDiff.hessian(wrap_cost, x)
@@ -174,7 +178,6 @@ function D3_var_NewtonOp(x_bkg::VecA(T), obs::VecA(T), state_cov::CovM(T), H_obs
     end
     return x
 end
-
 
 ##############################################################################################
 # end module
