### 2022-09-21 6459ce9
NewtonOp documentation correction
