### 2022-09-26 e1c936e
Parallel experiments and plotting updates for obs information
