+++ src/experiments/ParallelExperimentDriver.jl
@@ -159,14 +159,14 @@ function D3_var_tuned_inflation()
                          ".jld2"
 
     # define ranges for filter parameters
-    γs = [1.0, 2.0, 3.0]
+    dims = collect(1, 1, 40)
     bkg_covs = ["ID", "clima"]
     s_infls = 0.005:0.005:1.0
 
     # load the experiments
     args = Vector{Any}()
     for s_infl in s_infls
-        for γ in γs
+        for dim in dims
             for bkg_cov in bkg_covs
                 tmp = (
                     time_series = time_series,
@@ -174,8 +174,8 @@ function D3_var_tuned_inflation()
                     seed        = 0,
                     nanl        = 3500,
                     obs_un      = 1.0,
-                    obs_dim     = 40,
-                    γ           = γ,
+                    obs_dim     = dim,
+                    γ           = 1.0,
                     s_infl      = s_infl,
                     )
                 push!(args, tmp)
