+++ src/analysis/plot_var_analysis.jl
@@ -9,21 +9,38 @@ using DataAssimilationBenchmarks
 
 ##############################################################################################
 
+bkg_covs = ["ID", "clima"]
+dims = collect(1, 1, 40)
+
 path = pkgdir(DataAssimilationBenchmarks) * "/src/data/D3-var-bkg-ID/"
-file = "bkg-ID_L96_state_seed_0000_diff_0.000_sysD_40_obsD_40_obsU_1.00_gamma_001.0_nanl_03500" *
-"_tanl_0.05_h_0.05_stateInfl_0.200.jld2"
 
-ts = load(path * file)::Dict{String,Any}
-nanl = ts["nanl"]::Int64
-filt_rmse = ts["filt_rmse"]::Array{Float64, 1}
-t = collect(1:1:nanl)
+stab_rmse_id = Vector{Float64}(undef, 40)
+stab_rmse_clima = Vector{Float64}(undef, 40)
 
-plot(t, filt_rmse, label = "Filter", title="Uninformative Prior: Filter RMSE vs. Cycles", 
-legend_position = :topright, margin=15mm, size=(800,500), dpi = 600)
-xlabel!("Cycles")
-ylabel!("Filter RMSE")
-print("Mean " * string(mean(filt_rmse[501:end]))  * " minutes\n")
-savefig("stab_I")
+for bkg_cov in bkg_covs
+    for dim in dims
+        file = "bkg-" * bkg_cov * "_L96_state_seed_0000_diff_0.000_sysD_40_obsD_" * lpad(dim, 2, "0") * 
+        "_obsU_1.00_gamma_001.0_nanl_03500_tanl_0.05_h_0.05_stateInfl_1.000.jld2"
+
+        ts = load(path * file)::Dict{String,Any}
+        nanl = ts["nanl"]::Int64
+        filt_rmse = ts["filt_rmse"]::Array{Float64, 1}
+
+        for cycle in cycles
+            if cmp(bkg_cov, "ID") == 0
+                stab_rmse_id[dim] = mean(filt_rmse[3000:nanl])
+            else
+                stab_rmse_clima[dim] = mean(filt_rmse[3000:nanl])
+            end
+        end
+    end
+end
+
+plot(dims, stab_rmse_id, label = "ID", title="Stabilized Filter RMSE vs. Cycles", legend_position = :topright, margin=15mm, size=(800,500), dpi = 600)
+plot!(dims, stab_rmse_clima, label = "ID")
+xlabel!("Dimensions")
+ylabel!("Stabilized RMSE")
+savefig("stab_both")
 
 ##############################################################################################
 # end module
