### 2022-09-26 852c348
Fixed several bugs in plotting scheme
