+++ src/methods/XdVAR.jl
@@ -15,11 +15,11 @@ export D3_var_cost, D3_var_grad, D3_var_hessian, D3_var_NewtonOp
 Computes the cost of the three-dimensional variational analysis increment from an initial state 
 proposal with a static background covariance
 
-'x' is a free argument used to evaluate the cost of the given state proposal versus other 
-proposal states, 'obs' is to the observation vector, 'x_bkg' is the initial state proposal 
-vector, state_cov is the background error covariance matrix, H_obs is a model mapping operator 
-for observations, and obs_cov is the observation error covariance matrix. 'kwargs' refers to 
-any additional arguments needed for the operation computation.
+`x` is a free argument used to evaluate the cost of the given state proposal versus other 
+proposal states, `obs` is to the observation vector, `x_bkg` is the initial state proposal 
+vector, `state_cov` is the background error covariance matrix, `H_obs` is a model mapping
+operator for observations, and `obs_cov` is the observation error covariance matrix.
+`kwargs` refers to any additional arguments needed for the operation computation.
 
 ```
 return  0.5*back_component + 0.5*obs_component
@@ -57,13 +57,13 @@ Computes the gradient of the three-dimensional variational analysis increment fr
 state proposal with a static background covariance using a wrapper function for automatic 
 differentiation
 
-'x' is a free argument used to evaluate the cost of the given state proposal versus other 
-proposal states, 'obs' is to the observation vector, 'x_bkg' is the initial state proposal 
-vector, state_cov is the background error covariance matrix, H_obs is a model mapping operator 
-for observations, and obs_cov is the observation error covariance matrix. 'kwargs' refers to 
-any additional arguments needed for the operation computation.
+`x` is a free argument used to evaluate the cost of the given state proposal versus other 
+proposal states, `obs` is to the observation vector, `x_bkg` is the initial state proposal 
+vector, `state_cov` is the background error covariance matrix, `H_obs` is a model mapping
+operator  for observations, and `obs_cov` is the observation error covariance matrix.
+`kwargs` refers to any additional arguments needed for the operation computation.
 
-'wrap_cost' is a function that allows differentiation with respect to the free argument 'x'
+`wrap_cost` is a function that allows differentiation with respect to the free argument `x`
 while treating all other hyperparameters of the cost function as constant.
 
 ```
@@ -87,17 +87,17 @@ end
     D3_var_hessian(x::VecA(T), obs::VecA(T), x_bkg::VecA(T), state_cov::CovM(T),
         H_obs::Function, obs_cov::CovM(T), kwargs::StepKwargs) where T <: Float64 
 
-Computes the hessian of the three-dimensional variational analysis increment from an initial 
+Computes the Hessian of the three-dimensional variational analysis increment from an initial 
 state proposal with a static background covariance using a wrapper function for automatic 
 differentiation
 
-'x' is a free argument used to evaluate the cost of the given state proposal versus other 
-proposal states, 'obs' is to the observation vector, 'x_bkg' is the initial state proposal 
-vector, state_cov is the background error covariance matrix, H_obs is a model mapping 
-operator for observations, and obs_cov is the observation error covariance matrix. 'kwargs' 
-refers to any additional arguments needed for the operation computation.
+`x` is a free argument used to evaluate the cost of the given state proposal versus other 
+proposal states, `obs` is to the observation vector, `x_bkg` is the initial state proposal 
+vector, `state_cov` is the background error covariance matrix, `H_obs` is a model mapping 
[…]
