### 2022-10-12 221d6e2
Updating syntax and fixing version for the production figure
