+++ test/TestFilterExps.jl
@@ -72,7 +72,7 @@ end
 
 
 ##############################################################################################
-# run and analyze the ETKF for state estimation with the Lorenz-96 model
+# run and analyze the ETKF for state estimation with inflation and the Lorenz-96 model
 
 function run_D3_var_filter_infl_L96()
     try
@@ -85,13 +85,14 @@ end
 
 function analyze_D3_var_filter_infl_L96()
     try
-        # test if the filter RMSE for standard simulation falls below adequate threshold
+        # test if the filter RMSE for inflated simulation falls below adequate threshold
         path = pkgdir(DataAssimilationBenchmarks) * "/src/data/D3-var-bkg-ID/"
         data = load(path * "bkg-ID_L96_state_seed_0123_diff_0.000_sysD_40_obsD_40_obsU_1.00_"*
         "gamma_001.0_nanl_03500_tanl_0.05_h_0.05_stateInfl_0.230.jld2")
         nanl = data["nanl"]
         rmse = data["filt_rmse"]
         display(mean(filter(!isnan,rmse[1000:nanl])))
+
         # note, we use a small burn-in to reach more regular cycles
         if mean(filter(!isnan,rmse[1000:nanl])) < 0.40
             true
