+++ src/experiments/SingleExperimentDriver.jl
@@ -124,7 +124,7 @@ enkf_exps = Dict{String, Any}(
 Pre-configured 3D-VAR style twin experiments for benchmark models. 
 """
 d3_var_exps = Dict{String, Any}(
-    # Lorenz-96 3D-VAR state estimation standard configuration
+    # Lorenz-96 3D-VAR state estimation configuration
     "L96_D3_var_state_test" => (
       time_series = path *
       "L96_time_series_seed_0000_dim_40_diff_0.000_F_08.0_tanl_0.05_nanl_05000_"*
@@ -135,20 +135,7 @@ d3_var_exps = Dict{String, Any}(
       obs_un      = 1.0,
       obs_dim     = 40,
       γ           = 1.00,
-      s_infl      = 0.20,
-     ),
-     # Lorenz-96 3D-VAR optimal inflation estimation configuration
-    "L96_D3_var_infl_test" => (
-      time_series = path *
-      "L96_time_series_seed_0123_dim_40_diff_0.000_F_08.0_tanl_0.05_nanl_06500_"*
-      "spin_1500_h_0.050.jld2",
-      bkg_cov     = "ID",
-      seed        = 123,
-      nanl        = 3500,
-      obs_un      = 1.0,
-      obs_dim     = 40,
-      γ           = 1.0,
-      s_infl      = 0.230,
+      s_infl      = 0.23,
      )
     )
 
