+++ CITATION.bib
@@ -1,7 +1,7 @@
 @misc{DataAssimilationBenchmarks.jl,
 	author  = {Colin Grudzien <cgrudz@mailbox.org> and contributors},
 	title   = {DataAssimilationBenchmarks.jl},
-	url     = {https://github.com/Colin Grudzien/DataAssimilationBenchmarks.jl},
+	url     = {https://github.com/cgrudz/DataAssimilationBenchmarks.jl},
 	version = {v0.2.0},
 	year    = {2021},
 	month   = {12}
