+++ Project.toml
@@ -1,7 +1,7 @@
 name = "DataAssimilationBenchmarks"
 uuid = "7c720533-a3c7-49b7-973c-85ef17efca74"
 authors = ["Colin Grudzien <cgrudzien@ucsd.edu> and contributors"]
-version = "0.2.0"
+version = "0.3.0"
 
 [deps]
 BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
@@ -27,6 +27,21 @@ StatsFuns = "4c63d2b9-4356-54db-8cca-17b64c39e42c"
 
 [compat]
 julia = ">= 1.6"
+BenchmarkTools = ">= 1.3"
+Distributions = ">= 0.25"
+Documenter = ">= 0.27"
+DocumenterTools = ">= 0.1"
+FileIO = ">= 1.16"
+ForwardDiff = ">= 0.10"
+HDF5 = ">= 0.16"
+JLD2 = ">= 0.4"
+LineSearches = ">= 7.2"
+LsqFit = ">= 0.13"
+Measures = ">= 0.3"
+Optim = ">= 1.7"
+Plots = ">= 1.35"
+StatsBase = ">= 0.33"
+StatsFuns = ">= 1.0"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
