+++ paper.bib
@@ -107,7 +107,7 @@
 
 @article{kalnay20074denkf,
   title={4-D-Var or ensemble Kalman filter?},
-  author={Kalnay, E and Li, H and Miyoshi, T and Yang, S.C. and Ballabrera-Poy, J.},
+  author={Kalnay, E. and Li, H. and Miyoshi, T. and Yang, S.C. and Ballabrera-Poy, J.},
   journal={Tellus A: Dynamic Meteorology and Oceanography},
   volume={59},
   number={5},
