+++ README.md
@@ -30,10 +30,10 @@ in the development branch.  Currently validated techniques are available in the 
 branch.
 
 This package supported the development of all numerical results and benchmark simulations
-in the pre-print
+in manuscript 
 [A fast, single-iteration ensemble Kalman smoother for sequential data
 assimilation](https://gmd.copernicus.org/preprints/gmd-2021-306/)
-available currently in open review in Geoscientific Model Development.
+pending final publication in Geoscientific Model Development.
 
 ### Documentation
 
