+++ Project.toml
@@ -26,22 +26,22 @@ StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 StatsFuns = "4c63d2b9-4356-54db-8cca-17b64c39e42c"
 
 [compat]
-julia = ">= 1.6"
-BenchmarkTools = ">= 1.3"
-Distributions = ">= 0.25"
-Documenter = ">= 0.27"
-DocumenterTools = ">= 0.1"
-FileIO = ">= 1.16"
-ForwardDiff = ">= 0.10"
-HDF5 = ">= 0.16"
-JLD2 = ">= 0.4"
-LineSearches = ">= 7.2"
-LsqFit = ">= 0.13"
-Measures = ">= 0.3"
-Optim = ">= 1.7"
-Plots = ">= 1.35"
-StatsBase = ">= 0.33"
-StatsFuns = ">= 1.0"
+julia = "1.8"
+BenchmarkTools = "1.3.1"
+Distributions = "0.25.76"
+Documenter = "0.27.23"
+DocumenterTools = "0.1.16"
+FileIO = "1.16.0"
+ForwardDiff = "1.16.0"
+HDF5 = "0.16.11"
+JLD2 = "0.4.25"
+LineSearches = "7.2.0"
+LsqFit = "0.13.0"
+Measures = "0.3.1"
+Optim = "1.7.3"
+Plots = "1.35.3"
+StatsBase = "0.33.21"
+StatsFuns = "1.0.1"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
