+++ README.md
@@ -30,7 +30,7 @@ in the development branch.  Currently validated techniques are available in the 
 branch.
 
 This package supported the development of all numerical results and benchmark simulations
-in manuscript 
+in the manuscript 
 [A fast, single-iteration ensemble Kalman smoother for sequential data
 assimilation](https://gmd.copernicus.org/preprints/gmd-2021-306/)
 pending final publication in Geoscientific Model Development.
