+++ Project.toml
@@ -32,7 +32,7 @@ Distributions = "0.25.76"
 Documenter = "0.27.23"
 DocumenterTools = "0.1.16"
 FileIO = "1.16.0"
-ForwardDiff = "1.16.0"
+ForwardDiff = "0.10.32"
 HDF5 = "0.16.11"
 JLD2 = "0.4.25"
 LineSearches = "7.2.0"
