### 2022-10-17 2281f2a
Updating citation and version numbering in package
