+++ CITATION.bib
@@ -1,8 +1,8 @@
 @misc{DataAssimilationBenchmarks.jl,
-	author  = {Colin Grudzien <cgrudz@mailbox.org> and contributors},
+	author  = {Colin Grudzien <cgrudzien@ucsd.edu> and contributors},
 	title   = {DataAssimilationBenchmarks.jl},
 	url     = {https://github.com/cgrudz/DataAssimilationBenchmarks.jl},
-	version = {v0.2.0},
-	year    = {2021},
-	month   = {12}
+	version = {v0.3.0},
+	year    = {2022},
+	month   = {10}
 }
