### 2022-10-17 #4129 @cgrudz
@diehlpk, thanks for reaching out, the timing is very good actually because we were just waiting to finish off some last test cases and push a new tagged version to the Julia General Registries.  I think everything is now complete with our revision, and the new tag is in-progress with the Julia tagbot action pending:

https://github.com/JuliaRegistries/General/pull/70433

Please feel free to start the next round of review, we are now satisfied with how we've addressed the original issues and how we have improved the quality of the software and its documentation.

Cheers,
Colin
