### 2022-10-22 #4129 @tmigot
@cgrudz I have a quick question. Is there a reason why you promote the `master` branch in the documentation and installation guidelines and not the stable release?
