### 2022-10-22 cgrudz/DataAssimilationBenchmarks.jl#29 @tmigot
Add badges for doc in README
Related to JOSS submission https://github.com/openjournals/joss-reviews/issues/4129

I think it would be great to have a badge instead of the link to the documentation in the readme, and also to "promote" the stable documentation additionally to the master-version.

```
[docs-stable-url]: https://cgrudz.github.io/DataAssimilationBenchmarks.jl/stable
[docs-dev-url]: https://cgrudz.github.io/DataAssimilationBenchmarks.jl/dev
```

Let me know if you think it's a good idea.
