### 2022-10-24 #4129 @cgrudz
> @cgrudz Please check the error message from here https://github.com/openjournals/joss-papers/actions/runs/3305148863/jobs/5455018587 and fix your paper.

@diehlpk, I think I caught the issue, can you reinitialize a build to test?
