### 2022-10-24 #4129 @cgrudz
> @cgrudz I have a quick question. Is there a reason why you promote the `master` branch in the documentation and installation guidelines and not the stable release?

@tmigot, currently Documenter.jl is configured in the project to automatically update the documentation when there is a commit to master. We are anticipating that the master branch and documentation will be kept consistent with each other, and as this is in continuous development, the master branch will be running ahead of the last tagged stable version.
