### 2022-10-24 d0f307a
Updating paper bib
