+++ paper.bib
@@ -10,14 +10,16 @@
 	doi={https://doi.org/10.1175/2009BAMS2618.1},
 }
 
-@article{grudzien2021fast,
-  title={A fast, single-iteration ensemble Kalman smoother for sequential data assimilation},
-  author={Grudzien, C., and Bocquet, M.},
-  journal={Geoscientific Model Development Discussions},
-  pages={1--62},
-  year={2021},
-  publisher={Copernicus GmbH},
-	doi={https://doi.org/10.5194/gmd-2021-306},
+@Article{grudzien2022fast,
+AUTHOR = {Grudzien, C. and Bocquet, M.},
+TITLE = {A fast, single-iteration ensemble Kalman smoother for sequential data assimilation},
+JOURNAL = {Geoscientific Model Development},
+VOLUME = {15},
+YEAR = {2022},
+NUMBER = {20},
+PAGES = {7641--7681},
+URL = {https://gmd.copernicus.org/articles/15/7641/2022/},
+DOI = {10.5194/gmd-15-7641-2022}
 }
 
 @article{vetra2018state,
