### 2022-10-24 #4129 @tmigot
@cgrudz you can also visualize the pdf here https://github.com/cgrudz/DataAssimilationBenchmarks.jl/actions/workflows/draft-pdf.yml in the Artifacts of each build
