### 2022-10-28 cgrudz/DataAssimilationBenchmarks.jl#15 @peanutfun
Thanks! I understand that "injecting" new models is not a goal of this project and I appreciate your effort in managing expectations of users who would want to do so.
