### 2022-10-28 10134e5
Remove URL prefixes from DOIs in paper.bib
