### 2022-10-28 ffdd681
Remove Dapper repo DOI from paper.bib
The JOSS paper renderer favors DOI over URL and drops the latter.
Therefore, remove the DOI altogether and reference the repo via URL.
