+++ paper.bib
@@ -40,7 +40,6 @@ DOI = {10.5194/gmd-15-7641-2022}
   month        = {September},
   year         = {2022},
   publisher    = {Zenodo},
-  doi          = {10.5281/zenodo.2029296},
   url          = {https://github.com/nansencenter/DAPPER},
 }
 
