### 2022-10-28 cgrudz/DataAssimilationBenchmarks.jl#8 @peanutfun
The documentation really has improved considerably. Thanks a lot for all the effort! ✌️
