### 2022-10-28 #4129 @peanutfun
@diehlpk @cgrudz I concluded my review and recommend DataAssimilationBenchmarks.jl for publication. I am happy to see so many improvements since the start of this review. The new documentation page in particular is a great addition. Thank you for the fun and constructive review process!
