### 2022-11-01 cgrudz/DataAssimilationBenchmarks.jl#32 @tmigot
Upper-cases in bibliography
Regarding the review https://github.com/openjournals/joss-reviews/issues/4129

The upper cases should be respected in the bibliography  https://github.com/cgrudz/DataAssimilationBenchmarks.jl/blob/master/paper.bib

For instance, `State-of-the-art stochastic data assimilation methods for high-dimensional non-Gaussian problems` is missing the upper case to `Gaussian`.
