### 2022-11-01 737256a
Fixing cases with bibtex
