+++ paper.bib
@@ -1,5 +1,5 @@
 @article{anderson2009data,
-  title={The data assimilation research testbed: A community facility},
+  title={The data assimilation research testbed: {A} community facility},
   author={Anderson, J., and Hoar, T., and Raeder, K., and Liu, H., and Collins, N., and Torn, R., and Avellano, A.},
   journal={Bulletin of the American Meteorological Society},
   volume={90},
@@ -12,7 +12,7 @@
 
 @Article{grudzien2022fast,
 AUTHOR = {Grudzien, C. and Bocquet, M.},
-TITLE = {A fast, single-iteration ensemble Kalman smoother for sequential data assimilation},
+TITLE = {A fast, single-iteration ensemble {K}alman smoother for sequential data assimilation},
 JOURNAL = {Geoscientific Model Development},
 VOLUME = {15},
 YEAR = {2022},
@@ -23,7 +23,7 @@ DOI = {10.5194/gmd-15-7641-2022}
 }
 
 @article{vetra2018state,
-  title={State-of-the-art stochastic data assimilation methods for high-dimensional non-Gaussian problems},
+  title={State-of-the-art stochastic data assimilation methods for high-dimensional non-{G}aussian problems},
   author={Vetra-Carvalho, S., and Van Leeuwen, P.J., and Nerger, L., and Barth, A., and Altaf, M.U., and Brasseur, P., and Kirchgessner, P., and Beckers, J.M.},
   journal={Tellus A: Dynamic Meteorology and Oceanography},
   volume={70},
@@ -46,7 +46,7 @@ DOI = {10.5194/gmd-15-7641-2022}
 
 @software{enkprocesses,
   author       = {Dunbar, O. and others},
-  title        = {CliMA/EnsembleKalmanProcesses.jl: v0.10.0},
+  title        = {{CliMA/EnsembleKalmanProcesses}.jl: v0.10.0},
   month        = {aug},
   year         = {2022},
   publisher    = {Zenodo},
@@ -56,7 +56,7 @@ DOI = {10.5194/gmd-15-7641-2022}
 }
 
 @article{grudzien2020numerical,
-  title={On the numerical integration of the Lorenz-96 model, with scalar additive noise, for benchmark twin experiments},
+  title={On the numerical integration of the {L}orenz-96 model, with scalar additive noise, for benchmark twin experiments},
   author={Grudzien, C., and Bocquet, M., and Carrassi, A.},
   journal={Geoscientific Model Development},
   volume={13},
@@ -77,7 +77,7 @@ DOI = {10.5194/gmd-15-7641-2022}
 }
 
 @article{carrassi2018data,
-  title={Data assimilation in the geosciences: An overview of methods, issues, and perspectives},
+  title={Data assimilation in the geosciences: {A}n overview of methods, issues, and perspectives},
   author={Carrassi, A., and Bocquet, M., and Bertino, L., and Evensen, G.},
   journal={Wiley Interdisciplinary Reviews: Climate Change},
   volume={9},
@@ -89,7 +89,7 @@ DOI = {10.5194/gmd-15-7641-2022}
 }
 
 @book{jazwinski2007stochastic,
-  title={Stochastic Processes and Filtering Theory},
+  title={Stochastic {P}rocesses and {F}iltering {T}heory},
   author={Jazwinski, A.H.},
   isbn={9780486785349},
   series={Dover Books on Electrical Engineering Series},
@@ -99,7 +99,7 @@ DOI = {10.5194/gmd-15-7641-2022}
 }
 
 @inproceedings{lam2015numba,
-  title={Numba: A llvm-based python jit compiler},
[…]
