### 2022-11-01 cgrudz/DataAssimilationBenchmarks.jl#32 @cgrudz
Thanks for the catch, this is now resolved in the latest draft:

https://github.com/openjournals/joss-reviews/issues/4129#issuecomment-1299170192
