### 2022-11-01 cgrudz/DataAssimilationBenchmarks.jl#32 @tmigot
This one https://dl.acm.org/doi/abs/10.1145/2833157.2833162 also needs upper-cases
