### 2022-11-01 b8da677
Fixing cases in bibtex
