+++ paper.bib
@@ -98,7 +98,7 @@ DOI = {10.5194/gmd-15-7641-2022}
 }
 
 @inproceedings{lam2015numba,
-  title={Numba: {A} llvm-based python jit compiler},
+  title={Numba: {A LLVM-based Python JIT compiler}},
   author={Lam, S.K. and Pitrou, A. and Seibert, S.},
   booktitle={Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC},
   pages={1--6},
