### 2022-11-01 #4129 @tmigot
@cgrudz @diehlpk I finished my review. The package looks very nice now, congrats. I still have a couple of comments.

- [ ] As asked above, https://github.com/openjournals/joss-reviews/issues/4129#issuecomment-1287949731 , I don't think it is a good idea to only focus on the master simply because there is a reason we use versioning. The user needs to rely on versions that are stable and working especially since the package has a lot of deps. Staying up-to-date with the master is great but only for developers that are actively using the package.

  The following two would solve the issue:
  - [ ] Add link to `dev` and `stable` docs to the readme https://github.com/cgrudz/DataAssimilationBenchmarks.jl/issues/29
  - [ ] Show how to install the stable version of the package
- [ ] Regarding the automated tests, it passes on my computer. However, I am slightly concerned about the code coverage as only 40% of the code is tested. This should be increased to a more satisfying level.
- [ ] Finally, I think the package still lacks a clear tutorial with a complete test scenario.
