+++ test/TestDataAssimilationBenchmarks.jl
@@ -1,0 +1,18 @@
+##############################################################################################
+module TestDataAssimilationBenchmarks
+##############################################################################################
+using DataAssimilationBenchmarks
+##############################################################################################
+function splash()
+    try
+        DataAssimilationBenchmarks.Info()
+        true
+    catch
+        false
+    end
+end
+
+##############################################################################################
+# end module
+
+end
