+++ Project.toml
@@ -1,7 +1,7 @@
 name = "DataAssimilationBenchmarks"
 uuid = "7c720533-a3c7-49b7-973c-85ef17efca74"
 authors = ["Colin Grudzien <cgrudzien@ucsd.edu> and contributors"]
-version = "0.3.0"
+version = "0.3.1"
 
 [deps]
 BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
