### 2022-11-02 cgrudz/DataAssimilationBenchmarks.jl#29 @cgrudz
Thanks for the suggestion, this is now resolved.
