+++ Project.toml
@@ -33,13 +33,13 @@ Documenter = "0.27.23"
 DocumenterTools = "0.1.16"
 FileIO = "1.16.0"
 ForwardDiff = "0.10.32"
-HDF5 = "0.16.11"
+HDF5 = "0.16.12"
 JLD2 = "0.4.25"
 LineSearches = "7.2.0"
 LsqFit = "0.13.0"
 Measures = "0.3.1"
 Optim = "1.7.3"
-Plots = "1.35.3"
+Plots = "1.35.7"
 StatsBase = "0.33.21"
 StatsFuns = "1.0.1"
 
