### 2022-11-02 #4129 @cgrudz
@tmigot, thanks for your suggestions.  I have resolved the issues that were raised in the main repository and I feel like the other items are now resolved as well.  @diehlpk, I would like to ask for your decision of the work in its current state.

@tmigot regarding your requests:

"Show how to install the stable version of the package"

This is performed by installing the latest stable official release in the Julia General Registries.  Please find the instructions [here](https://cgrudz.github.io/DataAssimilationBenchmarks.jl/dev/home/Getting%20Started/#Install-a-dev-package).

"Regarding the automated tests, it passes on my computer. However, I am slightly concerned about the code coverage as only 40% of the code is tested. This should be increased to a more satisfying level."

Please note that I have included additional test now to verify that the parallel experiment constructors work.  This has raised the coverage already, but also please not that the [Analysis Directory](https://app.codecov.io/gh/cgrudz/DataAssimilationBenchmarks.jl/tree/master/src/analysis) contains 900 tracked lines, none of which are officially supported at this time as discussed in the documentation.  If you consider the coverage of lines (1670) relative to the total tracked lines (3181) minus the non-supported Analysis codes (900), the current coverage is closer to 70% of the supported code.  I could always remove these non-supported lines, but the plan is to eventually integrate them in native Julia, though for now I feel like it doesn't hurt anyone to have a template for performing their own analysis.

"Finally, I think the package still lacks a clear tutorial with a complete test scenario."

I feel like this is adequately covered in the documentation between the discussion of the basic workflow

https://cgrudz.github.io/DataAssimilationBenchmarks.jl/dev/submodules/experiments/Workflow/

and the discussion of the parallel experiment driver for running a sensitivity test

https://cgrudz.github.io/DataAssimilationBenchmarks.jl/dev/submodules/experiments/ParallelExperimentDriver/#ParallelExperimentDriver

The kinds of experiments that can be run are furthermore clearly discussed in the open-access manuscript this software supported:

https://gmd.copernicus.org/articles/15/7641/2022/gmd-15-7641-2022.html

Between the documentation itself, and its works referenced, I think the usage of the software is adequate to run an analysis.
