### 2022-11-03 #4129 @diehlpk
> "Regarding the automated tests, it passes on my computer. However, I am slightly concerned about the code coverage as only 40% of the code is tested. This should be increased to a more satisfying level."
>
> Please note that I have included additional test now to verify that the parallel experiment constructors work. This has raised the coverage already, but also please not that the [Analysis Directory](https://app.codecov.io/gh/cgrudz/DataAssimilationBenchmarks.jl/tree/master/src/analysis) contains 900 tracked lines, none of which are officially supported at this time as discussed in the documentation. If you consider the coverage of lines (1670) relative to the total tracked lines (3181) minus the non-supported Analysis codes (900), the current coverage is closer to 70% of the supported code. I could always remove these non-supported lines, but the plan is to eventually integrate them, and for now I feel like it doesn't hurt anyone to have a template for performing their own analysis.

@cgrudz I do not know Julia well enough, but could you just exclude the one file from the analysis?
