### 2022-11-03 #4129 @diehlpk
> This is performed by installing the latest stable official release in the Julia General Registries. Please find the instructions [here](https://cgrudz.github.io/DataAssimilationBenchmarks.jl/dev/home/Getting%20Started/#Install-a-dev-package).

@cgrudz I think the information is good, but I would make it more prominent. Maybe with a own header? I think many people, who want to use the package and not add features might prefer the stable version. I think making it easier to find how to do that would benefit your users.
