### 2022-11-03 #4129 @diehlpk
> "Finally, I think the package still lacks a clear tutorial with a complete test scenario."

I think it is ok what you have in the documentation. However, a code example would be nice. So your users can adapt these examples do their needs.
