+++ src/analysis/ProcessExperimentData.jl
@@ -16,6 +16,8 @@ export process_filter_state, process_smoother_state, process_smoother_param, pro
 #
 # These scripts are designed to try to load every file according to the standard naming conventions
 # and if these files cannot be loaded, to save inf as a dummy variable for missing or corrupted data.
+# NOTE: these scripts are currently deprecated and there is only a single method provided as an example
+# Future plans include merging in methods for processing data that are more consistent.
 ########################################################################################################################
 
 function process_filter_state()
@@ -219,2227 +221,5 @@ end
 
 
 ########################################################################################################################
-
-function process_smoother_state()
-    # Create an array of the time-average RMSE and spread for each experiment, past
-    # a burn-in period to reach stationary statistics for the experiment.
-    # Lag is increasing form the origin on the first axis, inflation is increasing
-    # from the origin on the second axis, ensemble size is increasing from the origin
-    # on the third axis.
-    
-    # time the operation
-    t1 = time()
-
-    # static parameters that are not varied 
-    seed = 0
-    tanl = 0.05
-    h = 0.01
-    obs_un = 1.0
-    obs_dim = 40
-    γ = 1.0
-    sys_dim = 40
-    nanl = 20000
-    burn = 5000
-    shift = 1
-    mda = true
-    diffusion = 0.00
-    
-    # parameters in ranges that will be used in loops
-    method_list = [
-                   "etks_classic", 
-                   "enks-n-primal_classic",
-                   "mles-n-transform_classic", 
-                   "etks_single_iteration", 
-                   "enks-n-primal_single_iteration",
-                   "mles-n-transform_single_iteration", 
-                   "ienks-transform",
-                   "ienks-n-transform",
-                   "lin-ienks-transform",
-                   "lin-ienks-n-transform",
-                  ]
-    
-    analysis_list = [
-                     "fore",
-                     "filt",
-                     "post"
-                    ]
-
-    stat_list = [
-                 "rmse",
-                 "spread"
-                ]
-                 
-
-    ensemble_sizes = 15:2:41
-    total_ensembles = length(ensemble_sizes)
-    inflations = LinRange(1.00, 1.10, 11)
-    total_inflations = length(inflations)
-    lags = 1:3:91
-    total_lags = length(lags)
-
-    # define the storage dictionary here, looping over the method list
-    data = Dict{String, Array{Float64}}()
-    for method in method_list
-        if method[1:6] == "enks-n" || 
-            method[1:6] == "mles-n" || 
-            method[1:7] == "ienks-n" ||
-            method[1:11] == "lin-ienks-n"
-                # multiplicative inflation parameter should always be one, there is no dimension for this variable
-                for analysis in analysis_list
-                    for stat in stat_list
-                        data[method * "_" * analysis * "_" * stat] = Array{Float64}(undef, total_lags, total_ensembles)
-                    end
-                end
-                if method[1:5] == "ienks"
-                    # for iterative schemes, additionally store statistics of iterations
-                    data[method * "_iteration_mean"] = Array{Float64}(undef, total_lags, total_ensembles)
-                    data[method * "_iteration_std"] = Array{Float64}(undef, total_lags, total_ensembles)
-                    data[method * "_iteration_median"] = Array{Float64}(undef, total_lags, total_ensembles)
-                end
-        else
-            # create storage for lags, inflations and ensembles
-            for analysis in analysis_list
-                for stat in stat_list
-                    data[method * "_" * analysis * "_" * stat ] = Array{Float64}(undef, total_lags, total_inflations, total_ensembles)
-                end
-            end
-            if method[1:5] == "ienks"
-                # for iterative schemes, additionally store statistics of iterations
[…]
