### 2022-11-03 0b32531
Uploading example from src, updating docs and README
