+++ README.md
@@ -1,13 +1,14 @@
 # DataAssimilationBenchmarks.jl
 
 ![DataAssimilationBenchmarks.jl logo](https://github.com/cgrudz/DataAssimilationBenchmarks.jl/blob/master/assets/dabenchmarks.png)
-
-[![docs-dev-img](https://img.shields.io/badge/docs-dev-purple.svg)](https://cgrudz.github.io/DataAssimilationBenchmarks.jl/dev)
-[![docs-stable-img](https://img.shields.io/badge/docs-stable-blue.svg)](https://cgrudz.github.io/DataAssimilationBenchmarks.jl/stable)
-[![status](https://joss.theoj.org/papers/478dcc0b1608d2a4d8c930edebb58736/status.svg)](https://joss.theoj.org/papers/478dcc0b1608d2a4d8c930edebb58736)
-[![Total lines of code without comments](https://tokei.rs/b1/github/cgrudz/DataAssimilationBenchmarks.jl?category=code)](https://github.com/cgrudz/DataAssimilationBenchmarks.jl)
-[![Build Status](https://app.travis-ci.com/cgrudz/DataAssimilationBenchmarks.jl.svg?branch=master)](https://app.travis-ci.com/cgrudz/DataAssimilationBenchmarks.jl)
-[![codecov](https://codecov.io/gh/cgrudz/DataAssimilationBenchmarks.jl/branch/master/graph/badge.svg?token=3XLYTH8YSZ)](https://codecov.io/gh/cgrudz/DataAssimilationBenchmarks.jl)
+|-----------|-----------|-----------|-----------|-----------|-----------|
+| [![docs-dev-img](https://img.shields.io/badge/docs-dev-purple.svg)](https://cgrudz.github.io/DataAssimilationBenchmarks.jl/dev) |
+| [![docs-stable-img](https://img.shields.io/badge/docs-stable-blue.svg)](https://cgrudz.github.io/DataAssimilationBenchmarks.jl/stable) |
+| [![status](https://joss.theoj.org/papers/478dcc0b1608d2a4d8c930edebb58736/status.svg)](https://joss.theoj.org/papers/478dcc0b1608d2a4d8c930edebb58736) |
+| [![Total lines of code without comments](https://tokei.rs/b1/github/cgrudz/DataAssimilationBenchmarks.jl?category=code)](https://github.com/cgrudz/DataAssimilationBenchmarks.jl) |
+| [![Build Status](https://app.travis-ci.com/cgrudz/DataAssimilationBenchmarks.jl.svg?branch=master)](https://app.travis-ci.com/cgrudz/DataAssimilationBenchmarks.jl) |
+| [![codecov](https://codecov.io/gh/cgrudz/DataAssimilationBenchmarks.jl/branch/master/graph/badge.svg?token=3XLYTH8YSZ)](https://codecov.io/gh/cgrudz/DataAssimilationBenchmarks.jl) |
+|-----------|-----------|-----------|-----------|-----------|-----------|
 
 Lines of code counter (without comments or blank lines) courtesy of [Tokei](https://github.com/XAMPPRocky/tokei).
 
