### 2022-11-03 #4129 @cgrudz
> > This is performed by installing the latest stable official release in the Julia General Registries. Please find the instructions [here](https://cgrudz.github.io/DataAssimilationBenchmarks.jl/dev/home/Getting%20Started/#Install-a-dev-package).
>
> @cgrudz I think the information is good, but I would make it more prominent. Maybe with a own header? I think many people, who want to use the package and not add features might prefer the stable version. I think making it easier to find how to do that would benefit your users.

@diehlpk, thanks for the suggestions, I have implemented the [further headings](https://cgrudz.github.io/DataAssimilationBenchmarks.jl/dev/home/Getting%20Started/#Installation), and I think I should clarify the scope of the package.  The use case for the software is developing the dev package for research purposes, where the standalone methods have very little functionality by themselves.  New experiments are to be constructed by directly editing the source, using the existing `ParallelExperimentDrivers.jl` and `SingleExperimentDrivers.jl` as templates.  This is also why there isn't much integration currently with the analysis scripts, as users are expected to analyze their data according to experimental parameters that they design.
