### 2022-11-03 #4129 @diehlpk
> > > "Finally, I think the package still lacks a clear tutorial with a complete test scenario."
> >
> > I think it is ok what you have in the documentation. However, a code example would be nice. So your users can adapt these examples do their needs.
>
> @diehlpk, I have [copied source code into the documentation](https://cgrudz.github.io/DataAssimilationBenchmarks.jl/dev/submodules/experiments/ParallelExperimentDriver/) directly from the `ParallelExperimentDriver.jl` as discussed above to illustrate how an experiment is run.  The expectation again is that a user needs to define the parameters of interest for themself and to create a similar constructor function to generate the NamedTuples that define an experimental configuration.  If an experiment lies outside of the scope of the existing defined experiments, a user furthermore needs to define their own experiment using the dev package, but can utilize all of the backend of the framework.

I think that improved the concerns. @tmigot  please have a look?
