### 2022-11-04 #4129 @tmigot
It looks great now, thanks.
Final comments:
- I open a PR to fix the link to the stable doc.
- I assume there is another typo here https://github.com/cgrudz/DataAssimilationBenchmarks.jl/blob/eb357ecaade4d10f6ed23e62f777788af1970757/docs/src/home/Getting%20Started.md?plain=1#L34 where it should be `pkg> add DataAssimilationBenchmarks`?

 I can recommend DataAssimilationBenchmarks.jl for publication. Congratulation for the package, and the many improvements.
