### 2022-11-04 #4129 @cgrudz
@tmigot, thanks so much for the PR's for the issue fixes, we appreciate the attention to detail with this.  I've merged the PR but I've had a busy morning and I'll push the fixes to the stable General Registries version of the repository shortly.  Likewise, thanks for your approval of the software, this was an excellent learning experience for myself and my collaborators that vastly improved our software.  We'll be looking forward to continuing our development and extending this as the framework for further research manuscripts.
