### 2022-11-10 #4129 @cgrudz
@diehlpk, the paper version above includes countries in the affiliations.  Please find below the new release and the associated Zenodo release

https://github.com/cgrudz/DataAssimilationBenchmarks.jl/releases/tag/v0.3.4

10.5281/zenodo.7311847

Let me know if you need any further actions and thanks for your assistance!
